% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{AbstractSparseNeuroVec-class}
\alias{AbstractSparseNeuroVec-class}
\title{AbstractSparseNeuroVec}
\description{
AbstractSparseNeuroVec
}
\section{Slots}{

\describe{
\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVol}} is used to map between spatial and index/row coordinates}
}}

