% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{anatomical_axes}
\alias{anatomical_axes}
\alias{LEFT_RIGHT}
\alias{RIGHT_LEFT}
\alias{ANT_POST}
\alias{POST_ANT}
\alias{INF_SUP}
\alias{SUP_INF}
\title{Pre-defined anatomical axes}
\format{
An object of class \code{NamedAxis} of length 1.

An object of class \code{NamedAxis} of length 1.

An object of class \code{NamedAxis} of length 1.

An object of class \code{NamedAxis} of length 1.

An object of class \code{NamedAxis} of length 1.

An object of class \code{NamedAxis} of length 1.
}
\usage{
LEFT_RIGHT

RIGHT_LEFT

ANT_POST

POST_ANT

INF_SUP

SUP_INF
}
\description{
These constants define standard anatomical axes used in neuroimaging.
Each axis has a defined direction vector in 3D space.
}
\keyword{datasets}
