% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{annotate_orientation}
\alias{annotate_orientation}
\title{Add L/R and A/P/S/I annotations (optional)}
\usage{
annotate_orientation(
  plane = c("axial", "coronal", "sagittal"),
  dims,
  gp = grid::gpar(col = "white", cex = 0.9, fontface = "bold")
)
}
\arguments{
\item{plane}{"axial", "coronal", or "sagittal"}

\item{dims}{c(nrow, ncol) of the slice matrix}

\item{gp}{grid::gpar style}
}
\value{
A ggplot2 layer with annotation_custom grobs
}
\description{
Add L/R and A/P/S/I annotations (optional)
}
