% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/filebacked_neurovec.R,
%   R/mapped_neurovec.R, R/neurohypervec.R, R/neurovec.R, R/neurovol.R
\name{apply_mask}
\alias{apply_mask}
\alias{apply_mask,FileBackedNeuroVec-method}
\alias{apply_mask,MappedNeuroVec-method}
\alias{apply_mask,NeuroHyperVec-method}
\alias{apply_mask,DenseNeuroVec-method}
\alias{apply_mask,AbstractSparseNeuroVec-method}
\alias{apply_mask,NeuroVol-method}
\alias{apply_mask,LogicalNeuroVol-method}
\alias{apply_mask,SparseNeuroVol-method}
\title{Apply a spatial mask to an image}
\usage{
apply_mask(x, mask)

\S4method{apply_mask}{FileBackedNeuroVec}(x, mask)

\S4method{apply_mask}{MappedNeuroVec}(x, mask)

\S4method{apply_mask}{NeuroHyperVec}(x, mask)

\S4method{apply_mask}{DenseNeuroVec}(x, mask)

\S4method{apply_mask}{AbstractSparseNeuroVec}(x, mask)

\S4method{apply_mask}{NeuroVol}(x, mask)

\S4method{apply_mask}{LogicalNeuroVol}(x, mask)

\S4method{apply_mask}{SparseNeuroVol}(x, mask)
}
\arguments{
\item{x}{A neuroimaging object.}

\item{mask}{A 3D mask supplied as a \code{LogicalNeuroVol}, numeric
\code{NeuroVol} (thresholded at \code{> 0}), logical array/vector, integer
voxel indices, or \code{NULL} for no masking.}
}
\value{
A masked neuroimaging object. Dense inputs remain dense; sparse
  inputs return a sparse object with the intersected mask.
}
\description{
Zeroes voxels outside a 3D spatial mask while preserving the geometry of the
input object. Unlike \code{\link{mask}}, which returns an object's spatial
domain, \code{apply_mask()} modifies image values.
}
\examples{
sp <- NeuroSpace(c(4, 4, 4), spacing = c(1, 1, 1))
vol <- NeuroVol(array(rnorm(64), c(4, 4, 4)), sp)
msk <- LogicalNeuroVol(array(runif(64) > 0.5, c(4, 4, 4)), sp)
masked <- apply_mask(vol, msk)
}
