% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Arith,NeuroVec,NeuroVol-method}
\alias{Arith,NeuroVec,NeuroVol-method}
\title{Arithmetic Operations for NeuroVec and NeuroVol}
\usage{
\S4method{Arith}{NeuroVec,NeuroVol}(e1, e2)
}
\arguments{
\item{e1}{A NeuroVec object.}

\item{e2}{A NeuroVol object.}
}
\value{
A DenseNeuroVec object resulting from the arithmetic operation.
}
\description{
This function performs arithmetic operations on a NeuroVec object and a NeuroVol object.
}
