% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Arith,NeuroVol,NeuroVec-method}
\alias{Arith,NeuroVol,NeuroVec-method}
\title{Arithmetic Operations for NeuroVol and NeuroVec}
\usage{
\S4method{Arith}{NeuroVol,NeuroVec}(e1, e2)
}
\arguments{
\item{e1}{A NeuroVol object.}

\item{e2}{A NeuroVec object.}
}
\value{
A DenseNeuroVec object resulting from the arithmetic operation.
}
\description{
This function performs arithmetic operations on a NeuroVol object and a
NeuroVec object.
}
