% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Arith,ROIVol,ROIVol-method}
\alias{Arith,ROIVol,ROIVol-method}
\title{This function performs arithmetic operations on two ROIVol objects.}
\usage{
\S4method{Arith}{ROIVol,ROIVol}(e1, e2)
}
\arguments{
\item{e1}{An ROIVol object.}

\item{e2}{An ROIVol object.}
}
\value{
An ROIVol object resulting from the arithmetic operation.
}
\description{
This function performs arithmetic operations on two ROIVol objects.
}
