% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ArrayLike3D-class}
\alias{ArrayLike3D-class}
\title{ArrayLike3D Class}
\description{
A virtual class for representing three-dimensional array-like objects.
It provides a common interface for 3D array operations.
}
