% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ArrayLike4D-class}
\alias{ArrayLike4D-class}
\title{ArrayLike4D Class}
\description{
A virtual class for representing four-dimensional array-like objects.
It is intended to serve as a base class for 4D array representations.
}
