% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ArrayLike5D-class}
\alias{ArrayLike5D-class}
\title{ArrayLike5D Class}
\description{
A virtual class for representing five-dimensional array-like objects.
This class serves as an interface for objects that mimic 5D arrays.
}
