% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{as,DenseNeuroVol,array}
\alias{as,DenseNeuroVol,array}
\title{Convert DenseNeuroVol to array}
\arguments{
\item{from}{A DenseNeuroVol object.}
}
\value{
An array resulting from the conversion.
}
\description{
This function converts a DenseNeuroVol object to an array.
}
\keyword{internal}
