% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R, R/neurovol.R, R/roi.R,
%   R/sparse_neurovec.R
\name{conversion from ClusteredNeuroVol to LogicalNeuroVol}
\alias{conversion from ClusteredNeuroVol to LogicalNeuroVol}
\alias{as}
\alias{as

conversion from \code{SparseNeuroVol} to \code{array}}
\alias{as
conversion from NeuroVol to LogicalNeuroVol}
\alias{as

conversion from NeuroVol to array}
\title{conversion from SparseNeuroVol to numeric}
\description{
conversion from SparseNeuroVol to numeric
}
