% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_obj.R
\docType{methods}
\name{as.array,NeuroObj-method}
\alias{as.array,NeuroObj-method}
\title{convert \code{NeuroObj} instance to array}
\usage{
\S4method{as.array}{NeuroObj}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{NeuroObj} instance to array
}
