% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.array}
\alias{as.array}
\title{Generic as.array Method}
\usage{
as.array(x, ...)
}
\arguments{
\item{x}{An object to be coerced to an array.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An \code{array} representation of the input \code{x}.
}
\description{
Coerces an object to a base array using S4 dispatch when available.
}
