% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.dense}
\alias{as.dense}
\title{Convert to dense representation}
\usage{
as.dense(x)
}
\arguments{
\item{x}{the object to densify}
}
\value{
A dense representation of the input object.
}
\description{
Convert to dense representation
}
