% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{as.list,NeuroVec-method}
\alias{as.list,NeuroVec-method}
\title{convert a \code{NeuroVec} to \code{list} of volumes.}
\usage{
\S4method{as.list}{NeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{NeuroVec} to \code{list} of volumes.
}
