% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\docType{methods}
\name{as.mask}
\alias{as.mask}
\alias{as.mask,NeuroVol,missing-method}
\alias{as.mask,NeuroVol,numeric-method}
\title{Convert to a LogicalNeuroVol}
\usage{
as.mask(x, indices)

\S4method{as.mask}{NeuroVol,missing}(x)

\S4method{as.mask}{NeuroVol,numeric}(x, indices)
}
\arguments{
\item{x}{the object to binarize}

\item{indices}{the indices to set to TRUE}
}
\description{
Convert to a LogicalNeuroVol
}
