% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.mask}
\alias{as.mask}
\title{Convert to a LogicalNeuroVol}
\usage{
as.mask(x, indices)
}
\arguments{
\item{x}{the object to binarize}

\item{indices}{the indices to set to TRUE}
}
\value{
A \code{LogicalNeuroVol} object with \code{TRUE} values at the specified \code{indices}.
}
\description{
Convert to a LogicalNeuroVol
}
