% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_neurovec.R, R/filebacked_neurovec.R,
%   R/mapped_neurovec.R, R/neurovec.R, R/neurovecseq.R, R/roi.R,
%   R/sparse_neurovec.R
\name{as.matrix,ClusteredNeuroVec-method}
\alias{as.matrix,ClusteredNeuroVec-method}
\alias{as.matrix,FileBackedNeuroVec-method}
\alias{as.matrix,MappedNeuroVec-method}
\alias{as.matrix,NeuroVec-method}
\alias{as.matrix,DenseNeuroVec-method}
\alias{as.matrix,NeuroVecSeq-method}
\alias{as.matrix,ROIVec-method}
\alias{as.matrix,AbstractSparseNeuroVec-method}
\title{convert a \code{NeuroVec} to a matrix}
\usage{
\S4method{as.matrix}{ClusteredNeuroVec}(x, by = c("cluster", "voxel"))

\S4method{as.matrix}{FileBackedNeuroVec}(x, ...)

\S4method{as.matrix}{MappedNeuroVec}(x)

\S4method{as.matrix}{NeuroVec}(x)

\S4method{as.matrix}{DenseNeuroVec}(x)

\S4method{as.matrix}{NeuroVecSeq}(x, ...)

\S4method{as.matrix}{ROIVec}(x)

\S4method{as.matrix}{AbstractSparseNeuroVec}(x, ...)
}
\arguments{
\item{x}{The object to convert to a matrix}

\item{by}{For ClusteredNeuroVec: controls the conversion target.
Defaults to "cluster" to return a TxK matrix of cluster time-series.
"voxel" is reserved for future use.}

\item{...}{Additional arguments}
}
\value{
A matrix with one row per voxel and one column per time point.

A matrix representation of the object
}
\description{
Converts a \code{\linkS4class{NeuroVecSeq}} to a dense voxel-by-time matrix.
}
