% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_obj.R
\docType{methods}
\name{as.matrix,NeuroObj-method}
\alias{as.matrix,NeuroObj-method}
\title{convert \code{NeuroObj} instance to matrix}
\usage{
\S4method{as.matrix}{NeuroObj}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{NeuroObj} instance to matrix
}
