% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.matrix}
\alias{as.matrix}
\title{Generic as.matrix Method}
\usage{
as.matrix(x, ...)
}
\arguments{
\item{x}{An object to be coerced to a matrix.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{matrix} representation of the input \code{x}.
}
\description{
Coerces an object to a matrix.
}
