% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R, R/roi.R
\name{as.numeric,SparseNeuroVol-method}
\alias{as.numeric,SparseNeuroVol-method}
\alias{as.numeric,ROIVol-method}
\title{Convert SparseNeuroVol to numeric}
\usage{
\S4method{as.numeric}{SparseNeuroVol}(x)

\S4method{as.numeric}{ROIVol}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
Convert SparseNeuroVol to numeric
}
