% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raster.R
\name{as.raster,NeuroSlice-method}
\alias{as.raster,NeuroSlice-method}
\alias{as.raster,NeuroVol-method}
\title{Convert neuroimaging objects to raster images}
\usage{
\S4method{as.raster}{NeuroSlice}(
  x,
  cmap = gray(seq(0, 1, length.out = 255)),
  irange = range(x, na.rm = TRUE),
  ...
)

\S4method{as.raster}{NeuroVol}(
  x,
  zlevel = ceiling(dim(x)[3]/2),
  cmap = gray(seq(0, 1, length.out = 255)),
  irange = range(x, na.rm = TRUE),
  ...
)
}
\arguments{
\item{x}{object to convert.}

\item{cmap}{palette name or hex-color vector for the background
(default \code{"grays"}).  See \code{\link{resolve_cmap}}.}

\item{irange}{numeric length-2 intensity range for the color scale.}

\item{...}{additional arguments passed to methods}

\item{zlevel}{slice index for \code{NeuroVol} objects. Defaults to the middle slice.}
}
\value{
A \code{raster} object
}
\description{
These methods convert \code{NeuroSlice} and \code{NeuroVol} objects
into \code{raster} images that can be displayed with
\code{grid::grid.raster} or base \code{plot}.
}
\seealso{
\code{\link{plot,NeuroSlice-method}}, \code{\link{plot,NeuroVol-method}}
}
