% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{as}
\alias{as}
\title{Coerce ROIVol to DenseNeuroVol}
\arguments{
\item{from}{An object of class \code{ROIVol} to be coerced to a \code{DenseNeuroVol}.}
}
\value{
A \code{DenseNeuroVol} object obtained by coercing the \code{ROIVol} object.
}
\description{
This function provides a method to coerce an object of class \code{ROIVol} to a \code{DenseNeuroVol}.
}
