% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_obj.R
\docType{methods}
\name{as.vector,NeuroObj-method}
\alias{as.vector,NeuroObj-method}
\title{convert \code{NeuroObj} instance to vector}
\usage{
\S4method{as.vector}{NeuroObj}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{NeuroObj} instance to vector
}
