% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{as.vector,SparseNeuroVol-method}
\alias{as.vector,SparseNeuroVol-method}
\title{Convert SparseNeuroVol to a base vector}
\usage{
\S4method{as.vector}{SparseNeuroVol}(x, mode = "any")
}
\arguments{
\item{x}{A `SparseNeuroVol` instance.}

\item{mode}{Optional coercion mode (see [base::as.vector]).}
}
\value{
A vector of length `prod(dim(x))`.
}
\description{
Supplies an `as.vector` S4 method that flattens sparse volumes to a dense
vector, keeping the same voxel ordering as `as.array`.
}
