% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{axcodes}
\alias{axcodes}
\alias{axcodes,NeuroSpace-method}
\alias{axcodes,NeuroObj-method}
\alias{axcodes,matrix-method}
\title{Get Orientation Axis Codes}
\usage{
axcodes(x)

\S4method{axcodes}{NeuroSpace}(x)

\S4method{axcodes}{NeuroObj}(x)

\S4method{axcodes}{matrix}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{NeuroVol}}, \code{\linkS4class{NeuroVec}},
\code{\linkS4class{NeuroSpace}}, or a 4x4 affine matrix.}
}
\value{
A character vector of length 3 with axis direction codes.
}
\description{
Returns a character vector of anatomical axis direction labels for a
neuroimaging object or space. For example, \code{c("R", "A", "S")} for a
standard RAS-oriented image.
}
\examples{
sp <- NeuroSpace(c(10L, 10L, 10L), c(2, 2, 2))
axcodes(sp)

vol <- DenseNeuroVol(array(0, c(10,10,10)), sp)
axcodes(vol)

}
