% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{axes}
\alias{axes}
\alias{axes,NeuroSpace-method}
\title{Extract Image Axes}
\usage{
axes(x)

\S4method{axes}{NeuroSpace}(x)
}
\arguments{
\item{x}{an object with a set of axes}
}
\value{
An object representing the \code{axes} of \code{x}.
}
\description{
Extract Image Axes
}
\examples{
x <- NeuroSpace(c(10,10,10), spacing=c(1,1,1))
class(axes(x)) == "AxisSet3D"

}
