% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{AxisSet-class}
\alias{AxisSet-class}
\title{AxisSet}
\description{
Virtual base class representing an ordered set of named axes.
}
\section{Slots}{

\describe{
\item{\code{ndim}}{the number of axes (or dimensions)}
}}

