% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{AxisSet1D}
\alias{AxisSet1D}
\title{Create a one-dimensional axis set}
\usage{
AxisSet1D(i)
}
\arguments{
\item{i}{A NamedAxis object representing the axis}
}
\value{
An AxisSet1D object
}
\description{
Create a one-dimensional axis set
}
\keyword{internal}
