% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{AxisSet2D}
\alias{AxisSet2D}
\title{Create a two-dimensional axis set}
\usage{
AxisSet2D(i, j)
}
\arguments{
\item{i}{A NamedAxis object representing the first axis}

\item{j}{A NamedAxis object representing the second axis}
}
\value{
An AxisSet2D object
}
\description{
Create a two-dimensional axis set
}
\keyword{internal}
