% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{AxisSet3D}
\alias{AxisSet3D}
\title{Create a three-dimensional axis set}
\usage{
AxisSet3D(i, j, k)
}
\arguments{
\item{i}{A NamedAxis object representing the first axis}

\item{j}{A NamedAxis object representing the second axis}

\item{k}{A NamedAxis object representing the third axis}
}
\value{
An AxisSet3D object
}
\description{
Create a three-dimensional axis set
}
\keyword{internal}
