% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BinaryReader-class}
\alias{BinaryReader-class}
\title{BinaryReader Class}
\description{
Class supporting reading of bulk binary data from a connection
}
\section{Slots}{

\describe{
\item{\code{input}}{The binary input connection}

\item{\code{byte_offset}}{The number of bytes to skip at the start of input}

\item{\code{data_type}}{The data type of the binary elements}

\item{\code{bytes_per_element}}{The number of bytes in each data element (e.g. 4 or 8 for floating point numbers)}

\item{\code{endian}}{The endianness of the binary input connection}

\item{\code{signed}}{Logical indicating whether the data type is signed}
}}

