% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{blocks}
\alias{blocks}
\title{Generic function to cut a vector-valued object into a list of sub-blocks}
\usage{
blocks(x, indices, ...)
}
\arguments{
\item{x}{the object that supplies the vector data}

\item{indices}{a vector of indices for the blocks. Must match the length of the inut vector.}

\item{...}{additional arguments}
}
\description{
Generic function to cut a vector-valued object into a list of sub-blocks
}
