% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{bootstrap_searchlight}
\alias{bootstrap_searchlight}
\title{bootstrap_searchlight}
\usage{
bootstrap_searchlight(mask, radius = 8, iter = 100)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight (can be a vector which is randomly sampled)}

\item{iter}{the total number of searchlights to sample (default is 100).}
}
\description{
Create a spherical searchlight iterator that samples regions from within a mask.
}
\details{
searchlight centers are sampled without replacement, but the same surround voxel can belong to multiple searchlight samples.
}
