% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CachedSparseNeuroVec-class}
\alias{CachedSparseNeuroVec-class}
\title{CachedSparseNeuroVec}
\description{
a sparse four-dimensional brain image, backed by a \code{matrix}, where each column represents
a non-zero vector spanning the fourth dimension (e.g. time), and defined by a volumetric mask.
The data is loaded on demand and retained in a memory-limited cache for faster repeated access.
}
\section{Slots}{

\describe{
\item{\code{meta}}{the file meta information of type \code{\linkS4class{FileMetaInfo}}}

\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVol}} is used to map between spatial and index/row coordinates}

\item{\code{bucket_vol}}{the volume contianing set of buckets or clusters used to define the data partitions}

\item{\code{cache_size}}{the maximum number of buckets to keep in cache}

\item{\code{cache}}{a sparse matrix holding the cached data}

\item{\code{cache_map}}{an array containing the array positions currently stored in memory}
}}

