% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CachedSparseNeuroVecSource-class}
\alias{CachedSparseNeuroVecSource-class}
\title{CachedSparseNeuroVecSource}
\description{
A class that is used to produce a \code{\linkS4class{CachedSparseNeuroVec}} instance
}
\section{Slots}{

\describe{
\item{\code{mask}}{the subset of voxels that will be stored in memory}

\item{\code{bucket_vol}}{the volume defining the cache partitions}

\item{\code{cache_size}}{the number of buckets to retain in memory cache.}
}}

