% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustervol.R, R/resample.R
\name{centroids}
\alias{centroids}
\alias{centroids,ClusteredNeuroVol-method}
\alias{resample,NeuroVol,NeuroVol-method}
\title{return a matrix of centroids of an object}
\usage{
centroids(x, ...)

\S4method{centroids}{ClusteredNeuroVol}(x, type = c("center_of_mass", "medoid"))

\S4method{resample}{NeuroVol,NeuroVol}(source, target, interpolation = 3L)
}
\arguments{
\item{x}{an object with multiple centroids (e.g. a \code{ClusteredNeuroVol})}

\item{...}{extra args}

\item{type}{the type of center of mass: one of "center_of_mass" or "medoid"}

\item{interpolation}{a single integer specifying the type of interpolation to be applied to the
final resampled image. May be 0 (nearest neighbour), 1 (trilinear) or 3 (cubic spline).
No other values are valid.}
}
\description{
return a matrix of centroids of an object
}
