% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgb.R
\name{cgb_smooth_loro}
\alias{cgb_smooth_loro}
\title{Leave-one-run-out smoothing helper}
\usage{
cgb_smooth_loro(runs, graphs, passes = 1L, lambda = 1)
}
\arguments{
\item{runs}{List of \code{\linkS4class{NeuroVec}} objects (one per run).}

\item{graphs}{List of graphs returned by \code{cgb_make_graph(..., leave_one_out=TRUE)}.}

\item{passes, lambda}{See \code{\link{cgb_smooth}}.}
}
\value{
A list of smoothed \code{NeuroVec} objects, one per run.
}
\description{
Leave-one-run-out smoothing helper
}
