% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{close,BinaryReader-method}
\alias{close,BinaryReader-method}
\alias{close,BinaryWriter-method}
\title{Close a BinaryReader or BinaryWriter}
\usage{
\S4method{close}{BinaryReader}(con)

\S4method{close}{BinaryWriter}(con)
}
\arguments{
\item{con}{The BinaryReader or BinaryWriter object to close.}
}
\value{
Invisibly returns NULL.
}
\description{
Closes the underlying connection associated with a BinaryReader or BinaryWriter object.
This should be called when you're done with the reader/writer to free system resources.
}
\examples{
\dontrun{
# Create and close a binary reader
reader <- BinaryReader("data.bin", byte_offset = 0L,
                      data_type = "double", bytes_per_element = 8L)
close(reader)

# Create and close a binary writer
writer <- BinaryWriter("output.bin", byte_offset = 0L,
                      data_type = "double", bytes_per_element = 8L)
close(writer)
}

}
