% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ClusteredNeuroVec-class}
\alias{ClusteredNeuroVec-class}
\title{ClusteredNeuroVec Class}
\description{
A class representing a 4D neuroimaging dataset where voxels are grouped into clusters.
Each cluster has a single time-series that is shared by all voxels within that cluster.
}
\section{Slots}{

\describe{
\item{\code{cvol}}{A \code{\linkS4class{ClusteredNeuroVol}} object defining cluster assignments}

\item{\code{ts}}{A numeric matrix of dimensions T x K (time points x clusters)}

\item{\code{cl_map}}{An integer vector mapping each voxel to its cluster ID (0 for outside mask)}

\item{\code{label}}{A character string label for the object}
}}

