% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R
\name{as,ClusteredNeuroVol,DenseNeuroVol}
\alias{as,ClusteredNeuroVol,DenseNeuroVol}
\title{Conversion from ClusteredNeuroVol to LogicalNeuroVol}
\description{
Conversion from ClusteredNeuroVol to LogicalNeuroVol
}
\keyword{internal}
