% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ColumnReader-class}
\alias{ColumnReader-class}
\title{ColumnReader}
\description{
A class that supports reading of data from a matrix-like storage format, such as a file or a database, in a column-wise manner.
}
\section{Slots}{

\describe{
\item{\code{nrow}}{An \code{integer} representing the number of rows in the matrix-like storage.}

\item{\code{ncol}}{An \code{integer} representing the number of columns in the matrix-like storage.}

\item{\code{reader}}{A \code{function} that takes a set of column indices as input and returns a \code{matrix} containing the requested columns from the storage.}
}}

