% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Compare-methods}
\alias{Compare-methods}
\alias{Compare,DenseNeuroVol,DenseNeuroVol-method}
\alias{Compare,DenseNeuroVol,numeric-method}
\alias{Compare,numeric,DenseNeuroVol-method}
\alias{Compare,SparseNeuroVol,numeric-method}
\alias{Compare,numeric,SparseNeuroVol-method}
\alias{Compare,NeuroVec,NeuroVec-method}
\title{Comparison Operations}
\usage{
\S4method{Compare}{DenseNeuroVol,DenseNeuroVol}(e1, e2)

\S4method{Compare}{DenseNeuroVol,numeric}(e1, e2)

\S4method{Compare}{numeric,DenseNeuroVol}(e1, e2)

\S4method{Compare}{SparseNeuroVol,numeric}(e1, e2)

\S4method{Compare}{numeric,SparseNeuroVol}(e1, e2)

\S4method{Compare}{NeuroVec,NeuroVec}(e1, e2)
}
\arguments{
\item{e1, e2}{Neuroimaging objects or numeric values.}
}
\value{
A \code{\linkS4class{LogicalNeuroVol}} for volume comparisons.
}
\description{
Methods for comparing neuroimaging objects.
  All volume comparisons return \code{\linkS4class{LogicalNeuroVol}} objects
  that preserve spatial metadata.
}
