% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{concat}
\alias{concat}
\title{Concatenate two objects}
\usage{
concat(x, y, ...)
}
\arguments{
\item{x}{the first object, typically \code{NeuroVol} or \code{BrainVector}}

\item{y}{the second object, typically \code{NeuroVol} or \code{BrainVector}}

\item{...}{additional objects}
}
\description{
Concatenate two objects
}
\details{
The \code{x} and \code{y} images must have compatible dimensions. a \code{NeuroVol} can be concatenated to \code{BrainVector}, and vice versa. See examples.
}
\examples{
bv1 <- NeuroVol(rep(1,1000), NeuroSpace(c(10,10,10), c(1,1,1)))
bv2 <- NeuroVol(rep(2,1000), NeuroSpace(c(10,10,10), c(1,1,1)))
bv3 <- concat(bv1,bv2)
inherits(bv3, "BrainVector")

bv4 <- concat(bv3, bv1)
dim(bv4)[4] == 3
bv5 <- concat(bv1, bv3)
dim(bv4)[4] == 3

bv6 <- concat(bv4,bv5)
dim(bv6)[4] == 6

}

