% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{conn_comp}
\alias{conn_comp}
\title{Find connected components}
\usage{
conn_comp(x, ...)
}
\arguments{
\item{x}{the image object}

\item{...}{additonal arguments}
}
\description{
Find connected components
}

