% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{coords}
\alias{coords}
\title{Extract coordinates}
\usage{
coords(x, ...)
}
\arguments{
\item{x}{the object to extract coordinates from}

\item{...}{additional arguments}
}
\description{
Extract coordinates
}
