% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{data_reader}
\alias{data_reader}
\alias{data_reader,NIFTIMetaInfo-method}
\alias{data_reader,AFNIMetaInfo-method}
\title{Generic function to create data reader}
\usage{
data_reader(x, offset)

\S4method{data_reader}{NIFTIMetaInfo}(x, offset = 0)

\S4method{data_reader}{AFNIMetaInfo}(x, offset = 0)
}
\arguments{
\item{x}{an object specifying the infromation required to produce the reader}

\item{offset}{the byte offset (number of bytes to skip before reading)}
}
\description{
Generic function to create data reader
}
