% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{DenseNeuroVec,H5NeuroVec}
\alias{DenseNeuroVec,H5NeuroVec}
\title{Convert DenseNeuroVec to H5NeuroVec}
\arguments{
\item{from}{A DenseNeuroVec object.}
}
\value{
An H5NeuroVec object resulting from the conversion.
}
\description{
This function converts a DenseNeuroVec object to an H5NeuroVec object.
}
\keyword{internal}
