% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{DenseNeuroVol,H5NeuroVol}
\alias{DenseNeuroVol,H5NeuroVol}
\title{Convert DenseNeuroVol to H5NeuroVol}
\arguments{
\item{from}{A DenseNeuroVol object.}
}
\value{
An H5NeuroVol object resulting from the conversion.
}
\description{
This function converts a DenseNeuroVol object to an H5NeuroVol object.
}
\keyword{internal}
