% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{dim,FileMetaInfo-method}
\alias{dim,FileMetaInfo-method}
\title{Get Dimensions of FileMetaInfo Object}
\usage{
\S4method{dim}{FileMetaInfo}(x)
}
\arguments{
\item{x}{A FileMetaInfo object}
}
\value{
Integer vector of image dimensions
}
\description{
Get Dimensions of FileMetaInfo Object
}
