% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_obj.R
\name{dim,NeuroObj-method}
\alias{dim,NeuroObj-method}
\title{dim of \code{NeuroObj} object}
\usage{
\S4method{dim}{NeuroObj}(x)
}
\arguments{
\item{x}{the object}
}
\description{
dim of \code{NeuroObj} object
}
