% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{dim,NeuroHyperVec-method}
\alias{dim,NeuroHyperVec-method}
\title{Get dimensions of a NeuroHyperVec object}
\usage{
\S4method{dim}{NeuroHyperVec}(x)
}
\arguments{
\item{x}{A NeuroHyperVec object}
}
\value{
A vector of dimensions
}
\description{
Get dimensions of a NeuroHyperVec object
}
