% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{drop,NeuroVec-method}
\alias{drop,NeuroVec-method}
\title{Drop a dimension}
\usage{
\S4method{drop}{NeuroVec}(x)
}
\arguments{
\item{x}{the object to drop a dimension from}
}
\value{
An object of the same class as x with reduced dimensions or elements.
}
\description{
Drop a dimension
}
