% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{drop,NeuroVec-method}
\alias{drop,NeuroVec-method}
\title{drop}
\usage{
\S4method{drop}{NeuroVec}(x)

\S4method{drop}{NeuroVec}(x)
}
\arguments{
\item{x}{The NeuroVec object.}
}
\value{
If the last dimension of the NeuroVec object is of length 1, a DenseNeuroVol
  object is returned. Otherwise, the original NeuroVec object is returned.

If the last dimension of the NeuroVec object is of length 1, a DenseNeuroVol
  object is returned. Otherwise, the original NeuroVec object is returned.
}
\description{
This function drops the last dimension of a NeuroVec object if it is of length 1.

This function drops the last dimension of a NeuroVec object if it is of length 1.
}
