% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{drop}
\alias{drop}
\title{Generic Drop Method}
\usage{
drop(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An object of the same class as \code{x} with reduced dimensions or elements.
}
\description{
Provides a mechanism to remove dimensions or elements from an object.
}
