% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{drop_dim,AxisSet2D,numeric-method}
\alias{drop_dim,AxisSet2D,numeric-method}
\title{Drop dimension from axis set}
\usage{
\S4method{drop_dim}{AxisSet2D,numeric}(x, dimnum)
}
\arguments{
\item{x}{An AxisSet2D or AxisSet3D object}

\item{dimnum}{Numeric index of dimension to drop}
}
\description{
Drop dimension from axis set
}
