% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{drop_dim,AxisSet3D,missing-method}
\alias{drop_dim,AxisSet3D,missing-method}
\title{Drop dimension from axis set}
\usage{
\S4method{drop_dim}{AxisSet3D,missing}(x, dimnum)
}
\arguments{
\item{x}{An AxisSet3D object}

\item{dimnum}{Numeric index of dimension to drop (optional)}
}
\value{
An axis set with one fewer dimension
}
\description{
Drop dimension from axis set
}
\examples{
\dontrun{
axes3d <- AxisSet3D(LEFT_RIGHT, ANT_POST, INF_SUP)
axes2d <- drop_dim(axes3d)  # Drop first dimension
}
}
