% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{drop_dim}
\alias{drop_dim}
\title{Generic function to drop a dimension from an object}
\usage{
drop_dim(x, dimnum)
}
\arguments{
\item{x}{a dimensioned object}

\item{dimnum}{the index of the dimension to drop}
}
\description{
Generic function to drop a dimension from an object
}
\examples{
x = NeuroSpace(c(10,10,10), c(1,1,1))
x1 <- dropDim(x)
ndim(x1) == 2
dim(x1)[2] == 10
}

