% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_extensions.R
\name{extension}
\alias{extension}
\alias{extension,NiftiExtensionList,numeric-method}
\title{Get Extension by Code}
\usage{
extension(x, ecode)

\S4method{extension}{NiftiExtensionList,numeric}(x, ecode)
}
\arguments{
\item{x}{A \code{\link{NiftiExtensionList-class}} object.}

\item{ecode}{Integer extension code to filter by.}
}
\value{
A \code{\link{NiftiExtensionList-class}} containing only extensions
  with the specified code.
}
\description{
Retrieve extensions with a specific extension code from a list.
}
