% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_like.R, R/neurovec.R, R/neurovol.R,
%   R/roi.R, R/sparse_neurovec.R
\name{[,DenseNeuroVol,numeric,missing,ANY-method}
\alias{[,DenseNeuroVol,numeric,missing,ANY-method}
\alias{[,DenseNeuroVol,integer,missing,ANY-method}
\alias{[[,NeuroVec,numeric-method}
\alias{[[,NeuroVec,character-method}
\alias{[,NeuroVol,ROICoords,missing,ANY-method}
\alias{[,NeuroVol,ROIVol,missing,ANY-method}
\alias{[,DenseNeuroVol,ROIVol,missing,ANY-method}
\alias{[,SparseNeuroVol,numeric,numeric,ANY-method}
\alias{[}
\alias{[,ROICoords,numeric,missing,ANY-method}
\alias{[,ROIVol,ROICoords,missing,ANY-method}
\alias{[,ROIVol,ROICoords,numeric,ANY-method}
\alias{[,ROIVol,logical,missing,ANY-method}
\alias{[,ROIVol,logical,numeric,ANY-method}
\alias{[,ROIVol,matrix,missing,ANY-method}
\alias{[,ROIVol,matrix,numeric,ANY-method}
\alias{[,ROIVol,missing,missing,ANY-method}
\alias{[,ROIVol,missing,numeric,ANY-method}
\alias{[,ROIVol,numeric,numeric,ANY-method}
\alias{[,ROIVol,numeric,missing,ANY-method}
\alias{[,AbstractSparseNeuroVec,numeric,numeric,ANY-method}
\title{[[}
\usage{
\S4method{[}{DenseNeuroVol,numeric,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{DenseNeuroVol,integer,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[[}{NeuroVec,numeric}(x, i)

\S4method{[[}{NeuroVec,character}(x, i)

\S4method{[}{NeuroVol,ROICoords,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NeuroVol,ROIVol,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{DenseNeuroVol,ROIVol,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{SparseNeuroVol,numeric,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ROIVol,logical,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,ROICoords,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,matrix,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,missing,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,missing,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,numeric,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,logical,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,ROICoords,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,matrix,numeric,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROICoords,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{NeuroVol,ROICoords,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{AbstractSparseNeuroVec,numeric,numeric,ANY}(x, i, j, k, m, ..., drop = TRUE)
}
\arguments{
\item{x}{An object of class \code{AbstractSparseNeuroVec}}

\item{i}{Numeric vector specifying the indices for the first dimension}

\item{j}{Numeric vector specifying the indices for the second dimension}

\item{k}{Numeric vector specifying the indices for the third dimension (optional)}

\item{...}{Additional arguments passed to methods}

\item{drop}{Logical indicating whether to drop dimensions of length one (default: TRUE)}

\item{m}{Numeric vector specifying the indices for the fourth dimension (optional)}
}
\value{
a DenseNeuroVol object

A subset of the input object, with dimensions depending on the indexing and the `drop` parameter.

An array containing the extracted subset
}
\description{
This function extracts a single volume from a NeuroVec object.

Extracts a subset of data from a sparse four-dimensional brain image based on provided indices.
}
