% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R
\docType{methods}
\name{[[,FileBackedNeuroVec,numeric-method}
\alias{[[,FileBackedNeuroVec,numeric-method}
\title{[[}
\usage{
\S4method{[[}{FileBackedNeuroVec,numeric}(x, i)
}
\arguments{
\item{i}{the volume index}
}
\description{
[[
}
