% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FileSource-class}
\alias{FileSource-class}
\title{FileSource}
\description{
Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{meta information for the data source}
}}

