% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_filter.R
\name{guided_filter}
\alias{guided_filter}
\title{Filter a volumetric image with a "guided" filter}
\usage{
guided_filter(vol, radius = 4, epsilon = 0.7^2)
}
\arguments{
\item{vol}{the image volume as a \code{NeuroVol}}

\item{radius}{the spatial radius of the filter}

\item{epsilon}{variance constant}
}
\value{
a filtered image of class \code{NeuroVol}
}
\description{
Filter a volumetric image with a "guided" filter
}
