% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5NeuroVec-class}
\alias{H5NeuroVec-class}
\title{H5NeuroVec}
\description{
A class representing a four-dimensional brain image (NeuroVec) backed by an HDF5 file. H5NeuroVec objects facilitate efficient handling of large-scale brain imaging data by storing and accessing the data using the HDF5 file format.
}
\section{Slots}{

\describe{
\item{\code{obj}}{An instance of class \code{H5File} from the \code{hdf5r} package, representing the underlying HDF5 file.}
}}

