% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{H5NeuroVecSource-class}
\alias{H5NeuroVecSource-class}
\title{H5NeuroVecSource}
\description{
A class that is used to produce a \code{\linkS4class{H5NeuroVecSource}} instance
}
\section{Slots}{

\describe{
\item{\code{file_name}}{the name of the hdf5 file.}
}}

\keyword{internal}
