% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_extensions.R
\name{has_extensions}
\alias{has_extensions}
\alias{has_extensions,NiftiExtensionList-method}
\alias{has_extensions,list-method}
\title{Check if Extensions are Present}
\usage{
has_extensions(x)

\S4method{has_extensions}{NiftiExtensionList}(x)

\S4method{has_extensions}{list}(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical indicating whether extensions are present.
}
\description{
Tests whether an object has any NIfTI extensions.
}
