% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{image}
\alias{image}
\title{Generic Image Method for Creating Visual Representations}
\usage{
image(x, ...)
}
\arguments{
\item{x}{An object to be rendered as an image.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An image object representing \code{x}.
}
\description{
Creates a visual representation (or image) from an object.
}
