% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{index_to_grid}
\alias{index_to_grid}
\title{Generic function to convert 1D indices to N-dimensional grid coordinates}
\usage{
index_to_grid(x, idx)
}
\arguments{
\item{x}{the object}

\item{idx}{the 1D \code{vector} of indices}
}
\value{
a matrix of grid coordinates
}
\description{
Generic function to convert 1D indices to N-dimensional grid coordinates
}
\examples{

 bvol <- NeuroVol(array(0, c(10,10,10)), NeuroSpace(c(10,10,10), c(1,1,1)))
 idx <- 1:10
 g <- index_to_grid(bvol, idx)
 bvol[g]

}

