% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{IndexLookupVolume-class}
\alias{IndexLookupVolume-class}
\title{IndexLookupVolume}
\description{
Three-dimensional brain image that can be used as a map between 1D grid indices and a table of values
Currently used in the \code{\linkS4class{SparseNeuroVec}} class.
}

