% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{indices}
\alias{indices}
\title{Extract indices}
\usage{
indices(x)
}
\arguments{
\item{x}{the object to extract indices}
}
\value{
A vector of indices from the object.
}
\description{
Extract indices
}
