% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{.isExtension}
\alias{.isExtension}
\alias{.concat4D}
\alias{.gridToIndex3D}
\alias{.gridToIndex}
\alias{.indexToGrid}
\alias{.getRStorage}
\alias{.getMMapMode}
\alias{.getDataStorage}
\alias{.getDataCode}
\alias{.getDataSize}
\alias{.getEndian}
\alias{.niftiExt}
\alias{.matrixToQuatern}
\alias{.quaternToMatrix}
\title{.isExtension}
\usage{
.isExtension(fname, extension)

.concat4D(x, y, ...)

.gridToIndex3D(dimensions, voxmat)

.gridToIndex(dimensions, vmat)

.indexToGrid(idx, array.dim)

.getRStorage(data_type)

.getMMapMode(code)

.getDataStorage(code)

.getDataCode(data_type)

.getDataSize(data_type)

.getEndian(conn)

.niftiExt(filetype)

.matrixToQuatern(mat)

.quaternToMatrix(quat, origin, stepSize, qfac)
}
\description{
.isExtension

.concat4D

.gridToIndex3D

.gridToIndex

.indexToGrid

.getRStorage

.getDataStorage

.getDataCode

.getDataSize

.getEndian
}
\keyword{internal}
