% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{inverse_trans}
\alias{inverse_trans}
\title{Generic getter to extract inverse image coordinate transformation}
\usage{
inverse_trans(x)
}
\arguments{
\item{x}{an object}
}
\description{
Generic getter to extract inverse image coordinate transformation
}
\examples{
bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
itrans <- inverseTrans(bspace)
identical(trans(bspace) \%*\% inverseTrans(bspace), diag(4))
}

