% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_neurovec.R
\name{labels,ClusteredNeuroVec-method}
\alias{labels,ClusteredNeuroVec-method}
\title{Get Labels from ClusteredNeuroVec}
\usage{
\S4method{labels}{ClusteredNeuroVec}(object)
}
\arguments{
\item{object}{A ClusteredNeuroVec object}
}
\description{
Get Labels from ClusteredNeuroVec
}
