% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/latent_neurovec.R
\docType{class}
\name{LatentNeuroVec-class}
\alias{LatentNeuroVec-class}
\alias{LatentNeuroVec}
\title{LatentNeuroVec}
\usage{
LatentNeuroVec(basis, loadings, space, mask, offset = NULL)
}
\arguments{
\item{basis}{an n-by-k matrix containing the latent vectors forming the reduced space}

\item{loadings}{a p-by-k matrix of p loadings}

\item{space}{a NeuroSpace instance}

\item{mask}{a 3D \code{array}, 1D \code{vector} of type \code{logical}, or an instance of type \code{LogicalNeuroVol}}

\item{centroid}{a 1-by-p offset vector}
}
\description{
a class that stores a represents a 4-dimensional array as a set of basis functions (dictionary) and
corresponding set of loadings (coefficients).

constructs a LatentNeuroVec object
}
\section{Slots}{

\describe{
\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{basis}}{the matrix of bases, were each column is a basis vector.}

\item{\code{loadings}}{the \code{sparseMatrix} of loadings}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVol}} is used to map between spatial and index/row coordinates}

\item{\code{offset}}{an offset vector}
}}

\examples{

bspace <- NeuroSpace(c(10,10,10,100), c(1,1,1))
mask <- array(rnorm(10*10*10) > .5, c(10,10,10))
mat <- matrix(rnorm(sum(mask)), 100, sum(mask))
svec <- SparseNeuroVec(mat, bspace,mask)
length(indices(svec)) == sum(mask)
}
