% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{LatentNeuroVecSource-class}
\alias{LatentNeuroVecSource-class}
\title{LatentNeuroVecSource}
\description{
A class that is used to produce a \code{\linkS4class{LatentNeuroVecSource}} instance
}
\section{Slots}{

\describe{
\item{\code{file_name}}{the name of the hdf5 file.}
}}

