% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{length,NeuroHyperVec-method}
\alias{length,NeuroHyperVec-method}
\title{Get the number of NeuroVec objects in a NeuroHyperVec}
\usage{
\S4method{length}{NeuroHyperVec}(x)
}
\arguments{
\item{x}{A NeuroHyperVec object}
}
\value{
The number of NeuroVec objects
}
\description{
Get the number of NeuroVec objects in a NeuroHyperVec
}
