% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R
\name{linear_access,FileBackedNeuroVec,numeric-method}
\alias{linear_access,FileBackedNeuroVec,numeric-method}
\title{Linear Access Method for FileBackedNeuroVec}
\usage{
\S4method{linear_access}{FileBackedNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{A FileBackedNeuroVec object}

\item{i}{A numeric vector of indices}
}
\value{
A numeric vector of values
}
\description{
Linear Access Method for FileBackedNeuroVec
}
