% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{linear_access,NeuroHyperVec,ANY-method}
\alias{linear_access,NeuroHyperVec,ANY-method}
\title{Linear access method for NeuroHyperVec}
\usage{
\S4method{linear_access}{NeuroHyperVec,ANY}(x, i, ...)
}
\arguments{
\item{x}{The NeuroHyperVec object}

\item{i}{The linear indices}

\item{...}{Additional arguments (not used)}
}
\value{
A numeric vector or array
}
\description{
Linear access method for NeuroHyperVec
}
