% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_extensions.R
\name{list_afni_attributes}
\alias{list_afni_attributes}
\title{List AFNI Attributes in Extension}
\usage{
list_afni_attributes(ext)
}
\arguments{
\item{ext}{A \code{\link{NiftiExtension-class}} object with ecode = 4,
or an xml_document from \code{\link{parse_afni_extension}}.}
}
\value{
Character vector of attribute names.
}
\description{
Returns a character vector of all attribute names in an AFNI extension.
}
\examples{
\dontrun{
# List all attributes in an AFNI extension
attrs <- list_afni_attributes(afni_ext)
print(attrs)
}

}
