% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{load_data}
\alias{load_data}
\title{Generic function to load data from a data source}
\usage{
load_data(x, ...)
}
\arguments{
\item{x}{a data source}

\item{...}{additional arguments}
}
\description{
Generic function to load data from a data source
}
