% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{loadVolumeList}
\alias{loadVolumeList}
\title{loadVolumeList}
\usage{
loadVolumeList(file_names, mask = NULL)
}
\arguments{
\item{file_names}{a list of files to load}

\item{mask}{an optional mask indicating subset of voxels to load}
}
\value{
an instance of class \code{\linkS4class{NeuroVec}}
}
\description{
load a list of image volumes and return a \code{\linkS4class{NeuroVec}} instance
}
