% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Logic-methods}
\alias{Logic-methods}
\alias{Logic,DenseNeuroVol,DenseNeuroVol-method}
\alias{Logic,SparseNeuroVol,SparseNeuroVol-method}
\alias{Logic,SparseNeuroVol,NeuroVol-method}
\alias{Logic,NeuroVol,SparseNeuroVol-method}
\alias{Logic,NeuroVol,logical-method}
\alias{Logic,logical,NeuroVol-method}
\title{Logic Operations for Neuroimaging Volumes}
\usage{
\S4method{Logic}{DenseNeuroVol,DenseNeuroVol}(e1, e2)

\S4method{Logic}{SparseNeuroVol,SparseNeuroVol}(e1, e2)

\S4method{Logic}{SparseNeuroVol,NeuroVol}(e1, e2)

\S4method{Logic}{NeuroVol,SparseNeuroVol}(e1, e2)

\S4method{Logic}{NeuroVol,logical}(e1, e2)

\S4method{Logic}{logical,NeuroVol}(e1, e2)
}
\arguments{
\item{e1, e2}{Neuroimaging volume objects or logical values.}
}
\value{
A \code{\linkS4class{LogicalNeuroVol}}.
}
\description{
Methods for performing logical operations (\code{&} and
  \code{|}) on neuroimaging volume objects. Results are always returned as
  \code{\linkS4class{LogicalNeuroVol}} objects that preserve spatial metadata.
}
\examples{
sp <- NeuroSpace(c(5L, 5L, 5L), c(1, 1, 1))
v1 <- DenseNeuroVol(array(sample(0:1, 125, replace = TRUE), c(5, 5, 5)), sp)
v2 <- DenseNeuroVol(array(sample(0:1, 125, replace = TRUE), c(5, 5, 5)), sp)
intersection <- v1 & v2
union_mask  <- v1 | v2

}
