% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neurovol.R
\docType{class}
\name{LogicalNeuroVol-class}
\alias{LogicalNeuroVol-class}
\alias{LogicalNeuroVol}
\title{LogicalNeuroVol}
\usage{
LogicalNeuroVol(data, space, label = "", indices = NULL)
}
\arguments{
\item{data}{a three-dimensional \code{array}, a 1D vector with length equal to \code{prod(dim(space))}, or a set of \code{indices} where elements are \code{TRUE}}

\item{space}{an instance of class \code{\linkS4class{NeuroSpace}}}

\item{label}{a \code{character} string}

\item{indices}{an optional 1-d index vector}
}
\value{
\code{\linkS4class{LogicalNeuroVol}} instance
}
\description{
Three-dimensional brain image where all values are either TRUE or FALSE

Construct a \code{\linkS4class{LogicalNeuroVol}} instance
}
