% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{make_vector}
\alias{make_vector}
\title{make_vector}
\usage{
make_vector(data, refdata, label = "")
}
\arguments{
\item{data}{a four-dimensional \code{array}}

\item{refdata}{an instance of class \code{\linkS4class{NeuroVec}} or \code{\linkS4class{NeuroVol}} containing the reference space for the new vector.}

\item{label}{a \code{character} string}
}
\value{
\code{\linkS4class{DenseNeuroVec}} instance
}
\description{
Construct a \code{\linkS4class{NeuroVec}} instance, using default (dense) implementation
}
