% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{matchAnatomy2D}
\alias{matchAnatomy2D}
\title{Find matching 2D anatomical orientation}
\usage{
matchAnatomy2D(axis1, axis2)
}
\arguments{
\item{axis1}{First NamedAxis object}

\item{axis2}{Second NamedAxis object}
}
\value{
Matching AxisSet2D object or NULL if no match found
}
\description{
Searches through pre-defined 2D orientations to find a match for given axes.
}
\keyword{internal}
