% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{matchAnatomy3D}
\alias{matchAnatomy3D}
\title{Find matching 3D anatomical orientation}
\usage{
matchAnatomy3D(axis1, axis2, axis3)
}
\arguments{
\item{axis1}{First NamedAxis object}

\item{axis2}{Second NamedAxis object}

\item{axis3}{Third NamedAxis object}
}
\value{
Matching AxisSet3D object or NULL if no match found
}
\description{
Searches through pre-defined 3D orientations to find a match for given axes.
}
\keyword{internal}
