% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{matchAxis}
\alias{matchAxis}
\title{Match anatomical axis abbreviation to full axis object}
\usage{
matchAxis(firstAxis)
}
\arguments{
\item{firstAxis}{Character string specifying the axis. Can be full name
(e.g. "LEFT") or abbreviation (e.g. "L")}
}
\value{
A NamedAxis object corresponding to the specified axis
}
\description{
Match anatomical axis abbreviation to full axis object
}
\examples{
\dontrun{
matchAxis("L")  # Returns LEFT_RIGHT axis
matchAxis("ANTERIOR")  # Returns ANT_POST axis
}
}
\keyword{internal}
