% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_neurovec.R
\name{matricized_access,LatentNeuroVec,matrix-method}
\alias{matricized_access,LatentNeuroVec,matrix-method}
\title{@rdname matricized_access-methods}
\usage{
\S4method{matricized_access}{LatentNeuroVec,matrix}(x, i)
}
\description{
@rdname matricized_access-methods
}
