% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{mean-methods}
\alias{mean-methods}
\alias{mean,DenseNeuroVec-method}
\alias{mean,SparseNeuroVec-method}
\alias{mean,NeuroVec-method}
\title{Temporal Mean of a NeuroVec}
\usage{
\S4method{mean}{DenseNeuroVec}(x, ...)

\S4method{mean}{SparseNeuroVec}(x, ...)

\S4method{mean}{NeuroVec}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{NeuroVec}} object.}

\item{...}{Ignored.}
}
\value{
A \code{\linkS4class{NeuroVol}} containing the temporal mean at
  each voxel.
}
\description{
Computes the voxel-wise mean across the 4th dimension (time), returning
a 3D \code{\linkS4class{DenseNeuroVol}} or \code{\linkS4class{SparseNeuroVol}}.
}
\examples{
bspace <- NeuroSpace(c(10, 10, 10, 20), c(1, 1, 1))
dat <- array(rnorm(10 * 10 * 10 * 20), c(10, 10, 10, 20))
vec <- DenseNeuroVec(dat, bspace)
mean_vol <- mean(vec)
dim(mean_vol)  # 10 10 10

}
