% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{MetaInfo-class}
\alias{MetaInfo-class}
\title{MetaInfo}
\description{
This class contains meta information for an image data type
}
\section{Slots}{

\describe{
\item{\code{data_type}}{the data type code, e.g. FLOAT}

\item{\code{dims}}{image dimensions}

\item{\code{spatial_axes}}{image axes for spatial dimensions (x,y,z)}

\item{\code{additional_axes}}{axes for dimensions > 3 (e.g. time, color band, direction)}

\item{\code{spacing}}{voxel dimensions}

\item{\code{origin}}{coordinate origin}

\item{\code{label}}{name(s) of images}
}}

