% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroslice.R
\name{neuro_slice}
\alias{neuro_slice}
\title{neuro_slice constructor}
\usage{
neuro_slice(data, space, indices = NULL)
}
\arguments{
\item{data}{data vector or matrix}

\item{space}{an instance of class BrainSpace}

\item{indices}{linear indices corresponding used if \code{data} is a 1D vector.}
}
\description{
neuro_slice constructor
}
\examples{
bspace <- BrainSpace(c(64,64), spacing=c(1,1))
dat <- array(rnorm(64*64), c(64,64))
bslice <- neuro_slice(dat,bspace)
print(bslice)
}
