% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurospace.R
\name{NeuroSpace}
\alias{NeuroSpace}
\title{Constructor function for \code{\linkS4class{NeuroSpace}} class}
\usage{
NeuroSpace(dim, spacing = NULL, origin = NULL, axes = NULL, trans = NULL)
}
\arguments{
\item{dim}{An integer vector describing the dimensions of the image grid.}

\item{spacing}{A numeric vector representing the real-valued voxel dimensions (e.g., in millimeters). If not provided, the default is set to a vector of ones with the same length as the dimensions.}

\item{origin}{A numeric vector representing the coordinate origin of the image space. If not provided, the default is set to a vector of zeroes with the same length as the dimensions.}

\item{axes}{An \code{\linkS4class{AxisSet}} object representing the image axes ordering. If not provided, the default axes are determined based on the NIFTI standard (Left-Posterior-Inferior).}

\item{trans}{A matrix representing the coordinate transformation associated with the image space. If not provided, the default is based on the NIFTI standard ("Neurological" orientation).}
}
\value{
An instance of the \code{\linkS4class{NeuroSpace}} class.
}
\description{
Constructor function for \code{\linkS4class{NeuroSpace}} class
}
\note{
Users rarely need to create a new \code{NeuroSpace} instance, as it will almost always be created automatically using information stored in an image header. If an existing image object is available, its \code{NeuroSpace} instance can be easily extracted with the \code{space} method.
}
\examples{
bspace <- NeuroSpace(c(64, 64, 64), origin = c(0, 0, 0), spacing = c(2, 2, 2))
bspace
origin(bspace)
axes(bspace)
trans(bspace)

}
