% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroim.R
\docType{package}
\name{neuroim2-package}
\alias{neuroim2-package}
\alias{neuroim2}
\title{neuroim2: neuroimaging data structures for analysis}
\description{
The neuroim2 package provides tools and functions for analyzing and 
manipulating neuroimaging data. It supports various neuroimaging formats 
and offers a range of analysis techniques.
}
\section{Main functions}{

\itemize{
  \item \code{\link{read_vol}}: Read neuroimaging volumes
  \item \code{\link{write_vol}}: Write neuroimaging volumes
  \item \code{\link{NeuroVol}}: Create NeuroVol objects
  \item \code{\link{NeuroVec}}: Create NeuroVec objects
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/neuroim2}
  \item \url{https://bbuchsbaum.github.io/neuroim2/}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neuroim2/issues}
}

}
\author{
\strong{Maintainer}: Bradley R Buchsbaum \email{brad.buchsbaum@gmail.com} [copyright holder]

}
