% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroim.R
\docType{package}
\name{neuroim2}
\alias{neuroim2-package}
\alias{neuroim2}
\title{neuroim2}
\description{
Data structures for neuroimaging data.
}
\details{
none
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/neuroim2}
  \item \url{https://bbuchsbaum.github.io/neuroim2/}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neuroim2/issues}
}

}
\author{
\strong{Maintainer}: Bradley R Buchsbaum \email{brad.buchsbaum@gmail.com} [copyright holder]

}
