% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroSpace-class}
\alias{NeuroSpace-class}
\title{NeuroSpace}
\description{
This class represents the geometry of a brain image
}
\section{Slots}{

\describe{
\item{\code{dim}}{the grid dimensions of the image}

\item{\code{origin}}{the coordinates of the spatial origin}

\item{\code{spacing}}{the dimensions (in mm) of the grid units (voxels)}

\item{\code{axes}}{the set of named spatial axes in the untransformed native grid space.}

\item{\code{trans}}{an affine transformation matrix that moves from grid -> real world coordinates}

\item{\code{inverse}}{an inverse matrix that moves from real world -> grid coordinates}
}}

