% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neurovec.R
\docType{class}
\name{NeuroVec-class}
\alias{NeuroVec-class}
\alias{NeuroVec}
\title{NeuroVec}
\usage{
NeuroVec(data, space = NULL, mask = NULL, label = "")
}
\arguments{
\item{data}{The image data, which can be a matrix, a 4D array, or a list of \code{NeuroVol} objects. If the latter, the geometric space of the data (\code{NeuroSpace}) will be inferred from the constituent volumes, which must all be identical.}

\item{space}{An optional \code{\linkS4class{NeuroSpace}} object. Does not need to be included if the \code{data} argument is a list of \code{NeuroVol} objects.}

\item{mask}{An optional \code{array} of type \code{logical}.}

\item{label}{A label of type \code{character}.}
}
\value{
A concrete instance of the \code{\linkS4class{NeuroVec}} class. If a \code{mask} is provided, the function returns a \code{\linkS4class{SparseNeuroVec}} object; otherwise, it returns a \code{\linkS4class{DenseNeuroVec}} object.
}
\description{
This S4 class represents a four-dimensional brain image, which can be used to store and process time series data such as fMRI or 4D functional connectivity maps. The class extends the basic functionality of \code{NeuroObj}.

Constructor function for the \code{\linkS4class{NeuroVec}} class. This function is used to create an instance of a NeuroVec object, which represents a four-dimensional brain image.
}
\section{Usage}{

To create a NeuroVec object, use the constructor function \code{NeuroVec()}.
}

\examples{
# Load an example 4D brain image
example_4d_image <- read_vol(system.file("extdata", "global_mask_v4.nii", package="neuroim2"))

# Create a NeuroVec object
neuro_vec <- NeuroVec(data=example_4d_image, space=space(example_4d_image))

# Load an example 4D brain image
example_4d_image <- read_vol(system.file("extdata", "global_mask_v4.nii", package="neuroim2"))

# Create a NeuroVec object
neuro_vec <- NeuroVec(data=example_4d_image, space=space(example_4d_image))

}
