% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{H5NeuroVec}
\alias{H5NeuroVec}
\title{H5NeuroVec}
\usage{
H5NeuroVec(file_name)
}
\arguments{
\item{file_name}{name of the 4-dimensional image in \code{neuroim2} hdf5 format}
}
\value{
an \code{\linkS4class{H5NeuroVec}} instance
}
\description{
Construct a \code{\linkS4class{H5NeuroVec}} object
}
