% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeurVecSeq-class}
\alias{NeurVecSeq-class}
\title{NeuroVecSeq}
\description{
A concatenated sequence of \code{\linkS4class{NeuroVec}} instances.
}
\section{Slots}{

\describe{
\item{\code{vecs}}{the sequences of \code{NeuroVec} instances}
}}

