% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{NeuroVecSeq}
\alias{NeuroVecSeq}
\title{Create a NeuroVecSeq Instance}
\usage{
NeuroVecSeq(...)
}
\arguments{
\item{...}{One or more instances of type \link{NeuroVec}.}
}
\value{
A NeuroVecSeq object containing the provided NeuroVec objects, along with the associated space and length information.
}
\description{
Constructs a NeuroVecSeq object to represent a variable-length list of NeuroVec objects.
}
\examples{
v1 <- NeuroVec(array(0, c(5, 5, 5, 2)), space = NeuroSpace(dim = c(5, 5, 5, 2)))
v2 <- NeuroVec(array(0, c(5, 5, 5, 4)), space = NeuroSpace(dim = c(5, 5, 5, 4)))
v3 <- NeuroVec(array(0, c(5, 5, 5, 6)), space = NeuroSpace(dim = c(5, 5, 5, 6)))
vs <- NeuroVecSeq(v1, v2, v3)
blks <- split_blocks(vs, rep(1:3, each = 4))
}
