% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroVecSource-class}
\alias{NeuroVecSource-class}
\title{NeuroVecSource Class}
\description{
A class used to produce a \code{\linkS4class{NeuroVec}} instance.
}
\section{Slots}{

\describe{
\item{\code{indices}}{An \code{integer} vector representing the indices of the volumes to be loaded.}
}}

\seealso{
\code{\link{FileSource-class}}, \code{\link{NeuroVec-class}}
}
