% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroVolSource-class}
\alias{NeuroVolSource-class}
\title{NeuroVolSource}
\description{
A class is used to produce a \code{\linkS4class{NeuroVol}} instance
}
\section{Slots}{

\describe{
\item{\code{index}}{the index of the volume to be read -- must be of length 1.}
}}

