% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{niftiDim}
\alias{niftiDim}
\title{Extract NIFTI Dimensions}
\usage{
niftiDim(nifti_header)
}
\arguments{
\item{nifti_header}{NIFTI header list}
}
\value{
Vector of image dimensions
}
\description{
Extract NIFTI Dimensions
}
\keyword{internal}
