% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{NIFTIMetaInfo}
\alias{NIFTIMetaInfo}
\title{Constructor for \code{\linkS4class{NIFTIMetaInfo}} class}
\usage{
NIFTIMetaInfo(descriptor, nifti_header)
}
\arguments{
\item{descriptor}{an instance of class \code{\linkS4class{NIFTIFormat}}}

\item{nifti_header}{a \code{list} returned by \code{readNIftiHeader}}
}
\value{
an instance of class \code{\linkS4class{NIFTIMetaInfo}}
}
\description{
Constructor for \code{\linkS4class{NIFTIMetaInfo}} class
}
