% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{not-methods}
\alias{not-methods}
\alias{!,DenseNeuroVol-method}
\alias{!,SparseNeuroVol-method}
\title{Logical Negation for Neuroimaging Volumes}
\usage{
\S4method{!}{DenseNeuroVol}(x)

\S4method{!}{SparseNeuroVol}(x)
}
\arguments{
\item{x}{A neuroimaging volume object.}
}
\value{
A \code{\linkS4class{LogicalNeuroVol}}.
}
\description{
Logical negation (\code{!}) for neuroimaging volumes. Returns a
  \code{\linkS4class{LogicalNeuroVol}} where non-zero voxels become
  \code{FALSE} and zero voxels become \code{TRUE}.
}
\examples{
sp <- NeuroSpace(c(5L, 5L, 5L), c(1, 1, 1))
mask <- LogicalNeuroVol(array(sample(c(TRUE, FALSE), 125, replace = TRUE),
                              c(5, 5, 5)), sp)
inv <- !mask

}
