% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{NullAxis}
\alias{NullAxis}
\title{Pre-defined null axis set}
\format{
An object of class \code{AxisSet} of length 1.
}
\usage{
NullAxis
}
\description{
Pre-defined null axis set
}
\keyword{datasets}
