% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R
\name{num_clusters,ClusteredNeuroVol-method}
\alias{num_clusters,ClusteredNeuroVol-method}
\title{Number of Clusters}
\usage{
\S4method{num_clusters}{ClusteredNeuroVol}(x)
}
\arguments{
\item{x}{A ClusteredNeuroVol object.}
}
\value{
An integer representing the number of clusters in the input object.
}
\description{
This function returns the number of clusters in a ClusteredNeuroVol object.
}
