% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{num_clusters}
\alias{num_clusters}
\title{Number of Clusters}
\usage{
num_clusters(x)
}
\arguments{
\item{x}{the object to extract number of clusters}
}
\value{
An integer representing the number of clusters in the object.
}
\description{
Number of Clusters
}
