% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{OrientationList2D}
\alias{OrientationList2D}
\title{Pre-defined 2D orientation configurations}
\format{
An object of class \code{list} of length 24.
}
\usage{
OrientationList2D
}
\description{
A list of standard 2D anatomical orientations used in neuroimaging.
Each orientation defines a pair of anatomical axes.
}
\keyword{datasets}
