% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{OrientationList3D}
\alias{OrientationList3D}
\title{Pre-defined 3D orientation configurations}
\format{
An object of class \code{list} of length 48.
}
\usage{
OrientationList3D
}
\description{
A list of standard 3D anatomical orientations used in neuroimaging.
Each orientation defines a triplet of anatomical axes.
}
\keyword{datasets}
