% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{partition}
\alias{partition}
\alias{partition,LogicalNeuroVol,integer-method}
\alias{partition,LogicalNeuroVol,numeric-method}
\title{partition an image into a set of disjoint clusters}
\usage{
partition(x, k, ...)

\S4method{partition}{LogicalNeuroVol,integer}(x, k)

\S4method{partition}{LogicalNeuroVol,numeric}(x, k)
}
\arguments{
\item{x}{the object to partition}

\item{k}{the number of clusters}

\item{...}{additional arguments}
}
\description{
partition an image into a set of disjoint clusters
}
