% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustervol.R
\docType{methods}
\name{partition_by}
\alias{partition_by}
\alias{partition_by,NeuroVol,ClusteredNeuroVol-method}
\title{Generic function to partition an image into a set of regions}
\usage{
partition_by(x, by, ...)

\S4method{partition_by}{NeuroVol,ClusteredNeuroVol}(x, by)
}
\arguments{
\item{x}{the object to partition}

\item{by}{the object defining the partitions over the image space.}

\item{...}{additional arguments}
}
\description{
Generic function to partition an image into a set of regions
}
\examples{
mask <- NeuroVol(array(1,c(10,10,10)), NeuroSpace(c(10,10,10)))
mask.idx <- which(mask != 0)
grid <- index_to_coord(mask, mask.idx)
vox <- index_to_grid(mask, mask.idx)
kres <- kmeans(grid, centers=50, iter.max=500)
kvol <- ClusteredNeuroVol(mask, kres$cluster)
klis <- partition_by(mask, kvol)
mask \%>\% partition_by(kvol) \%>\% purrr::map(~ mean(values(.)))
}
