% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{patch_set,NeuroVol,numeric,missing-method}
\alias{patch_set,NeuroVol,numeric,missing-method}
\alias{patch_set,NeuroVol,numeric,LogicalNeuroVol-method}
\title{Create a patch set from a NeuroVol object}
\usage{
\S4method{patch_set}{NeuroVol,numeric,missing}(x, dims, mask, ...)

\S4method{patch_set}{NeuroVol,numeric,LogicalNeuroVol}(x, dims, mask, ...)
}
\arguments{
\item{x}{a NeuroVol object}

\item{dims}{the dimensions of the patch}

\item{mask}{the mask defining the valid patch centers}

\item{...}{additional args}
}
\value{
A deferred list of patches.

A deferred list of patches.
}
\description{
This function creates a patch set from a NeuroVol object given specified dimensions

This function creates a patch set from a NeuroVol object given specified dimensions and a mask.
}
