% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{perm_mat,AxisSet2D-method}
\alias{perm_mat,AxisSet2D-method}
\title{Get permutation matrix from axis set}
\usage{
\S4method{perm_mat}{AxisSet2D}(x, ...)
}
\arguments{
\item{x}{An AxisSet2D object}

\item{...}{Additional arguments (not used)}
}
\value{
A matrix representing the axis directions
}
\description{
Get permutation matrix from axis set
}
\examples{
\dontrun{
axes <- AxisSet2D(LEFT_RIGHT, ANT_POST)
perm_mat(axes)
}
}
