% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{methods}
\name{perm_mat,AxisSet2D-method}
\alias{perm_mat,AxisSet2D-method}
\alias{perm_mat,AxisSet3D-method}
\alias{perm_mat,NeuroSpace-method}
\title{perm_mat}
\usage{
\S4method{perm_mat}{AxisSet2D}(x, ...)

\S4method{perm_mat}{AxisSet3D}(x, ...)

\S4method{perm_mat}{NeuroSpace}(x, ...)
}
\description{
perm_mat

perm_mat

perm_mat
}
