% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{perm_mat,NeuroSpace-method}
\alias{perm_mat,NeuroSpace-method}
\title{Get permutation matrix from axis set}
\usage{
\S4method{perm_mat}{NeuroSpace}(x, ...)
}
\arguments{
\item{x}{A NeuroSpace object}

\item{...}{Additional arguments (not used)}
}
\value{
A matrix representing the axis directions
}
\description{
Get permutation matrix from axis set
}
\examples{
\dontrun{
ns <- NeuroSpace(axes=AxisSet3D(LEFT_RIGHT, ANT_POST, INF_SUP))
perm_mat(ns)
}
}
