% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{perm_mat}
\alias{perm_mat}
\title{Extract permutation matrix}
\usage{
perm_mat(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
Extract permutation matrix
}

