% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,NeuroVol-method}
\title{plot}
\usage{
plot(x, y, ...)

\S4method{plot}{NeuroVol}(x, cmap = gray(seq(0, 1, length.out = 255)),
  zlevels = unique(round(seq(1, dim(x)[3], length.out = 6))),
  irange = range(x), thresh = c(0, 0), bgvol = NULL)
}
\arguments{
\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{zlevels}{the series of slice indices to display.}

\item{irange}{the intensity range indicating the low and high values of the color scale.}
}
\description{
plot

plot
}
\examples{

dat <- matrix(rnorm(100*100), 100, 100)
slice <- NeuroSlice(dat, NeuroSpace(c(100,100)))
plot(slice)
}
