% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{print_,AxisSet2D-method}
\alias{print_,AxisSet2D-method}
\title{Print method for AxisSet2D objects}
\usage{
\S4method{print_}{AxisSet2D}(x, ...)
}
\arguments{
\item{x}{An AxisSet2D object}

\item{...}{Additional arguments (not used)}
}
\value{
Character string representing the axis set
}
\description{
Print method for AxisSet2D objects
}
\keyword{internal}
