% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{print_,AxisSet3D-method}
\alias{print_,AxisSet3D-method}
\title{Print method for AxisSet3D objects}
\usage{
\S4method{print_}{AxisSet3D}(x, ...)
}
\arguments{
\item{x}{An AxisSet3D object}

\item{...}{Additional arguments (not used)}
}
\value{
Character string representing the axis set
}
\description{
Print method for AxisSet3D objects
}
\keyword{internal}
