% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{print_,NamedAxis-method}
\alias{print_,NamedAxis-method}
\title{Print method for NamedAxis objects}
\usage{
\S4method{print_}{NamedAxis}(x, ...)
}
\arguments{
\item{x}{A NamedAxis object}

\item{...}{Additional arguments (not used)}
}
\value{
Character string representing the axis
}
\description{
Print method for NamedAxis objects
}
\keyword{internal}
