% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_io.R
\name{read_afni_header}
\alias{read_afni_header}
\title{read_afni_header}
\usage{
read_afni_header(file_name)
}
\arguments{
\item{file_name}{the name of the AFNI header file (ending in .HEAD)}
}
\value{
a \code{list} representation of an AFNI header
}
\description{
read_afni_header
}
\keyword{internal}
