% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{read_header}
\alias{read_header}
\title{read header information of an image file}
\usage{
read_header(file_name)
}
\arguments{
\item{file_name}{the name of the file to read}
}
\value{
an instance of class \code{\linkS4class{FileMetaInfo}}
}
\description{
read header information of an image file
}
