% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_data}
\alias{read_mapped_data}
\title{Read mapped data from a 4D image file}
\usage{
read_mapped_data(meta, idx)
}
\arguments{
\item{meta}{An object containing metadata about the file to be read.}

\item{idx}{A vector of indices specifying which elements to read.}
}
\value{
A vector of values read from the memory-mapped file.
}
\description{
This function reads data from a memory-mapped 4D image file.
}
\keyword{internal}
