% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_series}
\alias{read_mapped_series}
\title{Read a mapped series from a 4D image file}
\usage{
read_mapped_series(meta, idx)
}
\arguments{
\item{meta}{An object containing metadata about the file to be read.}

\item{idx}{A vector of indices specifying which elements to read.}
}
\value{
A matrix of values read from the memory-mapped file, transposed.
}
\description{
This function reads a series of data from a memory-mapped 4D image file.
}
\keyword{internal}
