% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_vols}
\alias{read_mapped_vols}
\title{Read mapped volumes from a 4D image file}
\usage{
read_mapped_vols(meta, idx)
}
\arguments{
\item{meta}{An object containing metadata about the file to be read.}

\item{idx}{A vector of indices specifying which volumes to read.}
}
\value{
A matrix of values read from the memory-mapped file.
}
\description{
This function reads volumes of data from a memory-mapped 4D image file.
}
\keyword{internal}
