% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{read_vol_list}
\alias{read_vol_list}
\title{read_vol_list}
\usage{
read_vol_list(file_names, mask = NULL)
}
\arguments{
\item{file_names}{A list of file names to load.}

\item{mask}{An optional mask defining the subset of voxels to load.}
}
\value{
An instance of the \code{\linkS4class{NeuroVec}} class.
}
\description{
This function loads a list of image volumes and returns a NeuroVec object.
}
