% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{render}
\alias{render}
\title{Render an image to create a drawable image.}
\usage{
render(x, width, height, colmap, ...)
}
\arguments{
\item{x}{the object, e.g. an instance of type \code{BrainSlice}}

\item{width}{width of the rendered image}

\item{height}{height of the rendered image}

\item{colmap}{the colors used to map from values to RGBA colors.}

\item{...}{additional arguments}
}
\description{
Render an image to create a drawable image.
}
