% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{render_slice}
\alias{render_slice}
\title{Render a slice at z coordinate}
\usage{
render_slice(x, zpos, width, height, colmap, ...)
}
\arguments{
\item{x}{the object, e.g. an instance of type \code{Layer} or \code{Overlay}}

\item{zpos}{the z coordinate to slice through.}

\item{width}{width of the rendered image}

\item{height}{height of the rendered image}

\item{colmap}{the colors used to map from values to RGBA colors.}

\item{...}{additional arguments}
}
\description{
Render a slice at z coordinate
}

