% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{reorient}
\alias{reorient}
\alias{reorient,NeuroSpace,character-method}
\title{Remap the grid-to-world coordinates mapping of an image.}
\usage{
reorient(x, orient)

\S4method{reorient}{NeuroSpace,character}(x, orient)
}
\arguments{
\item{x}{the object}

\item{orient}{the orientation code indicating the "remapped" axes.}
}
\value{
a reoriented image space
}
\description{
Remap the grid-to-world coordinates mapping of an image.
}
