% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/resample.R
\name{resample}
\alias{resample}
\alias{resample,NeuroVol,NeuroVol-method}
\alias{resample,NeuroVol,NeuroSpace-method}
\title{resample an image to match the space of another image}
\usage{
resample(source, target, ...)

\S4method{resample}{NeuroVol,NeuroVol}(source, target, interpolation = 3L)

\S4method{resample}{NeuroVol,NeuroSpace}(source, target, interpolation = 3L)
}
\arguments{
\item{source}{the source image}

\item{target}{the target image}

\item{...}{extra args}

\item{interpolation}{a single integer specifying the type of interpolation to be applied to the
final resampled image. May be 0 (nearest neighbor), 1 (trilinear) or 3 (cubic spline).
No other values are valid.}
}
\description{
resample an image to match the space of another image
}
