% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample,NeuroVol,NeuroVol-method}
\alias{resample,NeuroVol,NeuroVol-method}
\title{Resample a NeuroVol object}
\usage{
\S4method{resample}{NeuroVol,NeuroVol}(source, target, interpolation = 3L)
}
\arguments{
\item{source}{A NeuroVol object representing the source volume to be resampled.}

\item{target}{A NeuroVol object representing the target volume to match the dimensions and orientation of the source volume.}

\item{interpolation}{A single integer specifying the type of interpolation to be applied to the final resampled image. May be 0 (nearest neighbor), 1 (trilinear), or 3 (cubic spline). No other values are valid.}
}
\value{
A NeuroVol object with the resampled source volume matched to the target volume's dimensions and orientation.
}
\description{
This method resamples a NeuroVol object (\code{source}) to match the dimensions and orientation of a target NeuroVol object (\code{target}).
}
