% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{resample}
\alias{resample}
\title{resample an image to match the space of another image}
\usage{
resample(source, target, ...)
}
\arguments{
\item{...}{extra args}

\item{the}{target image}
}
\description{
resample an image to match the space of another image
}
