% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ROICoords-class}
\alias{ROICoords-class}
\title{ROICoords}
\description{
A class representing a region of interest (ROI) in a brain image, defined by a set of coordinates. This class stores the geometric space of the image and the coordinates of the voxels within the ROI.
}
\section{Slots}{

\describe{
\item{\code{space}}{An instance of class \code{\linkS4class{NeuroSpace}} representing the geometric space of the image data.}

\item{\code{coords}}{A \code{matrix} containing the coordinates of the voxels within the ROI.
Each row represents a coordinate as, e.g. (i,   j,  k).}
}}

