% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROIVec}
\alias{ROIVec}
\title{Create an instance of class \code{\linkS4class{ROIVec}}}
\usage{
ROIVec(vspace, coords, data = rep(nrow(coords), 1))
}
\arguments{
\item{vspace}{an instance of class \code{NeuroSpace}}

\item{coords}{matrix of voxel coordinates}

\item{data}{the \code{matrix} of data values}
}
\value{
an instance of class \code{ROIVec}
}
\description{
Create an instance of class \code{\linkS4class{ROIVec}}
}
