% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scale}
\alias{scale}
\title{Generic Scale Method}
\usage{
scale(x, ...)
}
\arguments{
\item{x}{The object to be scaled.}

\item{...}{Additional arguments for scaling methods.}
}
\value{
A scaled version of the input object.
}
\description{
Scales an object by (typically) subtracting the mean and dividing by the standard deviation.
}
