% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{series,NeuroHyperVec,ANY-method}
\alias{series,NeuroHyperVec,ANY-method}
\title{Series method for NeuroHyperVec}
\usage{
\S4method{series}{NeuroHyperVec,ANY}(x, i, j, k, ...)
}
\arguments{
\item{x}{The NeuroHyperVec object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{...}{Additional arguments (not used)}
}
\description{
Series method for NeuroHyperVec
}
\details{
when x is a NeuroHyperVec object, the series method returns a 2D array with dimensions [features x trials]
}
