% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{show,AxisSet1D-method}
\alias{show,AxisSet1D-method}
\title{Show method for AxisSet1D objects}
\usage{
\S4method{show}{AxisSet1D}(object)
}
\arguments{
\item{object}{An AxisSet1D object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for AxisSet1D objects
}
