% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{show,AxisSet2D-method}
\alias{show,AxisSet2D-method}
\title{Show method for AxisSet2D objects}
\usage{
\S4method{show}{AxisSet2D}(object)
}
\arguments{
\item{object}{An AxisSet2D object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for AxisSet2D objects
}
