% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{show,AxisSet4D-method}
\alias{show,AxisSet4D-method}
\title{Show method for AxisSet4D objects}
\usage{
\S4method{show}{AxisSet4D}(object)
}
\arguments{
\item{object}{An AxisSet4D object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for AxisSet4D objects
}
