% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{show,FileMetaInfo-method}
\alias{show,FileMetaInfo-method}
\title{show a \code{FileMetaInfo}}
\usage{
\S4method{show}{FileMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
show a \code{FileMetaInfo}
}
