% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_neurovec.R
\name{show,MappedNeuroVec-method}
\alias{show,MappedNeuroVec-method}
\title{Show Method for MappedNeuroVec Objects}
\usage{
\S4method{show}{MappedNeuroVec}(object)
}
\arguments{
\item{object}{A MappedNeuroVec object}
}
\description{
Displays a formatted summary of a MappedNeuroVec object.
}
\keyword{internal}
