% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/neuroslice.R, R/neurospace.R,
%   R/neurovec.R, R/neurovecseq.R, R/roi.R
\name{show,IndexLookupVol-method}
\alias{show,IndexLookupVol-method}
\alias{show,NeuroSlice-method}
\alias{show,NeuroSpace-method}
\alias{show,DenseNeuroVec-method}
\alias{show,NeuroVecSeq-method}
\alias{show}
\alias{show,ROICoords-method}
\title{Show method for IndexLookupVol Objects}
\usage{
\S4method{show}{IndexLookupVol}(object)

\S4method{show}{NeuroSlice}(object)

\S4method{show}{NeuroSpace}(object)

\S4method{show}{DenseNeuroVec}(object)

\S4method{show}{NeuroVecSeq}(object)

\S4method{show}{NeuroVecSeq}(object)

\S4method{show}{ROICoords}(object)
}
\arguments{
\item{object}{A \code{NeuroSlice} object}
}
\value{
Invisibly returns NULL after printing the object summary.
}
\description{
Displays a formatted summary of a \code{NeuroSlice} object's properties using
color-coded output for improved readability.
}
\details{
Display NeuroSlice Information
}
\examples{
slice <- NeuroSlice(matrix(1:100, 10, 10), NeuroSpace(c(10, 10)))
show(slice)

}
