% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{show,NamedAxis-method}
\alias{show,NamedAxis-method}
\title{Show method for NamedAxis objects}
\usage{
\S4method{show}{NamedAxis}(object)
}
\arguments{
\item{object}{A NamedAxis object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for NamedAxis objects
}
