% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{show,NeuroVec-method}
\alias{show,NeuroVec-method}
\title{show a NeuroVec object}
\usage{
\S4method{show}{NeuroVec}(object)
}
\arguments{
\item{object}{The NeuroVec object to display.}
}
\description{
This function prints a summary of a NeuroVec object.
}
