% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{show,NeuroVecSource-method}
\alias{show,NeuroVecSource-method}
\title{show a NeuroVecSource object}
\usage{
\S4method{show}{NeuroVecSource}(object)
}
\arguments{
\item{object}{The NeuroVecSource object to display.}
}
\description{
This function prints a summary of a NeuroVecSource object.
}
