% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\docType{methods}
\name{show,NeuroVol-method}
\alias{show,NeuroVol-method}
\title{show a \code{NeuroVol}}
\usage{
\S4method{show}{NeuroVol}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NeuroVol}
}
