% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_neurovec.R
\name{show,SparseNeuroVec-method}
\alias{show,SparseNeuroVec-method}
\title{show a \code{SparseNeuroVec}}
\usage{
\S4method{show}{SparseNeuroVec}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{SparseNeuroVec}
}
