% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{show,SparseNeuroVol-method}
\alias{show,SparseNeuroVol-method}
\title{show a \code{SparseNeuroVol}}
\usage{
\S4method{show}{SparseNeuroVol}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{SparseNeuroVol}
}
