% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{slice}
\alias{slice}
\title{Extract a 2D slice from an image volume}
\usage{
slice(x, zlevel, along, orientation, ...)
}
\arguments{
\item{x}{the object}

\item{zlevel}{coordinate (in voxel units) along the sliced axis}

\item{along}{the axis along which to slice}

\item{orientation}{the target orientation of the 2D slice}

\item{...}{additional arguments}
}
\description{
Extract a 2D slice from an image volume
}
