% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{slices}
\alias{slices}
\alias{slices,NeuroVol-method}
\title{extract an ordered series of 2d slices}
\usage{
slices(x, ...)

\S4method{slices}{NeuroVol}(x)
}
\arguments{
\item{x}{the object that supplies the slices}

\item{...}{additional arguments}
}
\description{
extract an ordered series of 2d slices
}
