% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{space}
\alias{space}
\title{Generic function to extract geometric properties of an image.}
\usage{
space(x, ...)
}
\arguments{
\item{x}{the object to query, e.g. an instance of \code{\linkS4class{NeuroVol}} or \code{\linkS4class{BrainVector}}}

\item{...}{additional arguments}
}
\value{
an object representing the geometric space of the image of type \code{\linkS4class{NeuroSpace}}
}
\description{
Generic function to extract geometric properties of an image.
}
\examples{
x = NeuroSpace(c(10,10,10), c(1,1,1))
vol <- NeuroVol(rnorm(10*10*10), x)
identical(x,space(vol))

}

