% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_neurovec.R
\name{[[,AbstractSparseNeuroVec,numeric-method}
\alias{[[,AbstractSparseNeuroVec,numeric-method}
\title{[[}
\usage{
\S4method{[[}{AbstractSparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{the volume index}
}
\value{
a SparseNeuroVol object
}
\description{
[[
}
