% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SparseNeuroVecSource-class}
\alias{SparseNeuroVecSource-class}
\title{SparseNeuroVecSource}
\description{
A class that is used to produce a \code{\linkS4class{SparseNeuroVec}} instance
}
\section{Slots}{

\describe{
\item{\code{mask}}{the subset of voxels that will be stored in memory}
}}

