% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{SparseNeuroVol,array}
\alias{SparseNeuroVol,array}
\title{Convert SparseNeuroVol to array}
\arguments{
\item{from}{A SparseNeuroVol object.}
}
\value{
An array object resulting from the conversion.
}
\description{
This function converts a SparseNeuroVol object to an array object.
}
\keyword{internal}
