% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neurovol.R
\docType{class}
\name{SparseNeuroVol-class}
\alias{SparseNeuroVol-class}
\alias{SparseNeuroVol}
\title{SparseNeuroVol}
\usage{
SparseNeuroVol(data, space, indices = NULL, label = "")
}
\arguments{
\item{data}{a numeric vector}

\item{space}{an instance of class \code{\linkS4class{NeuroSpace}}}

\item{indices}{a index vector indicating the 1-d coordinates of the data values}

\item{label}{a \code{character} string}
}
\value{
\code{\linkS4class{SparseNeuroVol}} instance
}
\description{
Three-dimensional brain image, backed by a \code{sparseVector} for \code{Matrix} package

Construct a \code{\linkS4class{SparseNeuroVol}} instance
}
\details{
Image data is backed by \code{Matrix::sparseVector}.
}
\section{Slots}{

\describe{
\item{\code{data}}{a \code{sparseVector} instance}
}}

\examples{
data <- 1:10
indices <- seq(1,1000, length.out=10)
bspace <- NeuroSpace(c(64,64,64), spacing=c(1,1,1))
sparsevol <- SparseNeuroVol(data,bspace,indices=indices)
densevol <- NeuroVol(data,bspace,indices=indices)
sum(sparsevol) == sum(densevol)



}
