% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{SparseNeuroVol,numeric}
\alias{SparseNeuroVol,numeric}
\title{Convert SparseNeuroVol to numeric}
\arguments{
\item{from}{A SparseNeuroVol object.}
}
\value{
A numeric object resulting from the conversion.
}
\description{
This function converts a SparseNeuroVol object to a numeric object.
}
\keyword{internal}
