% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{split_clusters}
\alias{split_clusters}
\title{Generic function to cut an object into a list of spatial or spatiotemporal clusters}
\usage{
split_clusters(x, clusters, ...)
}
\arguments{
\item{x}{the object to split}

\item{clusters}{a vector of cluster indices to split by}

\item{...}{additional arguments}
}
\description{
Generic function to cut an object into a list of spatial or spatiotemporal clusters
}
