% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_like.R
\docType{methods}
\name{[,ArrayLike4D,numeric,missing,ANY-method}
\alias{[,ArrayLike4D,numeric,missing,ANY-method}
\alias{[,ArrayLike4D,numeric,missing,ANY-method}
\title{extractor}
\usage{
\S4method{[}{ArrayLike4D,numeric,missing,ANY}(x, i, j, k, m, ..., drop = TRUE)

\S4method{[}{ArrayLike4D,numeric,missing,ANY}(x, i, j, k, m, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{m}{third index}

\item{...}{additional args}

\item{drop}{drop dimension}

\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{m}{the fourth index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor

extractor
}
