% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{[,NeuroHyperVec,ANY,ANY,ANY-method}
\alias{[,NeuroHyperVec,ANY,ANY,ANY-method}
\title{[ method for NeuroHyperVec}
\usage{
\S4method{[}{NeuroHyperVec,ANY,ANY,ANY}(x, i, j, k, l, m, ..., drop = TRUE)
}
\arguments{
\item{x}{The NeuroHyperVec object}

\item{i, j, k, l, m}{Indices for each dimension}

\item{...}{Additional arguments (not used)}

\item{drop}{Whether to drop dimensions of length 1}
}
\value{
A subset of the NeuroHyperVec data
}
\description{
[ method for NeuroHyperVec
}
