% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{[,NeuroVecSeq,numeric,missing,ANY-method}
\alias{[,NeuroVecSeq,numeric,missing,ANY-method}
\title{extractor}
\usage{
\S4method{[}{NeuroVecSeq,numeric,missing,ANY}(x, i, j, k, m, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{m}{the fourth index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}
