% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\docType{methods}
\name{[,SparseNeuroVol,numeric,missing,ANY-method}
\alias{[,SparseNeuroVol,numeric,missing,ANY-method}
\title{extractor}
\usage{
\S4method{[}{SparseNeuroVol,numeric,missing,ANY}(x, i, j, k, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}
