% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{[[,NeuroHyperVec,ANY-method}
\alias{[[,NeuroHyperVec,ANY-method}
\title{Get a single NeuroVec from a NeuroHyperVec}
\usage{
\S4method{[[}{NeuroHyperVec,ANY}(x, i)
}
\arguments{
\item{x}{A NeuroHyperVec object}

\item{i}{Index of the NeuroVec to retrieve}
}
\value{
A NeuroVec object
}
\description{
Get a single NeuroVec from a NeuroHyperVec
}
