% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{[[,NeuroVec,numeric-method}
\alias{[[,NeuroVec,numeric-method}
\title{[[}
\usage{
\S4method{[[}{NeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{The NeuroVec object.}

\item{i}{The volume index to extract.}
}
\value{
A DenseNeuroVol object representing the extracted volume.
}
\description{
This function extracts a single volume from a NeuroVec object.
}
