% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_neurovec.R, R/clustervol.R
\name{sub_clusters,ClusteredNeuroVec,integer-method}
\alias{sub_clusters,ClusteredNeuroVec,integer-method}
\alias{sub_clusters,ClusteredNeuroVec,numeric-method}
\alias{sub_clusters,ClusteredNeuroVec,character-method}
\alias{sub_clusters-methods}
\alias{sub_clusters,ClusteredNeuroVol,integer-method}
\alias{sub_clusters,ClusteredNeuroVol,numeric-method}
\alias{sub_clusters,ClusteredNeuroVol,character-method}
\title{Select a Subset of Clusters}
\usage{
\S4method{sub_clusters}{ClusteredNeuroVec,integer}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVec,numeric}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVec,character}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,integer}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,numeric}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,character}(x, ids, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{ClusteredNeuroVol}} or
\code{\linkS4class{ClusteredNeuroVec}} object.}

\item{ids}{Integer cluster IDs, numeric (coerced to integer), or character
cluster names (looked up in the label map) identifying the clusters to
retain.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of the same class as \code{x} containing only the
  selected clusters.
}
\description{
Select a Subset of Clusters
}
