% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustered_neurovec.R,
%   R/clustervol.R
\name{sub_clusters}
\alias{sub_clusters}
\alias{sub_clusters,ClusteredNeuroVec,integer-method}
\alias{sub_clusters,ClusteredNeuroVec,numeric-method}
\alias{sub_clusters,ClusteredNeuroVec,character-method}
\alias{sub_clusters,ClusteredNeuroVol,integer-method}
\alias{sub_clusters,ClusteredNeuroVol,numeric-method}
\alias{sub_clusters,ClusteredNeuroVol,character-method}
\title{Select a Subset of Clusters}
\usage{
sub_clusters(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVec,integer}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVec,numeric}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVec,character}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,integer}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,numeric}(x, ids, ...)

\S4method{sub_clusters}{ClusteredNeuroVol,character}(x, ids, ...)
}
\arguments{
\item{x}{A clustered neuroimaging object.}

\item{ids}{Integer cluster IDs, numeric (coerced to integer), or character
cluster names to retain.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of the same class as \code{x} containing only the
  selected clusters.
}
\description{
Return a new object containing only the requested clusters. Clusters can
be identified by integer ID or by name (matched against the label map).
}
\examples{
sp <- NeuroSpace(c(10L, 10L, 10L), c(1, 1, 1))
mask <- LogicalNeuroVol(array(c(rep(TRUE, 500), rep(FALSE, 500)),
                              c(10, 10, 10)), sp)
clusters <- rep(1:5, length.out = 500)
cvol <- ClusteredNeuroVol(mask, clusters,
          label_map = list(A = 1, B = 2, C = 3, D = 4, E = 5))
# By integer ID
sub <- sub_clusters(cvol, c(1L, 3L))
# By name
sub2 <- sub_clusters(cvol, c("A", "C"))

}
