% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{sub_vector}
\alias{sub_vector}
\title{Generic function to extract a sub-vector from a \code{BrainVector} object.}
\usage{
sub_vector(x, i, ...)
}
\arguments{
\item{x}{four-dimensional image}

\item{i}{the indices of the volume(s) to extract}

\item{...}{additional arguments}
}
\value{
a  \code{BrainVector} object that is a sub-sequence of the supplied object.
}
\description{
Generic function to extract a sub-vector from a \code{BrainVector} object.
}
\examples{
bvec <- BrainVector(array(rnorm(24*24*24*24), c(24,24,24,24)), NeuroSpace(c(24,24,24,24), c(1,1,1)))
vec <- sub_vector(bvec,1:2)
all.equal(2, dim(vec)[4])

vec <- sub_vector(bvec, c(1,3,5,7))
all.equal(4, dim(vec)[4])

mask <- LogicalNeuroVol(rep(TRUE, 24*24*24), NeuroSpace(c(24,24,24), c(1,1,1)))
svec <- SparseBrainVector(array(rnorm(24*24*24*24), c(24,24,24,24)),
NeuroSpace(c(24,24,24,24), c(1,1,1)), mask)
vec <- sub_vector(svec, c(1,3,5))
all.equal(3, dim(vec)[4])
}

