% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Summary,DenseNeuroVol-method}
\alias{Summary,DenseNeuroVol-method}
\title{Summary of DenseNeuroVol}
\usage{
\S4method{Summary}{DenseNeuroVol}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A DenseNeuroVol object.}

\item{...}{Additional arguments (currently ignored).}

\item{na.rm}{A logical value indicating whether NA values should be removed before computation.}
}
\value{
A summary of the input DenseNeuroVol object.
}
\description{
This function computes a summary of a DenseNeuroVol object.
}
