% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Summary,SparseNeuroVec-method}
\alias{Summary,SparseNeuroVec-method}
\alias{Summary,SparseNeuroVol-method}
\alias{Summary,DenseNeuroVol-method}
\title{Summary method for Neuroimaging objects}
\usage{
\S4method{Summary}{SparseNeuroVec}(x, ..., na.rm = FALSE)

\S4method{Summary}{SparseNeuroVol}(x, ..., na.rm = FALSE)

\S4method{Summary}{DenseNeuroVol}(x, ..., na.rm = FALSE)

\S4method{Summary}{DenseNeuroVol}(x, ..., na.rm = FALSE)

\S4method{Summary}{DenseNeuroVol}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A SparseNeuroVec object}

\item{...}{Additional arguments passed to methods}

\item{na.rm}{Logical indicating whether to remove NA values before computation}
}
\description{
Summary method for Neuroimaging objects
}
