% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary-neuro-methods}
\alias{summary-neuro-methods}
\alias{summary,NeuroVol-method}
\alias{summary,DenseNeuroVec-method}
\alias{summary,SparseNeuroVec-method}
\title{Summary of Neuroimaging Objects}
\usage{
\S4method{summary}{NeuroVol}(object, ...)

\S4method{summary}{DenseNeuroVec}(object, ...)

\S4method{summary}{SparseNeuroVec}(object, ...)
}
\arguments{
\item{object}{A neuroimaging object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class \code{"summary.NeuroVol"} or
  \code{"summary.NeuroVec"} (invisibly), with a print method.
}
\description{
Provides a concise summary of neuroimaging volume and vector objects,
including spatial metadata and data statistics.
}
\examples{
vol <- DenseNeuroVol(array(rnorm(27), c(3,3,3)),
                     NeuroSpace(c(3L,3L,3L), c(1,1,1)))
summary(vol)

}
