% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Summary,SparseNeuroVec-method}
\alias{Summary,SparseNeuroVec-method}
\title{Summary of SparseNeuroVec}
\usage{
\S4method{Summary}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{A SparseNeuroVec object.}
}
\value{
A summary of the input SparseNeuroVec object.
}
\description{
This function computes a summary of a SparseNeuroVec object.
}
