% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Summary,SparseNeuroVol-method}
\alias{Summary,SparseNeuroVol-method}
\title{Summary of SparseNeuroVol}
\usage{
\S4method{Summary}{SparseNeuroVol}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A SparseNeuroVol object.}

\item{...}{Additional arguments (currently ignored).}

\item{na.rm}{A logical value indicating whether NA values should be removed before computation.}
}
\value{
A summary of the input SparseNeuroVol object.
}
\description{
This function computes a summary of a SparseNeuroVol object.
}
