% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_neuro}
\alias{theme_neuro}
\title{A minimal, publication-friendly theme for image slices}
\usage{
theme_neuro(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\description{
Quiet axes, thin panel border, no grid, generous margins, slim legend.
}
