% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{TIME}
\alias{TIME}
\title{Time axis}
\format{
An object of class \code{NamedAxis} of length 1.
}
\usage{
TIME
}
\description{
Represents the temporal dimension in neuroimaging data
}
\keyword{datasets}
