% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{TimeAxis}
\alias{TimeAxis}
\title{Time axis set}
\format{
An object of class \code{AxisSet1D} of length 1.
}
\usage{
TimeAxis
}
\description{
A one-dimensional axis set representing time
}
\keyword{datasets}
