% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/roi.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,ROIVol-method}
\alias{values,ROIVec-method}
\title{Generic function to extract data values of object}
\usage{
values(x, ...)

\S4method{values}{ROIVol}(x, ...)

\S4method{values}{ROIVec}(x, ...)
}
\arguments{
\item{x}{the object to get values from}

\item{...}{additional arguments}
}
\description{
Generic function to extract data values of object
}
