% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R, R/roi.R,
%   R/sparse_neurovec.R
\name{vectors}
\alias{vectors}
\alias{vectors,NeuroVec,missing-method}
\alias{vectors,NeuroVec,numeric-method}
\alias{vectors,NeuroVec,logical-method}
\alias{vectors,ROIVec,missing-method}
\alias{vectors,ROIVec,integer-method}
\alias{vectors,ROIVec,numeric-method}
\alias{vectors,ROIVec,logical-method}
\alias{vectors,SparseNeuroVec,missing-method}
\title{extract an ordered list of 1d vectors}
\usage{
vectors(x, subset, ...)

\S4method{vectors}{NeuroVec,missing}(x)

\S4method{vectors}{NeuroVec,numeric}(x, subset)

\S4method{vectors}{NeuroVec,logical}(x, subset)

\S4method{vectors}{ROIVec,missing}(x)

\S4method{vectors}{ROIVec,integer}(x, subset)

\S4method{vectors}{ROIVec,numeric}(x, subset)

\S4method{vectors}{ROIVec,logical}(x, subset)

\S4method{vectors}{SparseNeuroVec,missing}(x, nonzero = FALSE)
}
\arguments{
\item{x}{the object that supplies the vector data}

\item{subset}{the subset of vectors to extract}

\item{...}{additional arguments}

\item{nonzero}{only include nonzero vectors in output list}
}
\description{
extract an ordered list of 1d vectors
}
