% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\docType{methods}
\name{[,ROIVol,numeric,missing,ANY-method}
\alias{[,ROIVol,numeric,missing,ANY-method}
\alias{[,ROIVol,logical,missing,ANY-method}
\title{subset an \code{ROIVol}}
\usage{
\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, drop)

\S4method{[}{ROIVol,logical,missing,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{drop}{drop dimension}
}
\description{
subset an \code{ROIVol}
}
