% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{voxels}
\alias{voxels}
\title{extract voxel coordinates}
\usage{
voxels(x, ...)
}
\arguments{
\item{x}{the object to extract voxels from}

\item{...}{additional arguments to function}
}
\description{
extract voxel coordinates
}

