% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{which_dim}
\alias{which_dim}
\alias{which_dim,NeuroSpace,NamedAxis-method}
\title{find dimensions of a given axis}
\usage{
which_dim(x, axis)

\S4method{which_dim}{NeuroSpace,NamedAxis}(x, axis)
}
\arguments{
\item{x}{the object}

\item{axis}{the axis to return dimension of}
}
\description{
find dimensions of a given axis
}
