% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{write_vol}
\alias{write_vol}
\title{Generic function to write a 3D image volume to disk}
\usage{
write_vol(x, file_name, format, dataType)
}
\arguments{
\item{x}{an image object, typically a \code{NeuroVol} instance.}

\item{file_name}{output file name}

\item{format}{file format string. Since "NIFTI" is the only currently supported format, this parameter can be safely ignored and omitted.}

\item{dataType}{output data type, If specified should be a \code{character} vector of: "BINARY", "UBYTE", "SHORT", "INT", "FLOAT", "DOUBLE".
Otherwise output format will be inferred from R the datatype of the image.}
}
\description{
Generic function to write a 3D image volume to disk
}
\details{
The output format will be inferred from file extension.

The output format will be inferred from file extension.
 \code{writeVolume(x, "out.nii")} outputs a NIFTI file.
 \code{writeVolume(x, "out.nii.gz")} outputs a gzipped NIFTI file.

No other file output formats are currently supported.
}
\examples{

bvol <- NeuroVol(array(0, c(10,10,10)), NeuroSpace(c(10,10,10), c(1,1,1)))
\dontrun{
writeVolume(bvol, "out.nii")
writeVolume(bvol, "out.nii.gz")
}
}
