% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurospace.R
\docType{methods}
\name{add_dim,NeuroSpace,numeric-method}
\alias{add_dim,NeuroSpace,numeric-method}
\title{add dimension to \code{\linkS4class{NeuroSpace}}}
\usage{
\S4method{add_dim}{NeuroSpace,numeric}(x, n)
}
\description{
add dimension to \code{\linkS4class{NeuroSpace}}
}
