% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{add_dim}
\alias{add_dim}
\alias{add_dim,NeuroSpace,numeric-method}
\title{Generic function to add a dimension to an object}
\usage{
add_dim(x, n)

\S4method{add_dim}{NeuroSpace,numeric}(x, n)
}
\arguments{
\item{x}{a dimensioned object}

\item{n}{the size of the dimension to add}
}
\description{
Generic function to add a dimension to an object

add dimension to \code{\linkS4class{NeuroSpace}}
}
\examples{
x = NeuroSpace(c(10,10,10), c(1,1,1))
x1 <- add_dim(x, 10)
ndim(x1) == 4
dim(x1)[4] == 10
}
