% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R, R/neurovol.R, R/roi.R
\name{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\title{conversion from ClusteredNeuroVol to LogicalNeuroVol}
\description{
conversion from ClusteredNeuroVol to LogicalNeuroVol

conversion from DenseNeuroVol to array

conversion from \code{SparseNeuroVol} to \code{array}

conversion from SparseNeuroVol to numeric

conversion from NeuroVol to LogicalNeuroVol

conversion from NeuroVol to array
}
