% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/roi.R
\docType{methods}
\name{as.dense}
\alias{as.dense}
\alias{as.dense,ROIVol-method}
\title{Convert to dense representation}
\usage{
as.dense(x)

\S4method{as.dense}{ROIVol}(x)
}
\arguments{
\item{x}{the object to densify}
}
\description{
Convert to dense representation
}
