% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R, R/roi.R, R/sparse_neurovec.R
\name{as.dense,ClusteredNeuroVol-method}
\alias{as.dense,ClusteredNeuroVol-method}
\alias{as.dense,ROIVol-method}
\alias{as.dense,SparseNeuroVec-method}
\title{Coerce ROIVol to DenseNeuroVol using as.dense method}
\usage{
\S4method{as.dense}{ClusteredNeuroVol}(x)

\S4method{as.dense}{ROIVol}(x)

\S4method{as.dense}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{An object of class \code{ROIVol} to be coerced to a \code{DenseNeuroVol}.}
}
\value{
A \code{\linkS4class{NeuroVol}} object representing the dense version of the clustered volume.

A \code{DenseNeuroVol} object obtained by coercing the \code{ROIVol} object.
}
\description{
This function provides a method to coerce an object of class \code{ROIVol} to a \code{DenseNeuroVol} using the \code{as.dense} method.
}
