% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R, R/sparse_neurovec.R
\docType{methods}
\name{as.list,NeuroVec-method}
\alias{as.list,NeuroVec-method}
\alias{as.list,SparseNeuroVec-method}
\title{convert a \code{NeuroVec} to \code{list} of volumes.}
\usage{
\S4method{as.list}{NeuroVec}(x)

\S4method{as.list}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{NeuroVec} to \code{list} of volumes.

convert SparseNeuroVec to list of \code{\linkS4class{DenseBrainVolume}}
}
