% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R, R/neurovec.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{as.list,FileBackedNeuroVec-method}
\alias{as.list,FileBackedNeuroVec-method}
\alias{as.list,NeuroVec-method}
\alias{as.list,SparseNeuroVec-method}
\title{as.list}
\usage{
\S4method{as.list}{FileBackedNeuroVec}(x)

\S4method{as.list}{NeuroVec}(x)

\S4method{as.list}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert FileBackedNeuroVec to list of \code{\linkS4class{DenseNeuroVol}}

convert SparseNeuroVec to list of \code{\linkS4class{DenseNeuroVol}}
}
