% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R, R/neurovec.R,
%   R/sparse_neurovec.R
\name{as.list,FileBackedNeuroVec-method}
\alias{as.list,FileBackedNeuroVec-method}
\alias{as.list,NeuroVec-method}
\alias{as.list,SparseNeuroVec-method}
\title{Convert FileBackedNeuroVec to List}
\usage{
\S4method{as.list}{FileBackedNeuroVec}(x)

\S4method{as.list}{NeuroVec}(x)

\S4method{as.list}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\value{
A list of DenseNeuroVol objects
}
\description{
Converts a FileBackedNeuroVec object to a list of DenseNeuroVol objects.

convert SparseNeuroVec to list of \code{\linkS4class{DenseNeuroVol}}
}
\details{
This method creates a deferred list, where each element is a DenseNeuroVol
object representing a single volume from the FileBackedNeuroVec.
}
