% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\docType{methods}
\name{as.logical,NeuroVol-method}
\alias{as.logical,NeuroVol-method}
\title{as.logical}
\usage{
\S4method{as.logical}{NeuroVol}(x)
}
\arguments{
\item{x}{the object}
}
\value{
an instance of \code{linkS4class{LogicalNeuroVol}}
}
\description{
Convert NeuroVol to \code{linkS4class{LogicalNeuroVol}}
}
\details{
the image values will be converted to using R base function \code{as.logical} and wrapped in \code{LogicalNeuroVol}
}
