% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{as.mask,NeuroVol,missing-method}
\alias{as.mask,NeuroVol,missing-method}
\alias{as.mask,NeuroVol,numeric-method}
\title{Convert NeuroVol to a mask}
\usage{
\S4method{as.mask}{NeuroVol,missing}(x)

\S4method{as.mask}{NeuroVol,numeric}(x, indices)
}
\arguments{
\item{x}{A NeuroVol object to convert to a mask.}

\item{indices}{A numeric vector containing the indices of the input NeuroVol that should be set to TRUE in the resulting mask.}
}
\value{
A LogicalNeuroVol object representing the mask created from the input NeuroVol.

A LogicalNeuroVol object representing the mask created from the input NeuroVol with specified indices.
}
\description{
This method converts a NeuroVol object to a mask by setting all positive values to TRUE and all non-positive values to FALSE.

This method converts a NeuroVol object to a mask by setting the specified indices to TRUE and the remaining elements to FALSE.
}
