% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R, R/neurovec.R, R/roi.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{as.matrix,FileBackedNeuroVec-method}
\alias{as.matrix,FileBackedNeuroVec-method}
\alias{as.matrix,NeuroVec-method}
\alias{as.matrix,ROIVec-method}
\alias{as.matrix,SparseNeuroVec-method}
\title{convert a \code{NeuroVec} to a matrix}
\usage{
\S4method{as.matrix}{FileBackedNeuroVec}(x)

\S4method{as.matrix}{NeuroVec}(x)

\S4method{as.matrix}{ROIVec}(x)

\S4method{as.matrix}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{NeuroVec} to a matrix

convert a \code{ROIVec} to a matrix

convert SparseNeuroVec to matrix
}
