% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R, R/roi.R, R/sparse_neurovec.R
\docType{methods}
\name{as.matrix,DenseNeuroVec-method}
\alias{as.matrix,DenseNeuroVec-method}
\alias{as.matrix,ROIVec-method}
\alias{as.matrix,SparseNeuroVec-method}
\title{convert a \code{DenseNeuroVec} to a matrix}
\usage{
\S4method{as.matrix}{DenseNeuroVec}(x)

\S4method{as.matrix}{ROIVec}(x)

\S4method{as.matrix}{SparseNeuroVec}(x)
}
\arguments{
\item{x}{the object}

\item{x}{the object}
}
\description{
convert a \code{DenseNeuroVec} to a matrix

convert a \code{ROIVec} to a matrix

convert SparseNeuroVec to matrix
}
