% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{as.matrix,DenseNeuroVec-method}
\alias{as.matrix,DenseNeuroVec-method}
\title{convert a \code{DenseNeuroVec} to a matrix}
\usage{
\S4method{as.matrix}{DenseNeuroVec}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{DenseNeuroVec} to a matrix
}
