% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raster.R
\name{as.raster,NeuroSlice-method}
\alias{as.raster,NeuroSlice-method}
\alias{as.raster,NeuroVol-method}
\title{Convert neuroimaging objects to raster images}
\usage{
\S4method{as.raster}{NeuroSlice}(x, cmap = gray(seq(0, 1, length.out = 255)), irange = range(x, na.rm = TRUE))
\S4method{as.raster}{NeuroVol}(x, zlevel = ceiling(dim(x)[3]/2), cmap = gray(seq(0, 1, length.out = 255)), irange = range(x, na.rm = TRUE))
}
\arguments{
\item{x}{object to convert}
\item{cmap}{color map used to map intensities to colors}
\item{irange}{intensity range used for scaling}
\item{zlevel}{slice index for \code{NeuroVol} objects}
}
\value{
A \code{raster} object suitable for plotting.
}
\description{
Convert \code{NeuroSlice} or \code{NeuroVol} objects to raster images.
}
\seealso{\code{\link{plot,NeuroSlice-method}},\code{\link{plot,NeuroVol-method}}}
