% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.raster}
\alias{as.raster}
\title{Generic Method for Converting Objects to Raster Format}
\arguments{
\item{x}{An object to be converted.}

\item{...}{Additional arguments passed to the conversion methods.}
}
\value{
A \code{raster} object representing \code{x}.
}
\description{
Converts an object to a raster (bitmap) representation.
}
