% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{as}
\alias{as}
\title{conversion from \code{\linkS4class{NeuroVol}} to \code{\linkS4class{LogicalNeuroVol}}}
\description{
conversion from \code{\linkS4class{NeuroVol}} to \code{\linkS4class{LogicalNeuroVol}}

conversion from \code{\linkS4class{NeuroVol}} to \code{array}
}
