% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R, R/roi.R, R/sparse_neurovec.R
\name{as}
\alias{as}
\title{conversion from \code{\linkS4class{NeuroVol}} to \code{\linkS4class{LogicalNeuroVol}}}
\arguments{
\item{from}{An object of class \code{ROIVol} to be coerced to a \code{DenseNeuroVol}.}
}
\value{
A \code{matrix} obtained by coercing the \code{ROIVec} object.

A \code{DenseNeuroVol} object obtained by coercing the \code{ROIVol} object.
}
\description{
This function provides a method to coerce an object of class \code{ROIVec} to a \code{matrix}.

This function provides a method to coerce an object of class \code{ROIVol} to a \code{DenseNeuroVol}.
}
