% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R, R/neurovol.R
\name{as.sparse,DenseNeuroVec,LogicalNeuroVol-method}
\alias{as.sparse,DenseNeuroVec,LogicalNeuroVol-method}
\alias{as.sparse,DenseNeuroVec,numeric-method}
\alias{as.sparse,DenseNeuroVol,LogicalNeuroVol-method}
\alias{as.sparse,DenseNeuroVol,numeric-method}
\title{Convert DenseNeuroVec to sparse representation using mask}
\usage{
\S4method{as.sparse}{DenseNeuroVec,LogicalNeuroVol}(x, mask)

\S4method{as.sparse}{DenseNeuroVec,numeric}(x, mask)

\S4method{as.sparse}{DenseNeuroVec,LogicalNeuroVol}(x, mask)

\S4method{as.sparse}{DenseNeuroVec,numeric}(x, mask)

\S4method{as.sparse}{DenseNeuroVol,LogicalNeuroVol}(x, mask)

\S4method{as.sparse}{DenseNeuroVol,numeric}(x, mask)
}
\arguments{
\item{x}{A DenseNeuroVec object to convert to a sparse representation.}

\item{mask}{A numeric vector representing the mask to apply during conversion.}
}
\value{
A SparseNeuroVec object resulting from the conversion.

A SparseNeuroVec object resulting from the conversion.

A SparseNeuroVec object resulting from the conversion.

A SparseNeuroVec object resulting from the conversion.
}
\description{
This method converts a DenseNeuroVec object to a sparse representation using a given LogicalNeuroVol mask.

This method converts a DenseNeuroVec object to a sparse representation using a given numeric mask.

This method converts a DenseNeuroVec object to a sparse representation using a given LogicalNeuroVol mask.

This method converts a DenseNeuroVec object to a sparse representation using a given numeric mask.
}
