% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{as_canonical}
\alias{as_canonical}
\title{Reorient Image to Canonical (RAS+) Orientation}
\usage{
as_canonical(x, target = c("R", "A", "S"))
}
\arguments{
\item{x}{A \code{\linkS4class{NeuroVol}} or \code{\linkS4class{NeuroVec}}
object.}

\item{target}{Character vector of length 3 giving the desired orientation.
Default is \code{c("R", "A", "S")} (RAS+).}
}
\value{
A reoriented object of the same class as \code{x}.
}
\description{
Reorients a neuroimaging volume or vector to the canonical RAS+
(Right-Anterior-Superior) orientation by permuting and flipping axes.
This is equivalent to nibabel's \code{as_closest_canonical()}.
}
\details{
The function works by computing the orientation transform from the
current axis codes to the target codes, then applying the necessary
axis permutations and flips. The affine matrix is updated to reflect
the new orientation while preserving world-coordinate mapping.
}
\examples{
sp <- NeuroSpace(c(10L, 10L, 10L), c(2, 2, 2))
vol <- DenseNeuroVol(array(rnorm(1000), c(10,10,10)), sp)
ras_vol <- as_canonical(vol)
axcodes(ras_vol)  # "R" "A" "S"

}
\seealso{
\code{\link{axcodes}}, \code{\link{reorient}}
}
