% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{axes}
\alias{axes}
\title{Generic getter function to extract image axes}
\usage{
axes(x)
}
\arguments{
\item{x}{an object with a set of axes}
}
\description{
Generic getter function to extract image axes
}

