% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{axes}
\alias{axes}
\alias{axes,NeuroSpace-method}
\title{Extract Image Axes}
\usage{
axes(x)

\S4method{axes}{NeuroSpace}(x)
}
\arguments{
\item{x}{an object with a set of axes}
}
\description{
Extract Image Axes
}
