% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BigNeuroVec-class}
\alias{BigNeuroVec-class}
\title{BigNeuroVec}
\description{
a sparse four-dimensional brain image that is backed by a disk-based big-matrix
}
\section{Slots}{

\describe{
\item{\code{data}}{an instance of class \code{FBM} from \code{bigstatsr} package which contains time-series data.}
}}

