% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BigNeuroVec-class}
\alias{BigNeuroVec-class}
\title{BigNeuroVec}
\description{
a four-dimensional brain image that is backed by a disk-based big-matrix
}
\section{Slots}{

\describe{
\item{\code{meta}}{the file meta information of type \code{\linkS4class{FileMetaInfo}}}
}}

