% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BinaryReader-class}
\alias{BinaryReader-class}
\title{BinaryReader}
\description{
This class supports reading of bulk binary data from a connection
}
\section{Slots}{

\describe{
\item{\code{input}}{the binary input connection}

\item{\code{byte_offset}}{the number of bytes to skip at the start of input}

\item{\code{data_type}}{the dataType of the binary Elements}

\item{\code{bytes_per_element}}{number of bytes in each data element (e.g. 4 or 8 for floating point numbers)}

\item{\code{endian}}{endianness of binary input connection}

\item{\code{endian}}{logical indicating whether data tyoe is signed}
}}

