% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BinaryWriter-class}
\alias{BinaryWriter-class}
\title{BinaryWriter}
\description{
This class supports writing of bulk binary data to a connection
}
\section{Slots}{

\describe{
\item{\code{output}}{the binary output connection}

\item{\code{byte_offset}}{the number of bytes to skip at the start of input}

\item{\code{data_type}}{the dataType of the binary Elements}

\item{\code{bytes_per_element}}{number of bytes in each data element (e.g. 4 or 8 for floating point numbers)}

\item{\code{endian}}{endianness of binary output connection}
}}

