% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R, R/roi.R
\name{centroid}
\alias{centroid}
\alias{centroid,NeuroSpace-method}
\alias{centroid,ROICoords-method}
\title{return the centroid of an object}
\usage{
centroid(x, ...)

\S4method{centroid}{NeuroSpace}(x)

\S4method{centroid}{ROICoords}(x)
}
\arguments{
\item{x}{an object with a centroid}

\item{...}{extra args}
}
\value{
the centroid of the object
}
\description{
return the centroid of an object
}
\examples{

bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
centroid(bspace)

}
