% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustered_neurovec.R,
%   R/clustervol.R
\name{centroids}
\alias{centroids}
\alias{centroids,ClusteredNeuroVec-method}
\alias{centroids,ClusteredNeuroVol-method}
\title{Return a matrix of centroids of an object}
\usage{
centroids(x, ...)

\S4method{centroids}{ClusteredNeuroVec}(x, type = c("center_of_mass", "medoid"))

\S4method{centroids}{ClusteredNeuroVol}(x, type = c("center_of_mass", "medoid"))
}
\arguments{
\item{x}{an object with multiple centroids (e.g. a \code{ClusteredNeuroVol})}

\item{...}{extra args}

\item{type}{the type of center of mass: one of "center_of_mass" or "medoid"}
}
\value{
A numeric \code{matrix} where each row represents the coordinates of a centroid.

A matrix of coordinates where each row represents the centroid of a cluster.
}
\description{
Return a matrix of centroids of an object
}
\details{
For `type = "center_of_mass"`, returns arithmetic mean coordinates; for `"medoid"`, returns the most central point.
}
