% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustervol.R
\docType{methods}
\name{centroids}
\alias{centroids}
\alias{centroids,ClusteredNeuroVol-method}
\title{return a matrix of centroids of an object}
\usage{
centroids(x, ...)

\S4method{centroids}{ClusteredNeuroVol}(x, type = c("center_of_mass",
  "medoid"))
}
\arguments{
\item{x}{an object with multiple centroids (e.g. a \code{ClusteredNeuroVol})}

\item{...}{extra args}

\item{type}{the type of center of mass: one of "center_of_mass" or "medoid"}
}
\description{
return a matrix of centroids of an object
}
