% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{close,BinaryReader-method}
\alias{close,BinaryReader-method}
\alias{close,BinaryWriter-method}
\title{Close a BinaryReader or BinaryWriter}
\usage{
\S4method{close}{BinaryReader}(con)

\S4method{close}{BinaryWriter}(con)
}
\arguments{
\item{con}{The BinaryReader or BinaryWriter object to close.}
}
\description{
This method closes the connection associated with a BinaryReader or BinaryWriter object.
}
