% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{clustered_searchlight}
\alias{clustered_searchlight}
\title{Create a clustered Searchlight iterator}
\usage{
clustered_searchlight(mask, csize)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{csize}{the number of clusters}
}
\value{
an \code{iter} class
}
\description{
Create a clustered Searchlight iterator
}
