% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/clustervol.R
\docType{class}
\name{ClusteredNeuroVol-class}
\alias{ClusteredNeuroVol-class}
\alias{ClusteredNeuroVol}
\title{ClusteredNeuroVol}
\usage{
ClusteredNeuroVol(mask, clusters, label_map = NULL, label = "")
}
\arguments{
\item{mask}{an instance of class \code{\linkS4class{LogicalNeuroVol}}}

\item{clusters}{a vector of clusters ids with length equal to number of nonzero voxels in mask \code{mask}}

\item{label_map}{an optional \code{list} that maps from cluster id to a cluster label, e.g. (1 -> "FFA", 2 -> "PPA")}

\item{label}{an optional \code{character} string used to label of the volume}
}
\value{
\code{\linkS4class{ClusteredNeuroVol}} instance
}
\description{
Three-dimensional brain image that is divided into N disjoint partitions

Construct a \code{\linkS4class{ClusteredNeuroVol}} instance
}
\section{Slots}{

\describe{
\item{\code{mask}}{the \code{logical} mask indicating the spatial domain of the set of clusters}

\item{\code{clusters}}{an integer index indicating the cluster number for each voxel in the mask}

\item{\code{label_map}}{a list mapping from name to cluster id}

\item{\code{cluster_map}}{an \code{environment} mapping from cluster id to the set of 1D spatial indices}
}}

\examples{

bspace <- NeuroSpace(c(16,16,16), spacing=c(1,1,1))
grid <- index_to_grid(bspace, 1:(16*16*16))
kres <- kmeans(grid, centers=10)
mask <- NeuroVol(rep(1, 16^3),bspace)
clusvol <- ClusteredNeuroVol(mask, kres$cluster)
}
