% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Compare,SparseNeuroVol,numeric-method}
\alias{Compare,SparseNeuroVol,numeric-method}
\alias{Compare,numeric,SparseNeuroVol-method}
\alias{Compare,NeuroVec,NeuroVec-method}
\title{Compare a SparseNeuroVol object with a numeric value}
\usage{
\S4method{Compare}{SparseNeuroVol,numeric}(e1, e2)

\S4method{Compare}{numeric,SparseNeuroVol}(e1, e2)

\S4method{Compare}{NeuroVec,NeuroVec}(e1, e2)
}
\arguments{
\item{e1}{A NeuroVec object to be compared.}

\item{e2}{A NeuroVec object to be compared.}
}
\value{
The result of the comparison between the SparseNeuroVol object's data and the numeric value.

The result of the comparison between the numeric value and the SparseNeuroVol object's data.

The result of the comparison between \code{e1} and \code{e2}.
}
\description{
This method compares the data of a SparseNeuroVol object (\code{e1}) with a numeric value (\code{e2}) using a generic comparison function.

This method compares a numeric value (\code{e1}) with the data of a SparseNeuroVol object (\code{e2}) using a generic comparison function.

This method compares two NeuroVec objects (\code{e1} and \code{e2}) using a generic comparison function.
The dimensions of both objects are checked for compatibility before performing the comparison.
}
