% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Compare-methods}
\alias{Compare-methods}
\alias{Compare,SparseNeuroVol,numeric-method}
\alias{Compare,numeric,SparseNeuroVol-method}
\alias{Compare,NeuroVec,NeuroVec-method}
\title{Comparison Operations}
\usage{
\S4method{Compare}{SparseNeuroVol,numeric}(e1, e2)

\S4method{Compare}{numeric,SparseNeuroVol}(e1, e2)

\S4method{Compare}{NeuroVec,NeuroVec}(e1, e2)
}
\arguments{
\item{e1}{A NeuroVec object to be compared.}

\item{e2}{A NeuroVec object to be compared.}
}
\value{
The result of the comparison between the SparseNeuroVol object's data and the numeric value.

The result of the comparison between \code{e1} and \code{e2}.
}
\description{
Methods for comparing neuroimaging objects

This method compares two NeuroVec objects (\code{e1} and \code{e2}) using a generic comparison function.
The dimensions of both objects are checked for compatibility before performing the comparison.
}
