% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{conn_comp}
\alias{conn_comp}
\alias{conn_comp,NeuroVol-method}
\title{Connected components}
\usage{
conn_comp(x, ...)

\S4method{conn_comp}{NeuroVol}(
  x,
  threshold = 0,
  cluster_table = TRUE,
  local_maxima = TRUE,
  local_maxima_dist = 15,
  ...
)
}
\arguments{
\item{x}{the image object}

\item{...}{additional arguments}

\item{threshold}{threshold defining lower intensity bound for image mask}

\item{cluster_table}{return cluster_table}

\item{local_maxima}{return table of local maxima}

\item{local_maxima_dist}{the distance used to define minum distance between local maxima}
}
\value{
An object representing the connected components of \code{x}.
}
\description{
Find connected components in an image
}
