% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{coord_to_grid}
\alias{coord_to_grid}
\title{Generic function to convert N-dimensional real world coordinates to grid coordinates}
\usage{
coord_to_grid(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of real world coordinates}
}
\value{
a matrix of grid coordinates
}
\description{
Generic function to convert N-dimensional real world coordinates to grid coordinates
}

