% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R, R/neurovol.R
\docType{methods}
\name{coord_to_grid}
\alias{coord_to_grid}
\alias{coord_to_grid,NeuroSpace,matrix-method}
\alias{coord_to_grid,NeuroSpace,numeric-method}
\alias{coord_to_grid,NeuroVol,matrix-method}
\title{Generic function to convert N-dimensional real world coordinates to grid coordinates}
\usage{
coord_to_grid(x, coords)

\S4method{coord_to_grid}{NeuroSpace,matrix}(x, coords)

\S4method{coord_to_grid}{NeuroSpace,numeric}(x, coords)

\S4method{coord_to_grid}{NeuroVol,matrix}(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of real world coordinates}
}
\value{
a matrix of grid coordinates
}
\description{
Generic function to convert N-dimensional real world coordinates to grid coordinates
}
