% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{coord_to_index}
\alias{coord_to_index}
\title{Generic function to convert N-dimensional real world coordinates to 1D indices}
\usage{
coord_to_index(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of real world coordinates}
}
\value{
a vector of indices
}
\description{
Generic function to convert N-dimensional real world coordinates to 1D indices
}

