% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/roi.R, R/sparse_neurovec.R
\name{coords,IndexLookupVol-method}
\alias{coords,IndexLookupVol-method}
\alias{coords,ROICoords-method}
\alias{coords,AbstractSparseNeuroVec-method}
\title{Extract Coordinates from an IndexLookupVol Object}
\usage{
\S4method{coords}{IndexLookupVol}(x, i)

\S4method{coords}{ROICoords}(x, real = FALSE)

\S4method{coords}{AbstractSparseNeuroVec}(x, i)
}
\arguments{
\item{x}{An IndexLookupVol object to extract coordinates from.}

\item{i}{The index into the lookup volume.}

\item{real}{if \code{TRUE}, return coordinates in real world units}
}
\value{
The extracted coordinates corresponding to the provided index.
        If the index is not found, it returns NA.
}
\description{
This method extracts the coordinates from an IndexLookupVol object based on a given index.
}
\examples{
\dontrun{
space <- NeuroSpace(c(64, 64, 64), c(1, 1, 1), c(0, 0, 0))
indices <- sample(1:262144, 10000)
ilv <- IndexLookupVol(space, indices)
coords(ilv, 1000)
}

}
