% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/roi.R, R/sparse_neurovec.R
\name{coords,IndexLookupVol-method}
\alias{coords,IndexLookupVol-method}
\alias{coords,ROICoords-method}
\alias{coords,AbstractSparseNeuroVec-method}
\title{Extract coordinates from an IndexLookupVol object}
\usage{
\S4method{coords}{IndexLookupVol}(x, i)

\S4method{coords}{ROICoords}(x, real = FALSE)

\S4method{coords}{AbstractSparseNeuroVec}(x, i)
}
\arguments{
\item{x}{An IndexLookupVol object to extract coordinates from.}

\item{i}{The index into the lookup volume.}

\item{real}{if \code{TRUE}, return coordinates in real world units}
}
\value{
The extracted coordinates corresponding to the provided index. If the index is not found, it returns NA.
}
\description{
This method extracts the coordinates from an IndexLookupVol object based on a given index.
}
