% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/index_vol.R, R/roi.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{coords}
\alias{coords}
\alias{coords,IndexLookupVol-method}
\alias{coords,ROIVol-method}
\alias{coords,SparseNeuroVec-method}
\title{Extract coordinates}
\usage{
coords(x, ...)

\S4method{coords}{IndexLookupVol}(x, i)

\S4method{coords}{ROIVol}(x, real = FALSE)

\S4method{coords}{SparseNeuroVec}(x, i)
}
\arguments{
\item{x}{the object to extract coordinates from}

\item{...}{additional arguments}

\item{i}{the index in to the lookup volume}

\item{real}{if \code{TRUE}, return coordinates in real world units}
}
\description{
Extract coordinates

coords
}
