% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{coords}
\alias{coords}
\title{Extract coordinates from an object}
\usage{
coords(x, ...)
}
\arguments{
\item{x}{The object to extract coordinates from.}

\item{...}{Additional arguments (not used in the generic function).}
}
\value{
A numeric \code{matrix} or \code{vector} containing the coordinates of \code{x}.
}
\description{
This function extracts the coordinates from an input object.
}
