% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{data_file}
\alias{data_file}
\title{Generic function to get the name of the data file, given a file name and a \code{\linkS4class{FileFormat}} instance.}
\usage{
data_file(x, file_name)
}
\arguments{
\item{x}{descriptor instance}

\item{file_name}{file name to be stripped of its extension}
}
\value{
the correct header name
}
\description{
Generic function to get the name of the data file, given a file name and a \code{\linkS4class{FileFormat}} instance.
}
