% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{data_file_matches}
\alias{data_file_matches}
\title{Generic function to test whether a file name conforms to the given a \code{\linkS4class{FileFormat}} instance.
Will test for match to data file only}
\usage{
data_file_matches(x, file_name)
}
\arguments{
\item{x}{object for which the file name is to matched to}

\item{file_name}{file name to be matched}
}
\value{
TRUE for match, FALSE otherwise
}
\description{
Generic function to test whether a file name conforms to the given a \code{\linkS4class{FileFormat}} instance.
Will test for match to data file only
}
