% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{data_reader}
\alias{data_reader}
\alias{data_reader,NIFTIMetaInfo-method}
\alias{data_reader,AFNIMetaInfo-method}
\title{Create a Data Reader}
\usage{
data_reader(x, offset)

\S4method{data_reader}{NIFTIMetaInfo}(x, offset = 0)

\S4method{data_reader}{AFNIMetaInfo}(x, offset = 0)
}
\arguments{
\item{x}{An object specifying the information required to produce the reader.
This could be a file path, a metadata object, or any other relevant information
depending on the specific method implementation.}

\item{offset}{A numeric value specifying the byte offset (number of bytes to skip
before reading). This allows for flexibility in reading data from different
parts of a file or data stream.}
}
\value{
A data reader object. The exact type and structure of this object will
  depend on the specific method implementation.
}
\description{
This generic function creates a data reader for neuroimaging data. It is designed
to be flexible and work with various types of neuroimaging data formats.
}
\details{
The `data_reader` function is a generic that should be implemented for different
types of neuroimaging data formats (e.g., NIFTI, AFNI). Each implementation should
return an appropriate reader object that can be used to access the data.
}
\examples{
\dontrun{
# Example usage (specific implementation may vary):
meta_info <- read_header("brain_scan.nii")
reader <- data_reader(meta_info, offset = 0)
}

}
\seealso{
\code{\link{read_vol}}, \code{\link{FileMetaInfo-class}}
}
