% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{data_reader,NIFTIMetaInfo-method}
\alias{data_reader,NIFTIMetaInfo-method}
\alias{data_reader,AFNIMetaInfo-method}
\title{Create Data Reader for AFNI Format}
\usage{
\S4method{data_reader}{NIFTIMetaInfo}(x, offset = 0)

\S4method{data_reader}{AFNIMetaInfo}(x, offset = 0)
}
\arguments{
\item{x}{AFNIMetaInfo object}

\item{offset}{Numeric byte offset}
}
\value{
BinaryReader object
}
\description{
Create Data Reader for AFNI Format
}
