% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R
\name{dim,FileMetaInfo-method}
\alias{dim,FileMetaInfo-method}
\title{dim of \code{FileMetaInfo}}
\usage{
\S4method{dim}{FileMetaInfo}(x)
}
\arguments{
\item{x}{the object}
}
\description{
dim of \code{FileMetaInfo}
}
