% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info.R, R/neuro_obj.R, R/roi.R
\name{dim,FileMetaInfo-method}
\alias{dim,FileMetaInfo-method}
\alias{dim,NeuroObj-method}
\alias{dim,ROICoords-method}
\alias{dim}
\alias{dim,ROIVol-method}
\title{Get Dimensions of FileMetaInfo Object}
\usage{
\S4method{dim}{FileMetaInfo}(x)

\S4method{dim}{NeuroObj}(x)

\S4method{dim}{ROICoords}(x)

\S4method{dim}{ROIVol}(x)

\S4method{dim}{ROICoords}(x)
}
\arguments{
\item{x}{the object}
}
\value{
A numeric vector of length 2 containing the dimensions of the ROICoords object.
}
\description{
Get Dimensions of FileMetaInfo Object

dim of \code{NeuroObj} object

Get dimensions of an object
}
