% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{dim_of}
\alias{dim_of}
\alias{dim_of,NeuroSpace,NamedAxis-method}
\title{get dimensions of an axis}
\usage{
dim_of(x, axis)

\S4method{dim_of}{NeuroSpace,NamedAxis}(x, axis)
}
\arguments{
\item{x}{the object}

\item{axis}{the axis to return the dimension of}
}
\description{
get dimensions of an axis
}
