% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{dim_of}
\alias{dim_of}
\alias{dim_of,NeuroSpace,NamedAxis-method}
\title{dim_of}
\usage{
dim_of(x, axis)

\S4method{dim_of}{NeuroSpace,NamedAxis}(x, axis)
}
\description{
dim_of
}
