% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{.read_mmap}
\alias{.read_mmap}
\title{Read Data Using Memory Mapping}
\usage{
.read_mmap(meta, idx)
}
\arguments{
\item{meta}{An object of class \linkS4class{ImageMetadata} containing file metadata}

\item{idx}{Integer vector of indices specifying elements to read}
}
\value{
A numeric vector of values read from the memory-mapped file
}
\description{
Read data from a memory-mapped file based on metadata and indices.
}
\keyword{internal}
