% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{drop,NeuroVec-method}
\alias{drop,NeuroVec-method}
\title{Drop singleton dimensions from a NeuroVec object}
\usage{
\S4method{drop}{NeuroVec}(x)

\S4method{drop}{NeuroVec}(x)
}
\arguments{
\item{x}{The NeuroVec object.}
}
\description{
This function drops the last dimension of a NeuroVec object if it is of length 1.

This function drops the last dimension of a NeuroVec object if it is of length 1.
}
