% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{drop_dim,AxisSet3D,numeric-method}
\alias{drop_dim,AxisSet3D,numeric-method}
\title{Drop dimension from axis set}
\usage{
\S4method{drop_dim}{AxisSet3D,numeric}(x, dimnum)
}
\arguments{
\item{x}{An AxisSet3D object}

\item{dimnum}{Numeric index of dimension to drop}
}
\value{
An object of the same class as the input with the specified dimension removed.
}
\description{
Drop dimension from axis set
}
