% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/roi.R
\name{embed_kernel}
\alias{embed_kernel}
\alias{embed_kernel,Kernel,NeuroSpace,numeric-method}
\title{Generic function to position kernel in a postion in image space}
\usage{
embed_kernel(x, sp, center_voxel, ...)

\S4method{embed_kernel}{Kernel,NeuroSpace,numeric}(x, sp, center_voxel, weight = 1)
}
\arguments{
\item{x}{the kernel object}

\item{sp}{the space to embed the kernel}

\item{center_voxel}{the voxel marking the center of the kernel in the embedded space}

\item{weight}{multiply all kernel weight by this value}
}
\description{
Generic function to position kernel in a postion in image space
}
