% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_like.R, R/neurovol.R, R/roi.R
\name{[,DenseNeuroVol,numeric,missing,ANY-method}
\alias{[,DenseNeuroVol,numeric,missing,ANY-method}
\alias{[,DenseNeuroVol,integer,missing,ANY-method}
\alias{[,NeuroVol,ROIVol,missing,ANY-method}
\alias{[,DenseNeuroVol,ROIVol,missing,ANY-method}
\alias{[}
\alias{[,NeuroVol,ROICoords,missing,ANY-method}
\alias{[,ROICoords,numeric,missing,ANY-method}
\alias{[,ROIVol,ROICoords,missing,ANY-method}
\alias{[,ROIVol,ROICoords,numeric,ANY-method}
\alias{[,ROIVol,logical,missing,ANY-method}
\alias{[,ROIVol,logical,numeric,ANY-method}
\alias{[,ROIVol,matrix,missing,ANY-method}
\alias{[,ROIVol,matrix,numeric,ANY-method}
\alias{[,ROIVol,missing,missing,ANY-method}
\alias{[,ROIVol,missing,numeric,ANY-method}
\alias{[,ROIVol,numeric,numeric,ANY-method}
\alias{[,SparseNeuroVol,numeric,numeric,ANY-method}
\alias{[,ROIVol,numeric,missing,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{DenseNeuroVol,numeric,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{DenseNeuroVol,integer,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{NeuroVol,ROIVol,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{DenseNeuroVol,ROIVol,missing,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ROIVol,logical,missing,ANY}(x, i, j, k, ..., drop = TRUE)
}
\arguments{
\item{x}{The object to extract from}

\item{i}{Index specifying elements to extract}

\item{j}{Second index (if applicable)}

\item{k}{Third index for 3D objects (if applicable)}

\item{...}{Additional arguments passed to methods}

\item{drop}{Whether to drop dimensions of length 1}
}
\value{
A subset of the input object, with dimensions depending on the indexing and the `drop` parameter.
}
\description{
Extract or replace parts of an object
}
