% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{[}
\alias{[}
\alias{[,NeuroVol,ROICoords,missing,ANY-method}
\alias{[,ROICoords,numeric,missing,ANY-method}
\alias{[,ROIVol,ROICoords,missing,ANY-method}
\alias{[,ROIVol,ROICoords,numeric,ANY-method}
\alias{[,ROIVol,logical,numeric,ANY-method}
\alias{[,ROIVol,matrix,missing,ANY-method}
\alias{[,ROIVol,matrix,numeric,ANY-method}
\alias{[,ROIVol,missing,missing,ANY-method}
\alias{[,ROIVol,missing,numeric,ANY-method}
\alias{[,ROIVol,numeric,numeric,ANY-method}
\alias{[,SparseNeuroVol,numeric,numeric,ANY-method}
\alias{[,ROIVol,numeric,missing,ANY-method}
\alias{[,ROIVol,logical,missing,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ROIVol,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ROIVol,logical,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{...}{additional arguments}

\item{drop}{drop dimension}
}
\value{
A subset of the input object, with dimensions depending on the indexing and the `drop` parameter.
}
\description{
Extract or replace parts of an object

subset an \code{ROIVol}
}
