% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/file_format.R
\name{file_matches}
\alias{file_matches}
\alias{file_matches,FileFormat,character-method}
\title{Generic function to test whether a file name conforms to the given \code{\linkS4class{FileFormat}} instance.
Will test for match to either header file or data file}
\usage{
file_matches(x, file_name)

\S4method{file_matches}{FileFormat,character}(x, file_name)
}
\arguments{
\item{x}{object for which the file name is to matched to}

\item{file_name}{file name to be matched}
}
\value{
TRUE for match, FALSE otherwise
}
\description{
Generic function to test whether a file name conforms to the given \code{\linkS4class{FileFormat}} instance.
Will test for match to either header file or data file
}
