% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FileBackedNeuroVec-class}
\alias{FileBackedNeuroVec-class}
\title{FileBackedNeuroVec}
\description{
a four-dimensional brain image that is read in to memory "on demand" using memory-mapped file access.
}
\section{Slots}{

\describe{
\item{\code{meta}}{the file meta information of type \code{\linkS4class{FileMetaInfo}}}
}}

