% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R
\name{FileBackedNeuroVec}
\alias{FileBackedNeuroVec}
\title{Constuct a \code{FileBackedNeuroVec} instance}
\usage{
FileBackedNeuroVec(file_name)
}
\arguments{
\item{the}{name of the image file}
}
\description{
Constuct a \code{FileBackedNeuroVec} instance
}
