% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R
\name{FileBackedNeuroVec}
\alias{FileBackedNeuroVec}
\title{Create a FileBackedNeuroVec Object}
\usage{
FileBackedNeuroVec(file_name)
}
\arguments{
\item{file_name}{A character string specifying the name of the image file.}
}
\value{
A new instance of class \code{FileBackedNeuroVec}.
}
\description{
Constructs a \code{FileBackedNeuroVec} instance, which represents a file-backed
neuroimaging vector object.
}
\details{
This function reads the header of the specified file, creates a NeuroSpace object,
and initializes a FileBackedNeuroVec with the appropriate metadata.
}
\examples{
\dontrun{
fbvec <- FileBackedNeuroVec("path/to/image.nii")
}

}
\seealso{
\code{\link{NeuroSpace}}, \code{\link{read_header}}
}
