% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filebacked_neurovec.R
\name{FileBackedNeuroVec}
\alias{FileBackedNeuroVec}
\title{FileBackedNeuroVec}
\usage{
FileBackedNeuroVec(file_name)
}
\arguments{
\item{file_name}{the name of the image file}
}
\description{
Construct a \code{FileBackedNeuroVec} instance
}
