% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FileFormat-class}
\alias{FileFormat-class}
\title{FileFormatDescriptor}
\description{
This class represents a neuroimaging file format
}
\section{Slots}{

\describe{
\item{\code{file_format}}{the name of the file format (e.g. NIfTI)}

\item{\code{header_encoding}}{the file encoding of the header file (e.g. 'raw' for binary, 'gzip' for gz compressed')}

\item{\code{header_extension}}{the file extension for the header file (e.g. 'nii' for NIfTI single files)}

\item{\code{data_encoding}}{the file encoding for the data file}

\item{\code{data_extension}}{the file extension for the data file (e.g. 'nii' for NIfTI single files)}
}}

