% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/meta_info.R
\docType{class}
\name{FileMetaInfo-class}
\alias{FileMetaInfo-class}
\alias{NIFTIMetaInfo-class}
\alias{AFNIMetaInfo-class}
\alias{NIFTIMetaInfo}
\alias{AFNIMetaInfo}
\title{FileMetaInfo}
\usage{
NIFTIMetaInfo(descriptor, nifti_header)

AFNIMetaInfo(descriptor, afni_header)
}
\arguments{
\item{descriptor}{an instance of class \code{AFNIFormat}}

\item{nifti_header}{a \code{list} returned by \code{readNIftiHeader}}

\item{afni_header}{a \code{list} returned by \code{read_afni_header}}
}
\value{
an instance of class \code{\linkS4class{NIFTIMetaInfo}}

an instance of class \code{AFNIMetaInfo}
}
\description{
This class contains meta information from an image data file

This class contains meta information for a NIfTI image file

This class contains meta information for a AFNI image file.

Constructor for \code{AFNIMetaInfo} class
}
\section{Slots}{

\describe{
\item{\code{header_file}}{name of the file containing meta information}

\item{\code{data_file}}{name of the file containing data}

\item{\code{descriptor}}{descriptor of image file format}

\item{\code{endian}}{byte order of data ('little' or 'big')}

\item{\code{data_offset}}{the number of bytes preceding the start of image data in data file}

\item{\code{bytes_per_element}}{number of bytes per element}

\item{\code{intercept}}{constant value added to image -- multiple values allowed (must equal numer of sub-images)}

\item{\code{slope}}{image multiplier -- multiple values allowed (must equal numer of sub-images)}

\item{\code{header}}{a list of format specific attributes}

\item{\code{nifti_header}}{a \code{list} of attributes specific to the NIfTI file format}

\item{\code{afni_header}}{a list of attributes specific to the AFNI file format}

\item{\code{afni_header}}{a \code{list} of attributes specific to the AFNI file format}
}}

