% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_filter.R
\name{gaussian_blur}
\alias{gaussian_blur}
\title{Blur a volumetric image with an isotropic discrete Gaussian kernel}
\usage{
gaussian_blur(vol, mask, sigma = 2, window = 1)
}
\arguments{
\item{vol}{A \code{\linkS4class{NeuroVol}} object representing the image volume to be smoothed.}

\item{mask}{An optional \code{\linkS4class{LogicalNeuroVol}} object representing the image mask that defines the region where the blurring is applied. If not provided, the entire volume is considered.}

\item{sigma}{A numeric value specifying the standard deviation of the Gaussian kernel (default is 2).}

\item{window}{An integer specifying the number of voxels around the center voxel to include on each side. For example, window=1 for a 3x3x3 kernel (default is 1).}
}
\value{
A smoothed image of class \code{\linkS4class{NeuroVol}}.
}
\description{
This function smooths a volumetric image (3D brain MRI data) by applying an isotropic discrete Gaussian kernel.
The blurring is performed within the specified image mask, and the kernel's standard deviation and window size can be customized.
}
\examples{
# Load an example brain volume
brain_vol <- read_vol(system.file("extdata", "example_volume.nii", package="neuroim2"))
# Load an example brain mask
brain_mask <- read_vol(system.file("extdata", "global_mask.nii", package="neuroim2"))

# Apply Gaussian blurring to the brain volume
blurred_vol <- gaussian_blur(brain_vol, brain_mask, sigma = 2, window = 1)

}
