% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_filter.R
\name{gaussian_blur}
\alias{gaussian_blur}
\title{Blur a volumetric image with an isotropic discrete Gaussian kernel}
\usage{
gaussian_blur(vol, mask, sigma = 2, window = 1)
}
\arguments{
\item{sigma}{the standard deviation of the Gaussian}

\item{window}{the number of voxels around the center voxel to include on each side (window=1 for a 3x3x3 kernel).}

\item{the}{image volume as a \code{NeuroVol}}

\item{an}{image mask as a \code{LogicalNeuroVol}}
}
\value{
a smoothed image of class \code{NeuroVol}
}
\description{
Blur a volumetric image with an isotropic discrete Gaussian kernel
}
