% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{grid_to_coord}
\alias{grid_to_coord}
\alias{grid_to_coord,NeuroSpace,matrix-method}
\alias{grid_to_coord,NeuroSpace,numeric-method}
\alias{grid_to_coord,NeuroVol,matrix-method}
\title{Generic function to convert N-dimensional grid coordinate coordinates to real world coordinates}
\usage{
grid_to_coord(x, coords)

\S4method{grid_to_coord}{NeuroSpace,matrix}(x, coords)

\S4method{grid_to_coord}{NeuroSpace,numeric}(x, coords)

\S4method{grid_to_coord}{NeuroVol,matrix}(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of grid coordinates}
}
\value{
a matrix of real coordinates
}
\description{
Generic function to convert N-dimensional grid coordinate coordinates to real world coordinates
}
