% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{grid_to_grid}
\alias{grid_to_grid}
\alias{grid_to_grid,NeuroSpace,matrix-method}
\alias{grid_to_grid,matrix,matrix-method}
\title{Generic function to convert voxel coordinates in the reference space (LPI) to native array space.}
\usage{
grid_to_grid(x, vox)

\S4method{grid_to_grid}{NeuroSpace,matrix}(x, vox)

\S4method{grid_to_grid}{matrix,matrix}(x, vox)
}
\arguments{
\item{x}{the object}

\item{vox}{a matrix of LPI voxel coordinates}
}
\value{
a matrix of native voxel coordinates
}
\description{
Generic function to convert voxel coordinates in the reference space (LPI) to native array space.
}
