% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuroslice.R, R/neurospace.R
\name{grid_to_index}
\alias{grid_to_index}
\alias{grid_to_index,NeuroSlice,matrix-method}
\alias{grid_to_index,NeuroSlice,numeric-method}
\alias{grid_to_index,NeuroSpace,matrix-method}
\alias{grid_to_index,NeuroSpace,numeric-method}
\alias{grid_to_index,NeuroVol,matrix-method}
\alias{grid_to_index,NeuroVol,numeric-method}
\title{Generic function to convert N-dimensional grid coordinate to 1D indices}
\usage{
grid_to_index(x, coords)

\S4method{grid_to_index}{NeuroSlice,matrix}(x, coords)

\S4method{grid_to_index}{NeuroSlice,numeric}(x, coords)

\S4method{grid_to_index}{NeuroSpace,matrix}(x, coords)

\S4method{grid_to_index}{NeuroSpace,numeric}(x, coords)

\S4method{grid_to_index}{NeuroVol,matrix}(x, coords)

\S4method{grid_to_index}{NeuroVol,numeric}(x, coords)
}
\arguments{
\item{x}{the object, typically a \code{NeuroVol} or \code{NeuroSpace} instance.}

\item{coords}{a matrix where each row is a coordinate or a vector of length equal to \code{ndim(x)}}
}
\value{
a vector of indices
}
\description{
Generic function to convert N-dimensional grid coordinate to 1D indices
}
