% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/file_format.R
\name{header_file}
\alias{header_file}
\alias{header_file,FileFormat,character-method}
\title{Generic function to get the name of the header file, given a file name and a \code{\linkS4class{FileFormat}} instance.}
\usage{
header_file(x, file_name)

\S4method{header_file}{FileFormat,character}(x, file_name)
}
\arguments{
\item{x}{descriptor instance}

\item{file_name}{file name to be stripped of its extension}
}
\value{
the correct header name
}
\description{
Generic function to get the name of the header file, given a file name and a \code{\linkS4class{FileFormat}} instance.
}
