% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{index_to_coord}
\alias{index_to_coord}
\alias{index_to_coord,NeuroSpace,numeric-method}
\alias{index_to_coord,NeuroVol,index-method}
\title{Generic function to convert 1D indices to N-dimensional real world coordinates}
\usage{
index_to_coord(x, idx)

\S4method{index_to_coord}{NeuroSpace,numeric}(x, idx)

\S4method{index_to_coord}{NeuroVol,index}(x, idx)
}
\arguments{
\item{x}{the object}

\item{idx}{the 1D indices}
}
\value{
a matrix of real coordinates
}
\description{
Generic function to convert 1D indices to N-dimensional real world coordinates
}
\examples{
bvol <- NeuroVol(array(0, c(10,10,10)), NeuroSpace(c(10,10,10), c(1,1,1)))
idx <- 1:10
g <- indexToCoord(bvol, idx)
idx2 <- coord_to_index(bvol, g)
all.equal(idx, idx2)
}
