% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuroslice.R, R/neurospace.R
\name{index_to_grid}
\alias{index_to_grid}
\alias{index_to_grid,NeuroSlice,numeric-method}
\alias{index_to_grid,NeuroSpace,numeric-method}
\alias{index_to_grid,NeuroVec,index-method}
\alias{index_to_grid,NeuroVec,integer-method}
\alias{index_to_grid,NeuroVol,index-method}
\alias{index_to_grid,NeuroVol,integer-method}
\title{convert 1d indices to n-dimensional grid coordinates}
\usage{
index_to_grid(x, idx)

\S4method{index_to_grid}{NeuroSlice,numeric}(x, idx)

\S4method{index_to_grid}{NeuroSpace,numeric}(x, idx)

\S4method{index_to_grid}{NeuroVec,index}(x, idx)

\S4method{index_to_grid}{NeuroVec,integer}(x, idx)

\S4method{index_to_grid}{NeuroVol,index}(x, idx)

\S4method{index_to_grid}{NeuroVol,integer}(x, idx)
}
\arguments{
\item{x}{the object}

\item{idx}{the 1d \code{vector} of indices}
}
\value{
a matrix of grid coordinates
}
\description{
convert 1d indices to n-dimensional grid coordinates
}
\examples{

 bvol <- NeuroVol(array(0, c(10,10,10)), NeuroSpace(c(10,10,10), c(1,1,1)))
 idx <- 1:10
 g <- index_to_grid(bvol, idx)
 bvol[g]

}
