% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/index_vol.R
\docType{class}
\name{IndexLookupVol-class}
\alias{IndexLookupVol-class}
\alias{IndexLookupVol}
\title{IndexLookupVol}
\usage{
IndexLookupVol(space, indices)
}
\arguments{
\item{space}{a NeuroSpace object}

\item{indices}{the set of 1-d indices defining the lookup map}
}
\description{
Three-dimensional brain image that can be used as a map between 1D grid indices and a table of values
Currently used in the \code{\linkS4class{SparseNeuroVec}} class.
}
