% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/index_vol.R
\docType{class}
\name{IndexLookupVol-class}
\alias{IndexLookupVol-class}
\alias{IndexLookupVol}
\title{IndexLookupVol Class}
\usage{
IndexLookupVol(space, indices)
}
\arguments{
\item{space}{A \code{\linkS4class{NeuroSpace}} object representing the 3D space of the brain image.}

\item{indices}{An integer vector containing the 1D indices of the voxels in the grid.}

\item{map}{An integer vector containing the mapping between the 1D indices and the table of values.}
}
\value{
An object of class \code{\linkS4class{IndexLookupVol}} representing the index lookup volume.

An object of class \code{\linkS4class{IndexLookupVol}} representing the index lookup volume.
}
\description{
A three-dimensional brain image class that serves as a map between 1D grid indices and a table of values.
It is primarily used in the \code{\linkS4class{SparseNeuroVec}} class.
}
\section{Constructor}{

\preformatted{
IndexLookupVol(space, indices, map)
}
}

\section{Slots}{

\describe{
  \item{\code{space}:}{A \code{\linkS4class{NeuroSpace}} object representing the 3D space of the brain image.}
  \item{\code{indices}:}{An integer vector containing the 1D indices of the voxels in the grid.}
  \item{\code{map}:}{An integer vector containing the mapping between the 1D indices and the table of values.}
}
}

\examples{
# Create a NeuroSpace object
space <- NeuroSpace(c(2, 2, 2, 10), c(1, 1, 1))

# Create a 3D mask
mask <- array(rnorm(2 * 2 * 2) > -100, c(2, 2, 2))

# Create indices and map for the IndexLookupVol
indices <- which(mask)
map <- seq_along(indices)

# Create an IndexLookupVol object
ilv <- IndexLookupVol(space, indices, map)
}
\seealso{
\code{\linkS4class{SparseNeuroVec}}

\code{\linkS4class{IndexLookupVol}}
}
