% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/roi.R, R/sparse_neurovec.R
\name{indices,IndexLookupVol-method}
\alias{indices,IndexLookupVol-method}
\alias{indices,ROIVol-method}
\alias{indices,ROIVec-method}
\alias{indices,AbstractSparseNeuroVec-method}
\title{Get Indices from an IndexLookupVol Object}
\usage{
\S4method{indices}{IndexLookupVol}(x)

\S4method{indices}{ROIVol}(x)

\S4method{indices}{ROIVec}(x)

\S4method{indices}{AbstractSparseNeuroVec}(x)
}
\arguments{
\item{x}{An IndexLookupVol object.}
}
\value{
An integer vector of indices.
}
\description{
This method retrieves the indices stored in an IndexLookupVol object.
}
\keyword{internal}
