% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/index_vol.R, R/roi.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{indices}
\alias{indices}
\alias{indices,IndexLookupVol-method}
\alias{indices,ROIVol-method}
\alias{indices,ROIVec-method}
\alias{indices,AbstractSparseNeuroVec-method}
\title{Extract indices}
\usage{
indices(x)

\S4method{indices}{IndexLookupVol}(x)

\S4method{indices}{ROIVol}(x)

\S4method{indices}{ROIVec}(x)

\S4method{indices}{AbstractSparseNeuroVec}(x)
}
\arguments{
\item{x}{the object to extract indices}
}
\description{
Extract indices

indices
}
