% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R
\name{initialize,IndexLookupVol-method}
\alias{initialize,IndexLookupVol-method}
\title{Initialize an IndexLookupVol Object}
\usage{
\S4method{initialize}{IndexLookupVol}(.Object, space, indices)
}
\arguments{
\item{.Object}{The IndexLookupVol object to initialize.}

\item{space}{A NeuroSpace object representing the 3D space.}

\item{indices}{An integer vector of indices.}
}
\value{
An initialized IndexLookupVol object.
}
\description{
This method initializes an IndexLookupVol object with the given space and indices.
}
\keyword{internal}
