% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/latent_neurovec.R
\docType{class}
\name{LatentNeuroVec-class}
\alias{LatentNeuroVec-class}
\alias{LatentNeuroVec}
\title{LatentNeuroVec}
\usage{
LatentNeuroVec(basis, loadings, space, mask, offset = NULL)
}
\arguments{
\item{basis}{A numeric n-by-k matrix containing the latent vectors forming the reduced space.}

\item{loadings}{A numeric p-by-k matrix of p loadings.}

\item{space}{A NeuroSpace instance defining the dimensions, spacing, origin, axes, and transformation of the neuroimaging space.}

\item{mask}{A 3D logical array, 1D logical vector, or an instance of LogicalNeuroVol class representing the brain mask.}

\item{offset}{An optional numeric 1-by-p offset vector. If not provided, it defaults to a zero vector.}
}
\value{
A new \code{\linkS4class{LatentNeuroVec}} instance representing the latent neuroimaging vectors.
}
\description{
a class that stores a represents a 4-dimensional array as a set of basis functions (dictionary) and
corresponding set of loadings (coefficients).

This function constructs a LatentNeuroVec object from a basis, associated loadings, a NeuroSpace instance, a mask, and an optional offset.
}
\section{Slots}{

\describe{
\item{\code{basis}}{the matrix of bases, were each column is a basis vector.}

\item{\code{loadings}}{the \code{sparseMatrix} of loadings}

\item{\code{offset}}{an offset vector}
}}

\section{Usage}{

\preformatted{
LatentNeuroVec(basis, loadings, space, mask, offset = NULL)
}
}

\section{Examples}{

\preformatted{
bspace <- NeuroSpace(c(2,2,2,10), c(1,1,1))
mask <- array(rnorm(2*2*2) > -100, c(2,2,2))
mat <- matrix(rnorm(sum(mask)), 10, sum(mask))
pres <- prcomp(mat)
svec <- LatentNeuroVec(pres$x, pres$rotation, bspace, mask, offset=colMeans(mat))
svec2 <- SparseNeuroVec(mat, bspace, mask)
length(indices(svec)) == sum(mask)

all.equal(svec2[1:prod(dim(mask))], svec[1:prod(dim(mask))])
}
}

