% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/latent_neurovec.R
\docType{class}
\name{LatentNeuroVec-class}
\alias{LatentNeuroVec-class}
\alias{LatentNeuroVec}
\title{LatentNeuroVec}
\usage{
LatentNeuroVec(basis, loadings, space, mask, offset = NULL)
}
\arguments{
\item{basis}{an n-by-k matrix containing the latent vectors forming the reduced space}

\item{loadings}{a p-by-k matrix of p loadings}

\item{space}{a NeuroSpace instance}

\item{mask}{a 3D \code{array}, 1D \code{vector} of type \code{logical}, or an instance of type \code{LogicalNeuroVol}}

\item{offset}{a 1-by-p offset vector}
}
\value{
a new \code{\linkS4class{LatentNeuroVec}} instance
}
\description{
a class that stores a represents a 4-dimensional array as a set of basis functions (dictionary) and
corresponding set of loadings (coefficients).

Construct a LatentNeuroVec object from a basis and associated loadings
}
\section{Slots}{

\describe{
\item{\code{basis}}{the matrix of bases, were each column is a basis vector.}

\item{\code{loadings}}{the \code{sparseMatrix} of loadings}

\item{\code{offset}}{an offset vector}
}}

\examples{

bspace <- NeuroSpace(c(2,2,2,10), c(1,1,1))
mask <- array(rnorm(2*2*2) > -100, c(2,2,2))
mat <- matrix(rnorm(sum(mask)), 10, sum(mask))
pres <- prcomp(mat)
svec <- LatentNeuroVec(pres$x, pres$rotation, bspace,mask, offset=colMeans(mat))
svec2 <- SparseNeuroVec(mat, bspace, mask)
length(indices(svec)) == sum(mask)

all.equal(svec2[1:prod(dim(mask))],svec[1:prod(dim(mask))])
}
