% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{length,NeuroVec-method}
\alias{length,NeuroVec-method}
\title{Get length of \code{NeuroVec}. This is the numbe rof volumes in the volume vector (e.g. the 4th image dimension)}
\usage{
\S4method{length}{NeuroVec}(x)
}
\description{
Get length of \code{NeuroVec}. This is the numbe rof volumes in the volume vector (e.g. the 4th image dimension)
}
