% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R, R/roi.R
\name{length,NeuroVec-method}
\alias{length,NeuroVec-method}
\alias{length,NeuroVecSeq-method}
\alias{length,ROIVol-method}
\title{Get the length of a NeuroVec object}
\usage{
\S4method{length}{NeuroVec}(x)

\S4method{length}{NeuroVecSeq}(x)

\S4method{length}{ROIVol}(x)
}
\arguments{
\item{x}{The NeuroVec object.}
}
\value{
The number of volumes in the NeuroVec object.
}
\description{
This function returns the number of volumes in a NeuroVec object.
}
