% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R, R/neurovecseq.R, R/roi.R
\name{length,NeuroVec-method}
\alias{length,NeuroVec-method}
\alias{length,NeuroVecSeq-method}
\alias{length,ROIVol-method}
\alias{length,ROICoords-method}
\title{Get length of NeuroVec object}
\usage{
\S4method{length}{NeuroVec}(x)

\S4method{length}{NeuroVecSeq}(x)

\S4method{length}{ROIVol}(x)

\S4method{length}{ROICoords}(x)

\S4method{length}{ROIVol}(x)

\S4method{length}{ROICoords}(x)

\S4method{length}{ROIVol}(x)
}
\arguments{
\item{x}{A NeuroVecSeq object}
}
\value{
Integer length (total number of time points)

An integer representing the number of coordinates in the ROICoords object.
}
\description{
Returns the number of time points (4th dimension) in a NeuroVec object.
This represents the temporal dimension of the neuroimaging data.

Returns the total number of time points across all vectors in the sequence
}
