% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovecseq.R
\name{linear_access,NeuroVecSeq,numeric-method}
\alias{linear_access,NeuroVecSeq,numeric-method}
\title{Linear Access to NeuroVecSeq Data}
\usage{
\S4method{linear_access}{NeuroVecSeq,numeric}(x, i)
}
\arguments{
\item{x}{A NeuroVecSeq object}

\item{i}{Numeric vector of indices for linear access}
}
\value{
Numeric vector of accessed values
}
\description{
Provides linear access to the data across all vectors in the sequence.
}
