% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{linear_access}
\alias{linear_access}
\title{Generic function to extract values from an array-like object using linear indexing}
\usage{
linear_access(x, i, ...)
}
\arguments{
\item{x}{a data source}

\item{...}{additional arguments}
}
\description{
Generic function to extract values from an array-like object using linear indexing
}
