% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/filebacked_neurovec.R,
%   R/mapped_neurovec.R, R/neurovec.R, R/neurovol.R, R/sparse_neurovec.R
\name{linear_access}
\alias{linear_access}
\alias{linear_access,FileBackedNeuroVec,numeric-method}
\alias{linear_access,MappedNeuroVec,numeric-method}
\alias{linear_access,NeuroVecSeq,numeric-method}
\alias{linear_access,SparseNeuroVol,numeric-method}
\alias{linear_access,AbstractSparseNeuroVec,numeric-method}
\title{extract values from an array-like object using linear indexing}
\usage{
linear_access(x, i, ...)

\S4method{linear_access}{FileBackedNeuroVec,numeric}(x, i)

\S4method{linear_access}{MappedNeuroVec,numeric}(x, i)

\S4method{linear_access}{NeuroVecSeq,numeric}(x, i)

\S4method{linear_access}{SparseNeuroVol,numeric}(x, i)

\S4method{linear_access}{AbstractSparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{a data source}

\item{i}{a vector of indices}

\item{...}{additional arguments}
}
\description{
extract values from an array-like object using linear indexing
}
