% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{linear_access}
\alias{linear_access}
\title{Extract values from an array-like object using linear indexing.}
\usage{
linear_access(x, i, ...)
}
\arguments{
\item{x}{a data source.}

\item{i}{a vector of indices.}

\item{...}{additional arguments to be passed to methods.}
}
\value{
A vector containing the values at the specified linear indices.
}
\description{
This function extracts the values of the elements in an array-like object using
linear indexing. Linear indexing is a way of indexing an array by a single index
that is computed from multiple indices using a formula.
}
