% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/mapped_neurovec.R,
%   R/neurovec.R, R/neurovol.R, R/sparse_neurovec.R
\name{load_data}
\alias{load_data}
\alias{load_data,MappedNeuroVecSource-method}
\alias{load_data,NeuroVecSource-method}
\alias{load_data,H5NeuroVecSource-method}
\alias{load_data,LatentNeuroVecSource-method}
\alias{load_data,NeuroVolSource-method}
\alias{load_data,SparseNeuroVecSource-method}
\title{load data from a data source}
\usage{
load_data(x, ...)

\S4method{load_data}{MappedNeuroVecSource}(x)

\S4method{load_data}{NeuroVecSource}(x)

\S4method{load_data}{H5NeuroVecSource}(x)

\S4method{load_data}{LatentNeuroVecSource}(x)

\S4method{load_data}{NeuroVolSource}(x)

\S4method{load_data}{SparseNeuroVecSource}(x)
}
\arguments{
\item{x}{a data source}

\item{...}{additional arguments}
}
\value{
an instance of class \code{\linkS4class{NeuroVec}}
}
\description{
load data from a data source

load_data

load a NeuroVol
}
