% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_neurovec.R, R/neurovec.R, R/neurovol.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{load_data,MappedNeuroVecSource-method}
\alias{load_data,MappedNeuroVecSource-method}
\alias{load_data,NeuroVecSource-method}
\alias{load_data,NeuroVolSource-method}
\alias{load_data,SparseNeuroVecSource-method}
\title{load_data}
\usage{
\S4method{load_data}{MappedNeuroVecSource}(x)

\S4method{load_data}{NeuroVecSource}(x)

\S4method{load_data}{NeuroVolSource}(x)

\S4method{load_data}{SparseNeuroVecSource}(x)
}
\value{
an instance of class \code{\linkS4class{NeuroVec}}
}
\description{
load_data

load a NeuroVol
}
