% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{load_data,NeuroVecSource-method}
\alias{load_data,NeuroVecSource-method}
\title{load_data}
\usage{
\S4method{load_data}{NeuroVecSource}(x)
}
\value{
an instance of class \code{\linkS4class{NeuroVec}}
}
\description{
load_data
}
