% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/mapped_neurovec.R,
%   R/neurovec.R, R/neurovol.R
\name{load_data}
\alias{load_data}
\alias{load_data,MappedNeuroVecSource-method}
\alias{load_data,NeuroVecSource-method}
\alias{load_data,H5NeuroVecSource-method}
\alias{load_data,LatentNeuroVecSource-method}
\alias{load_data,NeuroVolSource-method}
\title{Load data from a data source.}
\usage{
load_data(x, ...)

\S4method{load_data}{MappedNeuroVecSource}(x)

\S4method{load_data}{NeuroVecSource}(x)

\S4method{load_data}{H5NeuroVecSource}(x)

\S4method{load_data}{LatentNeuroVecSource}(x)

\S4method{load_data}{NeuroVolSource}(x)
}
\arguments{
\item{x}{a data source.}

\item{...}{additional arguments to be passed to methods.}
}
\value{
an instance of class \code{\linkS4class{NeuroVec}}
}
\description{
This function loads data from a data source and returns it in a format that is compatible with
other functions in the neuroim2 package. The format of the returned data depends on the type
of data source used.
}
