% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{load_data}
\alias{load_data}
\title{Load data from a data source.}
\usage{
load_data(x, ...)
}
\arguments{
\item{x}{a data source.}

\item{...}{additional arguments to be passed to methods.}
}
\value{
A data object in a format compatible with neuroim2 package functions.
}
\description{
This function loads data from a data source and returns it in a format that is compatible with
other functions in the neuroim2 package. The format of the returned data depends on the type
of data source used.
}
\keyword{internal}
