% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neurovol.R
\docType{class}
\name{LogicalNeuroVol-class}
\alias{LogicalNeuroVol-class}
\alias{LogicalNeuroVol}
\title{LogicalNeuroVol}
\usage{
LogicalNeuroVol(data, space, label = "", indices = NULL)
}
\arguments{
\item{data}{A three-dimensional \code{array}, a 1D vector with length equal to \code{prod(dim(space))}, or a set of \code{indices} where elements are \code{TRUE}.}

\item{space}{An instance of class \code{\linkS4class{NeuroSpace}}.}

\item{label}{A \code{character} string.}

\item{indices}{An optional 1-d index vector.}
}
\value{
A \code{\linkS4class{LogicalNeuroVol}} instance.
}
\description{
This S4 class represents a three-dimensional brain image, where all values are either TRUE or FALSE. It is particularly useful for creating and managing binary masks for brain images.

This function constructs a \code{\linkS4class{LogicalNeuroVol}} instance.
}
\examples{
# Load an example brain mask
brain_mask <- read_vol(system.file("extdata", "global_mask.nii", package="neuroim2"))

# Convert the brain mask to a LogicalNeuroVol
logical_vol <- LogicalNeuroVol(brain_mask, space(brain_mask))

}
