% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/sparse_neurovec.R
\name{lookup,IndexLookupVol,numeric-method}
\alias{lookup,IndexLookupVol,numeric-method}
\alias{lookup,AbstractSparseNeuroVec,numeric-method}
\title{Lookup Values in an IndexLookupVol Object}
\usage{
\S4method{lookup}{IndexLookupVol,numeric}(x, i)

\S4method{lookup}{AbstractSparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{An IndexLookupVol object.}

\item{i}{A numeric vector of indices to look up.}
}
\value{
A vector of lookup values corresponding to the input indices.
}
\description{
This method performs a lookup operation on an IndexLookupVol object.
}
\keyword{internal}
