% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{lookup}
\alias{lookup}
\title{Index Lookup operation}
\usage{
lookup(x, i, ...)
}
\arguments{
\item{x}{the object to query}

\item{i}{the index to lookup}

\item{...}{additional arguments}
}
\description{
Index Lookup operation
}
