% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vol.R, R/sparse_neurovec.R
\name{lookup,IndexLookupVol,numeric-method}
\alias{lookup,IndexLookupVol,numeric-method}
\alias{lookup,AbstractSparseNeuroVec,numeric-method}
\title{lookup}
\usage{
\S4method{lookup}{IndexLookupVol,numeric}(x, i)

\S4method{lookup}{AbstractSparseNeuroVec,numeric}(x, i)
}
\description{
lookup
}
\keyword{internal}
