% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/index_vol.R,
%   R/sparse_neurovec.R
\docType{methods}
\name{lookup}
\alias{lookup}
\alias{lookup,IndexLookupVol,numeric-method}
\alias{lookup,SparseNeuroVec,numeric-method}
\title{Index Lookup operation}
\usage{
lookup(x, i, ...)

\S4method{lookup}{IndexLookupVol,numeric}(x, i)

\S4method{lookup}{SparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{the object to query}

\item{i}{the index to lookup}

\item{...}{additional arguments}
}
\description{
Index Lookup operation

lookup
}
