% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{map}
\alias{map}
\title{Generic function to apply a function to an object}
\usage{
map(x, m, ...)
}
\arguments{
\item{x}{the object that is mapped}

\item{m}{the mapping object}

\item{...}{additional arguments}
}
\description{
Generic function to apply a function to an object
}
