% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\docType{methods}
\name{map_values}
\alias{map_values}
\alias{map_values,NeuroVol,list-method}
\alias{map_values,NeuroVol,matrix-method}
\title{Generic function to map values from one set to another using a user-supplied lookup table}
\usage{
map_values(x, lookup)

\S4method{map_values}{NeuroVol,list}(x, lookup)

\S4method{map_values}{NeuroVol,matrix}(x, lookup)
}
\arguments{
\item{x}{the object to map values from}

\item{lookup}{the lookup table. The first column is the "key" the second column is the "value".}
}
\value{
a new object where the original values have been filled in with the values in the lookup table
}
\description{
Generic function to map values from one set to another using a user-supplied lookup table
}
\examples{
x <- NeuroSpace(c(10,10,10), c(1,1,1))
vol <- NeuroVol(sample(1:10, 10*10*10, replace=TRUE), x)

## lookup table is list
lookup <- lapply(1:10, function(i) i*10)
ovol <- map_values(vol, lookup)

## lookup table is matrix. First column is key, second column is value
names(lookup) <- 1:length(lookup)
lookup.mat <- cbind(as.numeric(names(lookup)), unlist(lookup))
ovol2 <- map_values(vol, lookup.mat)
all.equal(as.vector(ovol2), as.vector(ovol))

}
