% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/mapped_neurovec.R
\docType{class}
\name{MappedNeuroVec-class}
\alias{MappedNeuroVec-class}
\alias{MappedNeuroVec}
\title{MappedNeuroVec}
\usage{
MappedNeuroVec(file_name)
}
\arguments{
\item{file_name}{the name of the 4D image file that containing the memory-mapped data source.}
}
\description{
Four-dimensional brain image, backed by an memory-mapped file.

constructs a MappedNeuroVec object
}
