% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/mapped_neurovec.R
\docType{class}
\name{MappedNeuroVec-class}
\alias{MappedNeuroVec-class}
\alias{MappedNeuroVec}
\title{MappedNeuroVec}
\usage{
MappedNeuroVec(file_name)
}
\arguments{
\item{file_name}{the name of the 4D image file that containing the memory-mapped data source.}
}
\value{
a new instance of type \code{MappedNeuroVec}
}
\description{
A class representing a four-dimensional brain image, backed by a memory-mapped file. MappedNeuroVec objects store their data in a memory-mapped file, providing efficient access to large brain images without loading the entire dataset into memory. This class inherits from the \code{\linkS4class{NeuroVec}} and implements the \code{ArrayLike4D} interface.

constructs a MappedNeuroVec object
}
\section{Slots}{

\describe{
\item{\code{filemap}}{An instance of the \code{mmap} class representing the memory-mapped file containing the brain image data.}

\item{\code{offset}}{An integer representing the offset (in bytes) within the memory-mapped file where the brain image data starts.}
}}

