% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/mapped_neurovec.R
\docType{class}
\name{MappedNeuroVecSource-class}
\alias{MappedNeuroVecSource-class}
\alias{MappedNeuroVecSource}
\title{MappedNeuroVecSource}
\usage{
MappedNeuroVecSource(file_name)
}
\arguments{
\item{file_name}{the name of the image file to be memory mapped}
}
\description{
A class that is used to produce a \code{\linkS4class{MappedNeuroVec}} instance

constructs a MappedNeuroVecSource object
}
