% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{matricized_access}
\alias{matricized_access}
\title{Extract values from a matricized (x,y,z) of a 4D tensor using a space-time coordinate matrix.}
\usage{
matricized_access(x, i, ...)
}
\arguments{
\item{x}{a data source.}

\item{i}{an index matrix specifying the space-time coordinates.}

\item{...}{additional arguments to be passed to methods.}
}
\description{
This function extracts the values of the elements in a 4D tensor, where the first three dimensions
(x,y,z) have been matricized into a single dimension, using a space-time coordinate matrix.
The input \code{i} is an index matrix, where each row specifies a (x,y,z,t) coordinate.
The output is a vector containing the values of the elements in \code{x} at the specified
space-time coordinates.
}
