% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/sparse_neurovec.R
\name{matricized_access}
\alias{matricized_access}
\alias{matricized_access,SparseNeuroVec,matrix-method}
\alias{matricized_access,SparseNeuroVec,integer-method}
\alias{matricized_access,SparseNeuroVec,numeric-method}
\alias{matricized_access,BigNeuroVec,matrix-method}
\alias{matricized_access,BigNeuroVec,integer-method}
\alias{matricized_access,BigNeuroVec,numeric-method}
\title{Extract values from a matricized (x,y,z) of a 4D tensor using a space-time coordinate matrix.}
\usage{
matricized_access(x, i, ...)

\S4method{matricized_access}{SparseNeuroVec,matrix}(x, i)

\S4method{matricized_access}{SparseNeuroVec,integer}(x, i)

\S4method{matricized_access}{SparseNeuroVec,numeric}(x, i)

\S4method{matricized_access}{BigNeuroVec,matrix}(x, i)

\S4method{matricized_access}{BigNeuroVec,integer}(x, i)

\S4method{matricized_access}{BigNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{a data source.}

\item{i}{an index matrix specifying the space-time coordinates.}

\item{...}{additional arguments to be passed to methods.}
}
\value{
A vector containing the values at the specified space-time coordinates.
}
\description{
This function extracts the values of the elements in a 4D tensor, where the first three dimensions
(x,y,z) have been matricized into a single dimension, using a space-time coordinate matrix.
The input \code{i} is an index matrix, where each row specifies a (x,y,z,t) coordinate.
The output is a vector containing the values of the elements in \code{x} at the specified
space-time coordinates.
}
