% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/meta_info.R
\docType{class}
\name{MetaInfo-class}
\alias{MetaInfo-class}
\alias{MetaInfo}
\title{MetaInfo}
\usage{
MetaInfo(
  Dim,
  spacing,
  origin = rep(0, length(spacing)),
  data_type = "FLOAT",
  label = "",
  spatial_axes = OrientationList3D$AXIAL_LPI,
  additional_axes = NullAxis
)
}
\arguments{
\item{Dim}{image dimensions}

\item{spacing}{voxel dimensions}

\item{origin}{coordinate origin}

\item{data_type}{the type of the data (e.g. "FLOAT")}

\item{label}{name(s) of images}

\item{spatial_axes}{image axes for spatial dimensions (x,y,z)}

\item{additional_axes}{axes for dimensions > 3 (e.g. time, color band, direction)}
}
\value{
an instance of class \code{\linkS4class{MetaInfo}}
}
\description{
This class contains meta information for an image data type
}
\section{Slots}{

\describe{
\item{\code{data_type}}{the data type code, e.g. FLOAT}

\item{\code{dims}}{image dimensions}

\item{\code{spatial_axes}}{image axes for spatial dimensions (x,y,z)}

\item{\code{additional_axes}}{axes for dimensions > 3 (e.g. time, color band, direction)}

\item{\code{spacing}}{voxel dimensions}

\item{\code{origin}}{coordinate origin}

\item{\code{label}}{name(s) of images}
}}

