% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{ndim,AxisSet-method}
\alias{ndim,AxisSet-method}
\title{Get number of dimensions in axis set}
\usage{
\S4method{ndim}{AxisSet}(x, ...)
}
\arguments{
\item{x}{An AxisSet object}

\item{...}{Additional arguments (not used)}
}
\value{
Integer number of dimensions
}
\description{
Get number of dimensions in axis set
}
\examples{
\dontrun{
axes <- AxisSet2D(LEFT_RIGHT, ANT_POST)
ndim(axes)  # Returns 2
}
}
