% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{ndim}
\alias{ndim}
\title{Generic function to extract the number of dimensions of an object}
\usage{
ndim(x, ...)
}
\arguments{
\item{x}{n-dimensional object}

\item{...}{additional arguments}
}
\description{
Generic function to extract the number of dimensions of an object
}
\examples{
x = NeuroSpace(c(10,10,10), c(1,1,1))
ndim(x) == 3
x = NeuroSpace(c(10,10,10,3), c(1,1,1,1))
ndim(x) == 4

}

