% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{NeuroHyperVec}
\alias{NeuroHyperVec}
\title{Constructor for NeuroHyperVec class}
\usage{
NeuroHyperVec(data, space, mask)
}
\arguments{
\item{data}{A matrix or three-dimensional array containing the data}

\item{space}{A NeuroSpace object defining the spatial dimensions}

\item{mask}{A mask volume (array, vector, or LogicalNeuroVol)}
}
\value{
A new NeuroHyperVec object
}
\description{
Constructor for NeuroHyperVec class
}
