% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{NeuroHyperVec}
\alias{NeuroHyperVec}
\title{Create an \code{NeuroHyperVec} instance from a variable length list of \code{NeuroVec} objects.}
\usage{
NeuroHyperVec(...)
}
\arguments{
\item{...}{one or more instances of type \code{NeuroVec}}
}
\description{
Create an \code{NeuroHyperVec} instance from a variable length list of \code{NeuroVec} objects.
}
\examples{

v1 <- NeuroVec(array(0,c(5,5,5,4)), space=NeuroSpace(dim=c(5,5,5,4)))
v2 <- NeuroVec(array(0,c(5,5,5,4)), space=NeuroSpace(dim=c(5,5,5,4)))
v3 <- NeuroVec(array(0,c(5,5,5,4)), space=NeuroSpace(dim=c(5,5,5,4)))
hv <- NeuroHyperVec(v1,v2,v3)
}
