% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroObj-class}
\alias{NeuroObj-class}
\title{NeuroObj
Base class for all data objects with a cartesion spatial represenetation}
\description{
NeuroObj
Base class for all data objects with a cartesion spatial represenetation
}
