% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroObj-class}
\alias{NeuroObj-class}
\title{NeuroObj Class}
\description{
Base class for all neuroimaging data objects with a Cartesian spatial representation.
This class provides a foundation for more specific neuroimaging data structures.
}
\section{Slots}{

\describe{
\item{\code{space}}{An object of class \code{\linkS4class{NeuroSpace}} representing the geometry of the image object.}
}}

\seealso{
\code{\link{NeuroSpace-class}}, \code{\link{NeuroSlice-class}}, \code{\link{NeuroVol-class}}
}
