% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroslice.R
\name{NeuroSlice}
\alias{NeuroSlice}
\title{Construct a NeuroSlice object}
\usage{
NeuroSlice(data, space, indices = NULL)
}
\arguments{
\item{data}{data vector or matrix}

\item{space}{an instance of class NeuroSpace}

\item{indices}{linear indices corresponding used if \code{data} is a 1D vector.}
}
\value{
a new instance of type \code{NeuroSlice}
}
\description{
Construct a NeuroSlice object
}
\examples{
bspace <- NeuroSpace(c(64,64), spacing=c(1,1))
dat <- array(rnorm(64*64), c(64,64))
bslice <- NeuroSlice(dat,bspace)
bslice
}
