% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{[[,NeuroVec,numeric,ANY-method}
\alias{[[,NeuroVec,numeric,ANY-method}
\title{[[}
\usage{
\S4method{[[}{NeuroVec,numeric,ANY}(x, i)
}
\arguments{
\item{i}{the volume index}
}
\description{
[[
}
