% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{H5NeuroVec}
\alias{H5NeuroVec}
\title{H5NeuroVec}
\usage{
H5NeuroVec(file_name)
}
\arguments{
\item{file_name}{The name of the 4-dimensional image file in neuroim2 HDF5 format.}
}
\value{
An instance of the \code{\linkS4class{H5NeuroVec}} class.
}
\description{
This function constructs an H5NeuroVec object, which represents a four-dimensional 
brain image stored in the neuroim2 HDF5 format.
}
