% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroVecSeq-class}
\alias{NeuroVecSeq-class}
\title{NeuroVecSeq Class}
\description{
A concatenated sequence of \code{\linkS4class{NeuroVec}} instances.
}
\section{Slots}{

\describe{
\item{\code{vecs}}{The sequences of \code{NeuroVec} instances}

\item{\code{lens}}{The number of volumes in each \code{NeuroVec} sequence}
}}

