% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{[[,NeuroVecSeq,numeric-method}
\alias{[[,NeuroVecSeq,numeric-method}
\title{[[}
\usage{
\S4method{[[}{NeuroVecSeq,numeric}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{the indices}
}
\description{
[[
}
