% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{NeuroVecSeq}
\alias{NeuroVecSeq}
\title{Create an \code{NeuroVecSeq} instance for a variable length list of \code{NeuroVec} objects.}
\usage{
NeuroVecSeq(...)
}
\arguments{
\item{...}{one or more instance of type \code{NeuroVec}}
}
\description{
Create an \code{NeuroVecSeq} instance for a variable length list of \code{NeuroVec} objects.
}
\examples{

v1 <- NeuroVec(array(0,c(5,5,5,2)), space=NeuroSpace(dim=c(5,5,5,2)))
v2 <- NeuroVec(array(0,c(5,5,5,4)), space=NeuroSpace(dim=c(5,5,5,4)))
v3 <- NeuroVec(array(0,c(5,5,5,6)), space=NeuroSpace(dim=c(5,5,5,6)))
vs <- NeuroVecSeq(v1,v2,v3)
blks <- split_blocks(vs, rep(1:3, each=4))
res <- blks \%>\% purrr::map( ~ dim(.))
}
