% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{NeuroVecSeq}
\alias{NeuroVecSeq}
\title{Create an \code{NeuroVecSeq} instance for a variable length list of \code{NeuroVec} objects.}
\usage{
NeuroVecSeq(...)
}
\arguments{
\item{...}{one or more instance of type \code{NeuroVec}}
}
\description{
Create an \code{NeuroVecSeq} instance for a variable length list of \code{NeuroVec} objects.
}
