% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroVecSource-class}
\alias{NeuroVecSource-class}
\title{NeuroVecSource}
\description{
A class that is used to produce a \code{\linkS4class{NeuroVec}} instance
}
\section{Slots}{

\describe{
\item{\code{indices}}{the index vector of the volumes to be loaded}
}}

