% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroVol-class}
\alias{NeuroVol-class}
\title{NeuroVol
Base class for image representing 3D volumetric data.}
\description{
NeuroVol
Base class for image representing 3D volumetric data.
}
