% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{NeuroVol}
\alias{NeuroVol}
\title{Construct a NeuroVol object}
\usage{
NeuroVol(data, space, label = "", indices = NULL)
}
\arguments{
\item{data}{a three-dimensional \code{array}}

\item{space}{an instance of class \code{\linkS4class{NeuroSpace}}}

\item{label}{a \code{character} string to identify volume}

\item{indices}{an 1D vector that gives the linear indices of the associated \code{data} vector}
}
\value{
a \code{\linkS4class{DenseNeuroVol}} instance
}
\description{
Construct a \code{\linkS4class{NeuroVol}} instance, using default (dense) implementation
}
\examples{
bspace <- NeuroSpace(c(64,64,64), spacing=c(1,1,1))
dat <- array(rnorm(64*64*64), c(64,64,64))
bvol <- NeuroVol(dat,bspace, label="test")
show(bvol)
}
