% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{NeuroVolSource}
\alias{NeuroVolSource}
\title{Constructor for NeuroVolSource}
\usage{
NeuroVolSource(input, index = 1)
}
\arguments{
\item{input}{the input file name}

\item{index}{the image subvolume index}
}
\value{
a new instance of type \code{NeuroVolSource}
}
\description{
Constructor for NeuroVolSource
}
