% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neurovol.R
\docType{class}
\name{NeuroVolSource-class}
\alias{NeuroVolSource-class}
\alias{NeuroVolSource}
\title{NeuroVolSource}
\usage{
NeuroVolSource(input, index = 1)
}
\arguments{
\item{input}{the input file name}

\item{index}{the image subvolume index}
}
\description{
A class is used to produce a \code{\linkS4class{NeuroVol}} instance

Constructor for NeuroVolSource
}
\section{Slots}{

\describe{
\item{\code{index}}{the index of the volume to be read -- must be of length 1.}
}}

