% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustervol.R
\docType{methods}
\name{num_clusters}
\alias{num_clusters}
\alias{num_clusters,ClusteredNeuroVol-method}
\title{num_clusters}
\usage{
num_clusters(x)

\S4method{num_clusters}{ClusteredNeuroVol}(x)
}
\arguments{
\item{x}{the object to extract number of clusters}
}
\description{
num_clusters

get number of clusters in a ClusteredNeuroVol
}
