% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{origin}
\alias{origin}
\alias{origin,NeuroSpace-method}
\title{extract image origin}
\usage{
origin(x)

\S4method{origin}{NeuroSpace}(x)
}
\arguments{
\item{x}{an object with an origin}
}
\description{
extract image origin
}
\examples{
bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
origin(bspace)

}
