% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{origin}
\alias{origin}
\alias{origin,NeuroSpace-method}
\alias{origin,NeuroVol-method}
\alias{origin,NeuroVec-method}
\title{Extract Image Origin}
\usage{
origin(x)

\S4method{origin}{NeuroSpace}(x)

\S4method{origin}{NeuroVol}(x)

\S4method{origin}{NeuroVec}(x)
}
\arguments{
\item{x}{an object with an origin}
}
\value{
the origin of the image
}
\description{
Extract Image Origin
}
\examples{
bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
stopifnot(origin(bspace) == c(0,0,0))

}
