% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{patch_set}
\alias{patch_set}
\alias{patch_set,NeuroVol,numeric,missing-method}
\alias{patch_set,NeuroVol,numeric,LogicalNeuroVol-method}
\title{Extract set of patches}
\usage{
patch_set(x, dims, mask, ...)

\S4method{patch_set}{NeuroVol,numeric,missing}(x, dims, mask, ...)

\S4method{patch_set}{NeuroVol,numeric,LogicalNeuroVol}(x, dims, mask, ...)
}
\arguments{
\item{x}{the object to extract patches from}

\item{dims}{a vector indicating the dimensions of the patches}

\item{mask}{mask indicating the valid patch area}

\item{...}{additional args}
}
\description{
generate a set of coordinate "patches" of fixed size from an image object.
}
