% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{patch_set}
\alias{patch_set}
\title{Generate a set of coordinate "patches" of fixed size from an image object.}
\usage{
patch_set(x, dims, mask, ...)
}
\arguments{
\item{x}{the object to extract patches from}

\item{dims}{a vector indicating the dimensions of the patches}

\item{mask}{mask indicating the valid patch area}

\item{...}{additional args}
}
\value{
A list of coordinate patches, each representing a fixed-size region of the input object.
}
\description{
Generate a set of coordinate "patches" of fixed size from an image object.
}
