% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/axis.R
\name{perm_mat}
\alias{perm_mat}
\alias{perm_mat,AxisSet2D-method}
\alias{perm_mat,AxisSet3D-method}
\alias{perm_mat,NeuroSpace-method}
\title{Extract permutation matrix}
\usage{
perm_mat(x, ...)

\S4method{perm_mat}{AxisSet2D}(x, ...)

\S4method{perm_mat}{AxisSet3D}(x, ...)

\S4method{perm_mat}{NeuroSpace}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
Extract permutation matrix

perm_mat

perm_mat

perm_mat
}
