% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{perm_mat,NeuroSpace-method}
\alias{perm_mat,NeuroSpace-method}
\title{Get permutation matrix from axis set}
\usage{
\S4method{perm_mat}{NeuroSpace}(x, ...)
}
\arguments{
\item{x}{A NeuroSpace object}

\item{...}{Additional arguments (not used)}
}
\description{
Get permutation matrix from axis set
}
\examples{
ns <- NeuroSpace(dim=c(10,10,10))
perm_mat(ns)

}
