% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroslice.R
\docType{methods}
\name{plot,NeuroSlice-method}
\alias{plot,NeuroSlice-method}
\title{plot}
\usage{
\S4method{plot}{NeuroSlice}(x, cmap = gray(seq(0, 1, length.out = 255)),
  irange = range(x))
}
\arguments{
\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{irange}{the intensity range indicating the low and high values of the color scale.}
}
\description{
plot
}
\examples{

dat <- matrix(rnorm(100*100), 100, 100)
slice <- NeuroSlice(dat, NeuroSpace(c(100,100)))
plot(slice)
}
