% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroslice.R, R/neurovol.R
\name{plot,NeuroSlice-method}
\alias{plot,NeuroSlice-method}
\alias{plot,NeuroVol-method}
\title{plot an image}
\usage{
\S4method{plot}{NeuroSlice}(x, cmap = gray(seq(0, 1, length.out = 255)), irange = range(x))

\S4method{plot}{NeuroVol}(
  x,
  cmap = gray(seq(0, 1, length.out = 255)),
  zlevels = unique(round(seq(1, dim(x)[3], length.out = 6))),
  irange = range(x),
  thresh = c(0, 0),
  alpha = 1,
  bgvol = NULL,
  bgcmap = gray(seq(0, 1, length.out = 255))
)
}
\arguments{
\item{x}{the object to display}

\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{irange}{the intensity range indicating the low and high values of the color scale.}

\item{zlevels}{the series of slice indices to display.}

\item{thresh}{a 2-element vector indicating the lower and upper transparency thresholds.}

\item{bgvol}{a background volume that serves as an image underlay (currently ignored).}
}
\description{
plot an image
}
\examples{

dat <- matrix(rnorm(100*100), 100, 100)
slice <- NeuroSlice(dat, NeuroSpace(c(100,100)))
#plot(slice)

dat <- matrix(rnorm(100*100), 100, 100)
slice <- NeuroSlice(dat, NeuroSpace(c(100,100)))
#plot(slice)
}
