% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{print_,NamedAxis-method}
\alias{print_,NamedAxis-method}
\title{print a \code{NamedAxis}}
\usage{
\S4method{print_}{NamedAxis}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
print a \code{NamedAxis}
}
