% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{random_searchlight}
\alias{random_searchlight}
\title{Create an spherical random searchlight iterator}
\usage{
random_searchlight(mask, radius)
}
\arguments{
\item{mask}{an volumetric image mask of type \code{\linkS4class{NeuroVol}} 
containing valid searchlight voxel set.}

\item{radius}{width in mm of spherical searchlight}
}
\description{
Create an spherical random searchlight iterator
}
