% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{random_searchlight}
\alias{random_searchlight}
\title{random_searchlight}
\usage{
random_searchlight(mask, radius)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}
}
\description{
Create an spherical random searchlight iterator
}
