% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_elements,BinaryReader,numeric-method}
\alias{read_elements,BinaryReader,numeric-method}
\title{Read Elements from Binary Reader}
\usage{
\S4method{read_elements}{BinaryReader,numeric}(x, num_elements)
}
\arguments{
\item{x}{Object of class \linkS4class{BinaryReader}}

\item{num_elements}{Integer specifying number of elements to read}
}
\value{
Numeric vector of read elements
}
\description{
Read a specified number of elements from a \linkS4class{BinaryReader} object.
}
\examples{
\dontrun{
reader <- BinaryReader("data.bin", byte_offset = 0L,
                      data_type = "double", bytes_per_element = 8L)
# Read 100 elements
data <- read_elements(reader, 100)
close(reader)
}
}
