% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/binary_io.R
\name{read_elements}
\alias{read_elements}
\alias{read_elements,BinaryReader,numeric-method}
\title{Read a sequence of elements from an input source}
\usage{
read_elements(x, num_elements)

\S4method{read_elements}{BinaryReader,numeric}(x, num_elements)
}
\arguments{
\item{x}{A BinaryReader object.}

\item{num_elements}{A numeric value specifying the number of elements to read.}
}
\value{
the elements as a vector

A vector of read elements.
}
\description{
This method reads a specified number of elements from a BinaryReader object.
}
