% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{read_elements}
\alias{read_elements}
\title{Read a sequence of elements from an input source}
\usage{
read_elements(x, num_elements)
}
\arguments{
\item{x}{the input channel}

\item{num_elements}{the number of elements to read}
}
\value{
the elements as a vector
}
\description{
Read a sequence of elements from an input source
}
