% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/binary_io.R
\docType{methods}
\name{read_elements}
\alias{read_elements}
\alias{read_elements,BinaryReader,numeric-method}
\title{Generic function to read a sequence of elements from an input source}
\usage{
read_elements(x, num_elements)

\S4method{read_elements}{BinaryReader,numeric}(x, num_elements)
}
\arguments{
\item{x}{the input channel}

\item{num_elements}{the number of elements to read}
}
\value{
the elements as a vector
}
\description{
Generic function to read a sequence of elements from an input source

read_elements
}
