% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_data}
\alias{read_mapped_data}
\title{Read Mapped Data from 4D Image}
\usage{
read_mapped_data(meta, idx)
}
\arguments{
\item{meta}{An object of class \linkS4class{ImageMetadata} containing file metadata}

\item{idx}{Integer vector of indices specifying elements to read}
}
\value{
A numeric vector of values
}
\description{
Read data from a memory-mapped 4D image file.
}
\keyword{internal}
