% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_series}
\alias{read_mapped_series}
\title{Read Mapped Series from 4D Image}
\usage{
read_mapped_series(meta, idx)
}
\arguments{
\item{meta}{An object of class \linkS4class{ImageMetadata} containing file metadata}

\item{idx}{Integer vector of indices specifying elements to read}
}
\value{
A numeric matrix of values with dimensions [time, voxels]
}
\description{
Read a time series of data from a memory-mapped 4D image file.
}
\keyword{internal}
