% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{read_mapped_vols}
\alias{read_mapped_vols}
\title{Read Mapped Volumes from 4D Image}
\usage{
read_mapped_vols(meta, idx)
}
\arguments{
\item{meta}{An object of class \linkS4class{ImageMetadata} containing file metadata}

\item{idx}{Integer vector of indices specifying volumes to read}
}
\value{
A numeric matrix with dimensions [time, voxels]
}
\description{
Read volume data from a memory-mapped 4D image file.
}
\keyword{internal}
