% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/file_format.R
\name{read_meta_info}
\alias{read_meta_info}
\alias{read_meta_info,NIFTIFormat-method}
\alias{read_meta_info,AFNIFormat-method}
\title{Generic function to read image meta info given a file and a \code{\linkS4class{FileFormat}} instance.}
\usage{
read_meta_info(x, file_name)

\S4method{read_meta_info}{NIFTIFormat}(x, file_name)

\S4method{read_meta_info}{AFNIFormat}(x, file_name)
}
\arguments{
\item{x}{descriptor instance}

\item{file_name}{file name contianing meta information}
}
\description{
Generic function to read image meta info given a file and a \code{\linkS4class{FileFormat}} instance.
}
