% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{read_vec}
\alias{read_vec}
\title{read_vec}
\usage{
read_vec(
  file_name,
  indices = NULL,
  mask = NULL,
  mode = c("normal", "mmap", "bigvec", "filebacked")
)
}
\arguments{
\item{file_name}{The name(s) of the file(s) to load. If more than one file_name is specified, 
the files are loaded and concatenated.}

\item{indices}{The indices of the sub-volumes to load (e.g. if the file is 4-dimensional).}

\item{mask}{A mask defining the spatial elements to load.}

\item{mode}{The IO mode which is one of "normal", "mmap", "bigvec", or "filebacked".}
}
\value{
An \code{\linkS4class{NeuroVec}} object.
}
\description{
This function loads an image volume from a file.
}
\note{
Memory-mapping a gzipped file is not currently allowed.
}
