% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\name{read_vec}
\alias{read_vec}
\title{read_vec}
\usage{
read_vec(file_name, indices = NULL, mask = NULL)
}
\arguments{
\item{file_name}{the name of the file to load}

\item{indices}{the indices of the sub-volumes to load (e.g. if the file is 4-dimensional)}

\item{mask}{a mask defining the spatial elements to load}
}
\value{
an \code{\linkS4class{NeuroVec}} object
}
\description{
load an image volume from a file
}
