% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{resample}
\alias{resample}
\title{Resample an Image to Match the Space of Another Image}
\usage{
resample(source, target, ...)
}
\arguments{
\item{source}{An object representing the source image to be resampled. This could be a 3D or 4D image object, depending on the use case.}

\item{target}{An object representing the target image, whose spatial properties will be used as the reference for resampling the source image.}

\item{...}{Additional arguments passed to the resampling function, such as interpolation method, boundary handling, or other resampling options.}
}
\value{
An object representing the resampled source image, with the same spatial properties as the target image.
}
\description{
This function resamples a source image to match the spatial properties (dimensions, resolution, and orientation) of a target image.
}
