% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ROIVecWindow-class}
\alias{ROIVecWindow-class}
\title{ROIVecWindow}
\description{
A class that represents a vector-valued volumetric region of interest
}
\section{Slots}{

\describe{
\item{\code{parent_index}}{the 1D index of the center voxel in the parent space.}

\item{\code{center_index}}{the location in the coordinate matrix of the center voxel in the window}
}}

