% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ROIVol-class}
\alias{ROIVol-class}
\title{ROIVol}
\description{
A class that represents a volumetric region of interest
}
\section{Slots}{

\describe{
\item{\code{data}}{the \code{numeric} data stored in ROI}

\item{\code{coords}}{the voxel coordinates of the ROI}
}}

