% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{bootstrap_searchlight}
\alias{bootstrap_searchlight}
\alias{searchlight_coords}
\alias{searchlight}
\alias{clustered_searchlight}
\title{bootstrap_searchlight}
\usage{
bootstrap_searchlight(mask, radius = 8, iter = 100)

searchlight_coords(mask, radius, nonzero = FALSE)

searchlight(mask, radius, eager = FALSE, nonzero = FALSE)

clustered_searchlight(mask, cvol = NULL, csize = NULL)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}

\item{iter}{the total number of searchlights to sample (default is 100).}

\item{nonzero}{only include nonzero coordinates}

\item{eager}{if TRUE, then all searchlight coordinates set are generated up front. This is faster but requires more memory to store all coordinates.}

\item{cvol}{a \code{ClusteredNeuroVol} instance}

\item{csize}{the number of clusters (ignored if \code{cvol} is provided)}
}
\value{
a list ofmatrices containing of integer-valued voxel coordinates

a list of \code{ROIVolWindow} objects

an \code{iter} class
}
\description{
Create a spherical searchlight iterator that samples regions from within a mask.

Create an exhaustive searchlight iterator that only returns voxel coordinates

Generates an \code{ROIVol} around each non-zero voxel in a 3D image mask

Create a clustered searchlight iterator
}
\details{
searchlight centers are sampled without replacement, but the same surround voxel can belong to multiple searchlight samples.
}
