% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{searchlight}
\alias{searchlight}
\title{searchlight}
\usage{
searchlight(mask, radius, eager = FALSE, nonzero = FALSE)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}

\item{nonzero}{only include nonzero values}
}
\value{
an \code{ROIVolWindow} class
}
\description{
Create an exhaustive searchlight iterator
}
