% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{searchlight_coords}
\alias{searchlight_coords}
\alias{searchlight}
\title{Create an exhaustive searchlight iterator that only returns voxel coordinates}
\usage{
searchlight_coords(mask, radius, nonzero = FALSE, cores = 0)

searchlight(mask, radius, eager = FALSE, nonzero = FALSE, cores = 0)
}
\arguments{
\item{mask}{A \code{\linkS4class{NeuroVol}} object representing the brain mask, containing valid central voxels for the roving searchlight.}

\item{radius}{A numeric value specifying the radius (in mm) of the spherical searchlight.}

\item{nonzero}{A logical value indicating whether to include only coordinates with nonzero values in the supplied mask (default is FALSE).}

\item{cores}{An integer specifying the number of cores to use for parallel computation (default is 0, which uses a single core).}

\item{eager}{A logical value specifying whether to eagerly compute the searchlight ROIs (default is FALSE, which uses lazy evaluation).}
}
\value{
A \code{\linkS4class{deferred_list}} object containing matrices of integer-valued voxel coordinates, each representing a searchlight region.

A \code{\linkS4class{deferred_list}} object containing matrices of integer-valued voxel coordinates, each representing a searchlight region.
}
\description{
This function generates an exhaustive searchlight iterator that returns voxel coordinates for each searchlight
sphere within the provided mask. The searchlight iterator visits every non-zero voxel in the mask as a potential center voxel.

This function generates an exhaustive searchlight iterator that returns voxel coordinates for each searchlight
sphere within the provided mask. The searchlight iterator visits every non-zero voxel in the mask as a potential center voxel.
}
\examples{
# Load an example brain mask
mask <- read_vol(system.file("extdata", "global_mask.nii", package="neuroim2"))

# Generate an exhaustive searchlight iterator with a radius of 6 mm
searchlights <- searchlight_coords(mask, radius = 6)

# Load an example brain mask
mask <- read_vol(system.file("extdata", "global_mask.nii", package="neuroim2"))

# Generate an exhaustive searchlight iterator with a radius of 6 mm
searchlights <- searchlight(mask, radius = 6, eager = TRUE)

}
