% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{searchlight_coords}
\alias{searchlight_coords}
\title{searchlight}
\usage{
searchlight_coords(mask, radius, nonzero = FALSE)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}

\item{nonzero}{only include nonzero coordinates}
}
\value{
a list of three-dimensional matrices containingof integer-valued voxel coordinates
}
\description{
Create an exhaustive searchlight iterator that only returns voxel coordinates
}
