% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R
\docType{methods}
\name{series_roi}
\alias{series_roi}
\alias{series}
\alias{series,NeuroVec,matrix-method}
\alias{series_roi,NeuroVec,matrix-method}
\alias{series,NeuroVec,ROIVolume-method}
\alias{series_roi,NeuroVec,ROIVolume-method}
\alias{series,NeuroVec,LogicalNeuroVol-method}
\alias{series_roi,NeuroVec,LogicalNeuroVol-method}
\alias{series,NeuroVec,numeric-method}
\alias{series_roi,NeuroVec,numeric-method}
\title{Extract one or more series from object and return as ROI object}
\usage{
series_roi(x, i, ...)

series(x, i, ...)

\S4method{series}{NeuroVec,matrix}(x, i)

\S4method{series_roi}{NeuroVec,matrix}(x, i)

\S4method{series}{NeuroVec,ROIVolume}(x, i)

\S4method{series_roi}{NeuroVec,ROIVolume}(x, i)

\S4method{series}{NeuroVec,LogicalNeuroVol}(x, i)

\S4method{series_roi}{NeuroVec,LogicalNeuroVol}(x, i)

\S4method{series}{NeuroVec,numeric}(x, i, j, k)

\S4method{series_roi}{NeuroVec,numeric}(x, i, j, k)
}
\arguments{
\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}

\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}
}
\description{
Extract one or more series from object and return as ROI object

Extract one or more series from object
}
