% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/cached_neurovec.R,
%   R/neurovec.R, R/sparse_neurovec.R
\docType{methods}
\name{series_roi}
\alias{series_roi}
\alias{series}
\alias{series,CachedSparseNeuroVec,matrix-method}
\alias{series,NeuroVec,matrix-method}
\alias{series_roi,NeuroVec,matrix-method}
\alias{series,NeuroVec,ROICoords-method}
\alias{series_roi,NeuroVec,ROICoords-method}
\alias{series,NeuroVec,LogicalNeuroVol-method}
\alias{series_roi,NeuroVec,LogicalNeuroVol-method}
\alias{series,NeuroVec,integer-method}
\alias{series,NeuroVec,numeric-method}
\alias{series_roi,NeuroVec,numeric-method}
\alias{series,NeuroVecSeq,integer-method}
\alias{series,NeuroVecSeq,numeric-method}
\alias{series,NeuroVecSeq,matrix-method}
\alias{series_roi,NeuroVecSeq,matrix-method}
\alias{series,SparseNeuroVec,matrix-method}
\alias{series,SparseNeuroVec,numeric-method}
\title{Extract one or more series from object and return as ROI object}
\usage{
series_roi(x, i, ...)

series(x, i, ...)

\S4method{series}{CachedSparseNeuroVec,matrix}(x, i)

\S4method{series}{NeuroVec,matrix}(x, i)

\S4method{series_roi}{NeuroVec,matrix}(x, i)

\S4method{series}{NeuroVec,ROICoords}(x, i)

\S4method{series_roi}{NeuroVec,ROICoords}(x, i)

\S4method{series}{NeuroVec,LogicalNeuroVol}(x, i)

\S4method{series_roi}{NeuroVec,LogicalNeuroVol}(x, i)

\S4method{series}{NeuroVec,integer}(x, i, j, k)

\S4method{series}{NeuroVec,numeric}(x, i, j, k)

\S4method{series_roi}{NeuroVec,numeric}(x, i, j, k)

\S4method{series}{NeuroVecSeq,integer}(x, i, j, k)

\S4method{series}{NeuroVecSeq,numeric}(x, i, j, k)

\S4method{series}{NeuroVecSeq,matrix}(x, i)

\S4method{series_roi}{NeuroVecSeq,matrix}(x, i)

\S4method{series}{SparseNeuroVec,matrix}(x, i)

\S4method{series}{SparseNeuroVec,numeric}(x, i, j, k)
}
\arguments{
\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}

\item{j}{index for 2nd dimension}

\item{k}{index for 3rd dimension}

\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}
}
\description{
Extract one or more series from object and return as ROI object

Extract one or more series from object
}
