% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{series_roi}
\alias{series}
\alias{series_roi}
\title{Extract one or more series from object and return as ROI object}
\usage{
series_roi(x, i, ...)

series(x, i, ...)
}
\arguments{
\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}

\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}
}
\description{
Extract one or more series from object and return as ROI object

Extract one or more series from object
}

