% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{series_reader}
\alias{series_reader}
\title{Create a series reader function for a 4D image file}
\usage{
series_reader(file_name)
}
\arguments{
\item{file_name}{The name of the 4D image file to read from.}
}
\value{
A function that can read series of data from the specified file.
}
\description{
This function creates a closure that can read series of data from a 4D image file.
}
\keyword{internal}
