% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{show,AxisSet3D-method}
\alias{show,AxisSet3D-method}
\title{Show method for AxisSet3D objects}
\usage{
\S4method{show}{AxisSet3D}(object)
}
\arguments{
\item{object}{An AxisSet3D object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for AxisSet3D objects
}
