% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustervol.R
\docType{methods}
\name{show,ClusteredNeuroVol-method}
\alias{show,ClusteredNeuroVol-method}
\title{show a \code{ClusteredNeuroVol}}
\usage{
\S4method{show}{ClusteredNeuroVol}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{ClusteredNeuroVol}
}
