% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R, R/clustered_neurovec.R, R/clustervol.R,
%   R/index_vol.R, R/mapped_neurovec.R, R/meta_info.R, R/neurohypervec.R,
%   R/neuroslice.R, R/neurospace.R, R/neurovec.R, R/neurovecseq.R, R/neurovol.R,
%   R/roi.R, R/sparse_neurovec.R
\name{show,NamedAxis-method}
\alias{show,NamedAxis-method}
\alias{show,AxisSet1D-method}
\alias{show,AxisSet2D-method}
\alias{show,AxisSet3D-method}
\alias{show,AxisSet4D-method}
\alias{show,ClusteredNeuroVec-method}
\alias{show,ClusteredNeuroVol-method}
\alias{show,IndexLookupVol-method}
\alias{show,MappedNeuroVec-method}
\alias{show,FileMetaInfo-method}
\alias{show,NeuroHyperVec-method}
\alias{show,NeuroSlice-method}
\alias{show,NeuroSpace-method}
\alias{show,NeuroVecSource-method}
\alias{show,NeuroVec-method}
\alias{show,DenseNeuroVec-method}
\alias{show,NeuroVecSeq-method}
\alias{show,NeuroVol-method}
\alias{show,SparseNeuroVol-method}
\alias{show,Kernel-method}
\alias{show,ROIVol-method}
\alias{show,ROICoords-method}
\alias{show,ROIVec-method}
\alias{show,SparseNeuroVec-method}
\title{Show method for NamedAxis objects}
\usage{
\S4method{show}{NamedAxis}(object)

\S4method{show}{AxisSet1D}(object)

\S4method{show}{AxisSet2D}(object)

\S4method{show}{AxisSet3D}(object)

\S4method{show}{AxisSet4D}(object)

\S4method{show}{ClusteredNeuroVec}(object)

\S4method{show}{ClusteredNeuroVol}(object)

\S4method{show}{IndexLookupVol}(object)

\S4method{show}{MappedNeuroVec}(object)

\S4method{show}{FileMetaInfo}(object)

\S4method{show}{NeuroHyperVec}(object)

\S4method{show}{NeuroSlice}(object)

\S4method{show}{NeuroSpace}(object)

\S4method{show}{NeuroVecSource}(object)

\S4method{show}{NeuroVec}(object)

\S4method{show}{DenseNeuroVec}(object)

\S4method{show}{NeuroVecSeq}(object)

\S4method{show}{NeuroVecSeq}(object)

\S4method{show}{NeuroVol}(object)

\S4method{show}{SparseNeuroVol}(object)

\S4method{show}{Kernel}(object)

\S4method{show}{ROIVol}(object)

\S4method{show}{ROICoords}(object)

\S4method{show}{ROIVec}(object)

\S4method{show}{SparseNeuroVec}(object)
}
\arguments{
\item{object}{A NamedAxis object}
}
\value{
Invisibly returns \code{NULL}, called for its side effect of displaying the object.
}
\description{
Show method for NamedAxis objects

Show method for ClusteredNeuroVec
}
