% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurohypervec.R
\name{show,NeuroHyperVec-method}
\alias{show,NeuroHyperVec-method}
\title{Show NeuroHyperVec Object}
\usage{
\S4method{show}{NeuroHyperVec}(object)
}
\arguments{
\item{object}{A NeuroHyperVec object to display.}
}
\description{
Displays a formatted summary of a NeuroHyperVec object, including dimensions,
sparsity information, and memory usage.
}
\details{
Display NeuroHyperVec Object
}
