% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroslice.R
\name{show,NeuroSlice-method}
\alias{show,NeuroSlice-method}
\title{show a \code{NeuroSlice}}
\usage{
\S4method{show}{NeuroSlice}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NeuroSlice}
}
