% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurospace.R
\name{show,NeuroSpace-method}
\alias{show,NeuroSpace-method}
\title{show a \code{NeuroSpace}}
\usage{
\S4method{show}{NeuroSpace}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NeuroSpace}
}
