% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{show,NeuroVec-method}
\alias{show,NeuroVec-method}
\title{show a \code{NeuroVec}}
\usage{
\S4method{show}{NeuroVec}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NeuroVec}
}
