% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovec.R
\docType{methods}
\name{show,NeuroVecSource-method}
\alias{show,NeuroVecSource-method}
\title{show a \code{NeuroVecSource}}
\usage{
\S4method{show}{NeuroVecSource}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NeuroVecSource}
}
