% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurovol.R
\name{show,NeuroVol-method}
\alias{show,NeuroVol-method}
\title{Display NeuroVol Object}
\usage{
\S4method{show}{NeuroVol}(object)
}
\arguments{
\item{object}{A \code{NeuroVol} object.}
}
\description{
Displays a formatted summary of a \code{NeuroVol} object.
}
