% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\docType{methods}
\name{show,ROIVec-method}
\alias{show,ROIVec-method}
\title{show an \code{\linkS4class{ROIVec}}}
\usage{
\S4method{show}{ROIVec}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{\linkS4class{ROIVec}}
}
