% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{slice}
\alias{slice}
\alias{slice,NeuroVol,numeric,numeric,missing-method}
\alias{slice,NeuroVol,numeric,NeuroSpace,AxisSet3D-method}
\title{Extract image slice}
\usage{
slice(x, zlevel, along, orientation, ...)

\S4method{slice}{NeuroVol,numeric,numeric,missing}(x, zlevel, along, orientation)

\S4method{slice}{NeuroVol,numeric,NeuroSpace,AxisSet3D}(x, zlevel, along, orientation)
}
\arguments{
\item{x}{the object}

\item{zlevel}{coordinate (in voxel units) along the sliced axis}

\item{along}{the axis along which to slice}

\item{orientation}{the target orientation of the 2D slice}

\item{...}{additional arguments}
}
\description{
Extract a 2D slice from an image volume
}
