% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\docType{methods}
\name{slices}
\alias{slices}
\alias{slices,NeuroVol-method}
\title{Generic function to extract an ordered series of 2D slices}
\usage{
slices(x, ...)

\S4method{slices}{NeuroVol}(x)
}
\arguments{
\item{x}{the object that supplies the slices}

\item{...}{additional arguments}
}
\description{
Generic function to extract an ordered series of 2D slices
}
