% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovol.R
\name{slices}
\alias{slices}
\alias{slices,NeuroVol-method}
\title{Extract an ordered series of 2D slices from a 3D or 4D object}
\usage{
slices(x, ...)

\S4method{slices}{NeuroVol}(x)
}
\arguments{
\item{x}{The 3D or 4D object to extract slices from}

\item{...}{Additional arguments to be passed to the underlying methods}
}
\value{
A list of 2D matrices, each containing a slice from the input object
}
\description{
This function extracts an ordered series of 2D slices from a 3D or 4D object. The returned slices are in the order they appear in the original object.
}
