% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/index_vol.R, R/neurospace.R
\docType{methods}
\name{space}
\alias{space}
\alias{space,IndexLookupVol-method}
\alias{space,NeuroSpace-method}
\title{Generic function to extract geometric properties of an image.}
\usage{
space(x, ...)

\S4method{space}{IndexLookupVol}(x)

\S4method{space}{NeuroSpace}(x)
}
\arguments{
\item{x}{the object to query, e.g. an instance of \code{\linkS4class{NeuroVol}} or \code{\linkS4class{NeuroVec}}}

\item{...}{additional arguments}
}
\value{
an object representing the geometric space of the image of type \code{\linkS4class{NeuroSpace}}
}
\description{
Generic function to extract geometric properties of an image.
}
\examples{
x = NeuroSpace(c(10,10,10), c(1,1,1))
vol <- NeuroVol(rnorm(10*10*10), x)
identical(x,space(vol))

}
