% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{spacing}
\alias{spacing}
\title{Generic function to extract the voxel dimensions of an image}
\usage{
spacing(x)
}
\arguments{
\item{x}{the object}
}
\value{
a numeric vector
}
\description{
Generic function to extract the voxel dimensions of an image
}
\examples{
bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
all.equal(spacing(bspace), c(2,2,2))
}

