% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SparseNeuroVec-class}
\alias{SparseNeuroVec-class}
\title{SparseNeuroVec}
\description{
a sparse four-dimensional brain image, backed by a \code{matrix}, where each column represents
a vector spanning the fourth dimension (e.g. time)
}
\section{Slots}{

\describe{
\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{data}}{the matrix of series, where rows span across voxel space and columns span the fourth dimensions}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVolume}} is used to map between spatial and index/row coordinates}
}}

