% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_neurovec.R, R/sparse_neurovec.R
\name{[[,LatentNeuroVec,numeric-method}
\alias{[[,LatentNeuroVec,numeric-method}
\alias{[[,SparseNeuroVec,numeric-method}
\title{[[}
\usage{
\S4method{[[}{LatentNeuroVec,numeric}(x, i)

\S4method{[[}{SparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{the volume index}
}
\description{
[[

[[
}
