% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_neurovec.R
\name{[[,SparseNeuroVec,numeric-method}
\alias{[[,SparseNeuroVec,numeric-method}
\title{[[}
\usage{
\S4method{[[}{SparseNeuroVec,numeric}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{the volume index}
}
\value{
a SparseNeuroVol object
}
\description{
[[
}
