% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_neurovec.R
\docType{methods}
\name{[[,SparseNeuroVec,numeric,ANY-method}
\alias{[[,SparseNeuroVec,numeric,ANY-method}
\title{[[}
\usage{
\S4method{[[}{SparseNeuroVec,numeric,ANY}(x, i)
}
\arguments{
\item{i}{the volume index}
}
\description{
[[
}
