% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/sparse_neurovec.R
\docType{class}
\name{SparseNeuroVecSource-class}
\alias{SparseNeuroVecSource-class}
\alias{SparseNeuroVecSource}
\title{SparseNeuroVecSource}
\usage{
SparseNeuroVecSource(meta_info, indices, mask)
}
\arguments{
\item{meta_info}{an object of class \code{\linkS4class{MetaInfo}}}

\item{indices}{a vector of 1D indices}

\item{mask}{a 3D \code{array} of type \code{logical}}
}
\description{
A class that is used to produce a \code{\linkS4class{SparseNeuroVec}} instance

constructs a SparseNeuroVecSource object
}
\section{Slots}{

\describe{
\item{\code{mask}}{the subset of voxels that will be stored in memory}
}}

