% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SparseNeuroVol-class}
\alias{SparseNeuroVol-class}
\title{SparseNeuroVol}
\description{
Three-dimensional brain image, backed by a \code{sparseVector} for \code{Matrix} package
}
\section{Slots}{

\describe{
\item{\code{data}}{a \code{sparseVector} instance}
}}

