% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R
\name{split_blocks}
\alias{split_blocks}
\alias{split_blocks,NeuroVec,integer-method}
\alias{split_blocks,NeuroVec,factor-method}
\title{cut a vector-valued object into a list of sub-blocks}
\usage{
split_blocks(x, indices, ...)

\S4method{split_blocks}{NeuroVec,integer}(x, indices, ...)

\S4method{split_blocks}{NeuroVec,factor}(x, indices, ...)
}
\arguments{
\item{x}{the object to split}

\item{indices}{a vector of indices for the blocks. Must match the length of the inut vector.}

\item{...}{additional arguments}
}
\description{
cut a vector-valued object into a list of sub-blocks
}
