% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/clustervol.R, R/neurovec.R
\docType{methods}
\name{split_clusters}
\alias{split_clusters}
\alias{split_clusters,NeuroVol,ClusteredNeuroVol-method}
\alias{split_clusters,NeuroVol,integer-method}
\alias{split_clusters,NeuroVol,numeric-method}
\alias{split_clusters,NeuroVec,integer-method}
\alias{split_clusters,NeuroVec,numeric-method}
\alias{split_clusters,NeuroVec,ClusteredNeuroVol-method}
\title{Generic function to cut an object into a list of spatial or spatiotemporal clusters}
\usage{
split_clusters(x, clusters, ...)

\S4method{split_clusters}{NeuroVol,ClusteredNeuroVol}(x, clusters)

\S4method{split_clusters}{NeuroVol,integer}(x, clusters)

\S4method{split_clusters}{NeuroVol,numeric}(x, clusters)

\S4method{split_clusters}{NeuroVec,integer}(x, clusters, ...)

\S4method{split_clusters}{NeuroVec,numeric}(x, clusters, ...)

\S4method{split_clusters}{NeuroVec,ClusteredNeuroVol}(x, clusters, ...)
}
\arguments{
\item{x}{the object to split}

\item{clusters}{a vector of cluster indices to split by}

\item{...}{additional arguments}
}
\description{
Generic function to cut an object into a list of spatial or spatiotemporal clusters

split_clusters
}
\examples{

## split 'NeuroVol' with a 'ClusteredNeuroVol'
vol <- NeuroVol(array(runif(10*10*10),c(10,10,10)), NeuroSpace(c(10,10,10)))
mask <- as.logical(vol > .5)
mask.idx <- which(mask != 0)
grid <- index_to_coord(mask, mask.idx)
vox <- index_to_grid(mask, mask.idx)

library(purrr)
## partition coordinates into 50 clusters using 'kmeans'
kres <- kmeans(grid, centers=50, iter.max=500)
kvol <- ClusteredNeuroVol(mask, kres$cluster)
klis <- split_clusters(mask, kvol)
ret1 <- vol \%>\% split_clusters(kvol) \%>\% purrr::map_dbl(~ mean(values(.)))

## split NeuroVol with 'integer' vector of clusters.
indices <- numeric(prod(dim(mask)[1:3]))

## locations with a cluster value of 0 are ignored
indices[mask.idx] <- kres$cluster

ret2 <- vol \%>\% split_clusters(as.integer(indices)) \%>\% purrr::map_dbl(~ mean(values(.)))
all(ret1 == ret1)

}
