% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/file_format.R
\name{strip_extension}
\alias{strip_extension}
\alias{strip_extension,FileFormat,character-method}
\title{Generic function to strip extension from file name, given a \code{\linkS4class{FileFormat}} instance.}
\usage{
strip_extension(x, file_name)

\S4method{strip_extension}{FileFormat,character}(x, file_name)
}
\arguments{
\item{x}{descriptor instance}

\item{file_name}{file name to be stripped of its extension}
}
\value{
file_name without extension
}
\description{
Generic function to strip extension from file name, given a \code{\linkS4class{FileFormat}} instance.
}
