% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{trans}
\alias{trans}
\title{Generic getter to extract image coordinate transformation}
\usage{
trans(x)
}
\arguments{
\item{x}{an object with a transformation}
}
\description{
Generic getter to extract image coordinate transformation
}
\details{
This function returns a transformation that can be used to go from "grid coordinates" to "real world coordinates" in millimeters.

This function returns a transformation that can be used to go from "grid coordinates" to "real world coordinates" in millimeters.
see \code{\linkS4class{NeuroSpace}}
}
\examples{
bspace <- NeuroSpace(c(10,10,10), c(2,2,2))
trans(bspace)
all.equal(dim(trans(bspace)), c(4,4))
}

