% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{values}
\alias{values}
\title{Generic function to extract data values of object}
\usage{
values(x, ...)
}
\arguments{
\item{x}{the object to get values from}

\item{...}{additional arguments}
}
\description{
Generic function to extract data values of object
}
