% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/roi.R
\name{values}
\alias{values}
\alias{values,ROIVol-method}
\alias{values,ROIVec-method}
\title{Extract Data Values of an Object}
\usage{
values(x, ...)

\S4method{values}{ROIVol}(x, ...)

\S4method{values}{ROIVec}(x, ...)
}
\arguments{
\item{x}{the object to get values from}

\item{...}{additional arguments}
}
\value{
A vector or array containing the values extracted from the object.
}
\description{
Extract Data Values of an Object
}
