% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R
\docType{methods}
\name{vectors}
\alias{vectors}
\alias{vectors,NeuroVec,missing-method}
\alias{vectors,NeuroVec,numeric-method}
\alias{vectors,NeuroVec,logical-method}
\alias{blocks,NeuroVec,integer-method}
\title{Generic function to extract an ordered series of 1D vectors}
\usage{
vectors(x, subset, ...)

\S4method{vectors}{NeuroVec,missing}(x)

\S4method{vectors}{NeuroVec,numeric}(x, subset)

\S4method{vectors}{NeuroVec,logical}(x, subset)

\S4method{blocks}{NeuroVec,integer}(x, indices, ...)
}
\arguments{
\item{x}{the object that supplies the vector data}

\item{subset}{the subset of vectors to extract}

\item{...}{additional arguments}
}
\description{
Generic function to extract an ordered series of 1D vectors
}
