% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R
\docType{methods}
\name{vols}
\alias{vols}
\alias{vols,NeuroVec,numeric-method}
\alias{vols,NeuroVec,missing-method}
\title{Generic function to extract an ordered series of 3D volumes}
\usage{
vols(x, indices, ...)

\S4method{vols}{NeuroVec,numeric}(x, indices)

\S4method{vols}{NeuroVec,missing}(x)
}
\arguments{
\item{x}{the object that supplies the volume data}

\item{indices}{the indices of the volumes}

\item{...}{additional arguments}
}
\description{
Generic function to extract an ordered series of 3D volumes
}
