% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurovec.R
\name{vols}
\alias{vols}
\alias{vols,NeuroVec,numeric-method}
\alias{vols,NeuroVec,missing-method}
\title{extract an ordered series of 3d volumes}
\usage{
vols(x, indices, ...)

\S4method{vols}{NeuroVec,numeric}(x, indices)

\S4method{vols}{NeuroVec,missing}(x)
}
\arguments{
\item{x}{the object that supplies the volume data}

\item{indices}{the subset of volumes to extract}

\item{...}{additional arguments}
}
\description{
extract an ordered series of 3d volumes
}
\examples{

vec <- read_vec(system.file("extdata", "global_mask_v5.nii.gz", package="neuroim2"))
vs <- vols(vec)
length(vs) == dim(vec)[4]

vs <- vols(vec, indices=1:3)
length(vs) == 3
}
