% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/roi.R
\name{voxels}
\alias{voxels}
\alias{voxels,Kernel-method}
\title{extract voxel coordinates}
\usage{
voxels(x, ...)

\S4method{voxels}{Kernel}(x, center_voxel = NULL)
}
\arguments{
\item{x}{the object to extract voxels from}

\item{...}{additional arguments to function}

\item{center_voxel}{the absolute location of the center of the voxel, default is (0,0,0)}
}
\description{
extract voxel coordinates
}
