% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\docType{methods}
\name{which_dim}
\alias{which_dim}
\alias{which_dim,NeuroSpace,NamedAxis-method}
\title{which_dim}
\usage{
which_dim(x, axis)

\S4method{which_dim}{NeuroSpace,NamedAxis}(x, axis)
}
\description{
which_dim
}
