% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neurospace.R
\name{which_dim}
\alias{which_dim}
\alias{which_dim,NeuroSpace,NamedAxis-method}
\title{Find Dimensions of a Given Axis}
\usage{
which_dim(x, axis)

\S4method{which_dim}{NeuroSpace,NamedAxis}(x, axis)
}
\arguments{
\item{x}{An object representing the input data, such as a matrix or an array.}

\item{axis}{An integer representing the axis for which the dimension is requested. For example, 1 for rows, 2 for columns, or 3 for the depth in a 3D array.}
}
\value{
An integer representing the dimension of the specified axis for the given object.
}
\description{
This function returns the dimension of the specified axis for a given object, such as a matrix or an array.
}
\examples{

x <- NeuroSpace(c(10,10,10), spacing=c(1,1,1))
which_dim(x, x@axes@j) == 2

}
