% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_io.R
\name{write_elements,BinaryWriter,numeric-method}
\alias{write_elements,BinaryWriter,numeric-method}
\title{Write Elements to Binary Writer}
\usage{
\S4method{write_elements}{BinaryWriter,numeric}(x, els)
}
\arguments{
\item{x}{Object of class \linkS4class{BinaryWriter}}

\item{els}{Numeric vector of elements to write}
}
\value{
None (called for side effect)
}
\description{
Write elements to a \linkS4class{BinaryWriter} object.
}
\examples{
\dontrun{
writer <- BinaryWriter("output.bin", byte_offset = 0L,
                      data_type = "double", bytes_per_element = 8L)
# Write some data
write_elements(writer, rnorm(100))
close(writer)
}
}
