% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{write_elements}
\alias{write_elements}
\title{Write a sequence of elements from an input source}
\usage{
write_elements(x, els)
}
\arguments{
\item{x}{the output channel}

\item{els}{the elements to write}
}
\description{
Write a sequence of elements from an input source
}
