% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{write_vec}
\alias{write_vec}
\title{Generic function to write a 4D image vector to disk}
\usage{
write_vec(x, file_name, format, dataType)
}
\arguments{
\item{x}{an image object, typically a \code{BrainVector} instance.}

\item{file_name}{output file name.}

\item{format}{file format string. Since "NIFTI" is the only currently supported format, this parameter can be safely ignored and omitted.}

\item{dataType}{the numeric data type. If specified should be a \code{character} vector of: "BINARY", "UBYTE", "SHORT", "INT", "FLOAT", "DOUBLE".
Otherwise output format will be inferred from R the datatype of the image.}
}
\description{
Generic function to write a 4D image vector to disk
}
\examples{

bvec <- BrainVector(array(0, c(10,10,10,10)), NeuroSpace(c(10,10,10,10), c(1,1,1)))
\dontrun{
writeVector(bvol, "out.nii")
writeVector(bvol, "out.nii.gz")
writeVector(bvec, "out.nii")
writeVector(bvec, "out.nii.gz")
}
}

